% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{aboutUI}
\alias{aboutUI}
\title{UIs for entering general metadata}
\usage{
aboutUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaCreateLoadServ}} or \code{\link{metaEditLoadServ}} and 
\code{\link{metaSaveServ}}.}
}
\description{
UIs for entering general metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates text boxes for general metadata. These
  include author, date created, confidentiality, date modified, version, 
  general description, source, collection method, and intended use.
}
\keyword{internal}
