% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accumarray}
\alias{accumarray}
\title{Accumarray fucntion}
\usage{
accumarray(subs, val, sz = NULL, func = sum, fillval = 0)
}
\arguments{
\item{subs}{subs}

\item{val}{val}

\item{sz}{sz}

\item{func}{set to sum}

\item{fillval}{set to 0}
}
\description{
Accumarray fucntion
}
