% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_dates}
\alias{add_missing_dates}
\title{Add missing dates to `MainDataTable`.}
\usage{
add_missing_dates(dataset, date, value, group = NULL, facet_by = NULL)
}
\arguments{
\item{dataset}{Object containing `MainDataTable`.}

\item{date}{String, name of date variable to find missing days.}

\item{value}{String, name of value variable to be aggregated by \code{agg_helper}.}

\item{group}{String, name of grouping variable(s).}

\item{facet_by}{String, name of variables(s) to be facetted (split).}
}
\description{
Add missing dates to `MainDataTable`.
}
\keyword{internal}
