% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spat_area.R
\name{add_polygon}
\alias{add_polygon}
\title{Add polygon to spatial data}
\usage{
add_polygon(poly, spat, spat.id, new.id = NULL, combine = FALSE)
}
\arguments{
\item{poly}{A valid polygon to add to spatial data.}

\item{spat}{Spatial dataset to add polygon too.}

\item{spat.id}{The ID column in \code{spat}.}

\item{new.id}{The ID for new polygon.}

\item{combine}{Whether to use \code{\link{combine_zone}}. This will turn 
the intersections between \code{poly} and \code{spat} into new polygons. 
Note that the new polygon IDs will be derived from \code{spat} and \code{new.id} 
will not be used.}
}
\description{
Add polygon to spatial data
}
\seealso{
\code{\link{combine_zone}}
}
\keyword{internal}
