% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prompter}
\alias{add_prompter}
\title{Wrapper for prompter::add_prompt()}
\usage{
add_prompter(
  ui_element,
  position = "bottom",
  message = NULL,
  type = NULL,
  size = NULL,
  permanent = FALSE,
  rounded = FALSE,
  animate = TRUE,
  bounce = FALSE,
  arrow = TRUE,
  shadow = TRUE
)
}
\description{
Wrapper for prompter::add_prompt()
}
\keyword{internal}
