% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vars_gui.R
\name{add_vars}
\alias{add_vars}
\title{View and select which variables to add to working data set}
\usage{
add_vars(working_dat, raw_dat, project)
}
\arguments{
\item{working_dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{raw_dat}{Main raw (unmodified data frame)}

\item{project}{Name of project. Parameter is used to generate meaningful table names in fishset_db database.}
}
\description{
View and select which variables to add to working data set
}
\details{
Opens an interactive table that allows uses to select which variables to included by clicking check boxes. 
Data should be loaded into the fishset_db database before running this function. Select variables that will be used to generate further variables, such as rates or cpue, and variables to be included in models.
}
\examples{
\dontrun{
select_vars('pcodMainDataTable')
}
}
