% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vars_gui.R
\name{add_vars_gui}
\alias{add_vars_gui}
\title{Add removed variables back into dataset}
\usage{
add_vars_gui(working_dat, raw_dat, project)
}
\arguments{
\item{working_dat}{Primary data containing information on hauls or trips. Table in FishSET database contains the string 'MainDataTable'.}

\item{raw_dat}{Unmodified raw version of the primary dataset. Should be a character specifying a table from the FishSET database containing the string ‘MainDataTable’ and date table was created.}

\item{project}{String, name of project.}
}
\description{
Add columns that have been removed from the primary dataset back into the primary dataset.
}
\details{
Opens an interactive table that allows users to select which variables to be added back into the working dataset. \cr
The removed variables are obtained from the \code{raw_dat} and merged into the working data based on a row identifier. The row identifier is created when the variable is removed using the \code{\link{select_vars}} function. The row identifier is used to match the raw data variables to \code{working_dat}.
}
\examples{
\dontrun{
select_vars_gui(pcodMainDataTable)
add_vars_gui(pcodMainDataTable, 'pcodMainDataTable20100101', 'pcod')
}
}
