\name{adfgstat6_map}
\docType{data}
\alias{adfg}
\title{Example grid data}
\description{
  This data set contains an example grid data.
}
\usage{data(adfgstat6_map)}
\format{A dataframe containing 5 columns and 362318 observations. X5 is the zoneID and X1 and X2 are longitude and latitude, respectively. X3 and X4 can be ignored.}
\source{}
\keyword{datasets}