% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alaskaNMFSAreas}
\alias{alaskaNMFSAreas}
\title{Alaskan NMFS Areas}
\format{
`alaskaNMFSAreas`
A simple features multipolygon dataframe with 10 features and two fields:
\describe{
  \item{NMFS_AREA}{NMFS area number.}
  \item{geometry}{Coordinates for NMFS area.}
}
}
\usage{
alaskaNMFSAreas
}
\description{
A spatial dataset of Alaskan NMFS areas.
}
\keyword{datasets}
