% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alaskaPorts}
\alias{alaskaPorts}
\title{Alaskan Ports}
\format{
`alaskaPorts`
A dataframe with three columns and 88 rows:
\describe{
  \item{PORT}{Name of port.}
  \item{LONGITUDE}{LONGITUDE of port.}
  \item{LATITUDE}{LATITUDE of port.}
}
}
\usage{
alaskaPorts
}
\description{
A dataframe of Alaskan ports.
}
\keyword{datasets}
