% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_column.R
\name{assignment_column}
\alias{assignment_column}
\title{Assign observations to fishing zones}
\usage{
assignment_column(
  dat,
  project,
  spat,
  lon_dat,
  lat_dat,
  zoneID_spat,
  name = "ZoneID",
  closest_pt = FALSE,
  bufferval = NULL,
  lon_spat = NULL,
  lat_spat = NULL,
  hull_polygon = FALSE,
  epsg = NULL,
  log_fun = TRUE
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in 
FishSET database contains the string 'MainDataTable'.}

\item{project}{name of project.}

\item{spat}{Spatial data containing information on fishery management or 
regulatory zones. \code{sf} objects are recommended, but \code{sp} objects
can be used as well. If using a spatial table read from a csv file, then
arguments \code{lon_spat} and \code{lat_spat} are required. To upload your
spatial data to the FishSETFolder see \code{\link{load_spatial}}.}

\item{lon_dat}{Longitude variable in \code{dat}.}

\item{lat_dat}{Latitude variable in \code{dat}.}

\item{zoneID_spat}{Variable or list in \code{spat} that identifies the individual areas 
or zones. If \code{spat} is class \code{sf}, \code{zoneID_spat} should be name of list 
containing information on zones.}

\item{name}{The name of the new assignment column. Defaults to \code{"ZoneID"}.}

\item{closest_pt}{Logical, if \code{TRUE}, observations that fall outside 
zones are classed as the closest zone polygon to the point.}

\item{bufferval}{Maximum buffer distance, in meters, for assigning observations 
to the closest zone polygon. If the observation is not within the defined 
\code{bufferval}, then it will not be assigned to a zone polygon. 
Required if \code{closest_pt = TRUE}.}

\item{lon_spat}{Variable or list from \code{spat} containing longitude data. 
Required for spatial tables read from csv files. Leave as \code{NULL} if 
\code{spat} is an \code{sf} or \code{sp} object.}

\item{lat_spat}{Variable or list from \code{spat} containing latitude data. 
Required for spatial tables read from csv files. Leave as \code{NULL} if 
\code{spat} is an \code{sf} or \code{sp} object.}

\item{hull_polygon}{Logical, if \code{TRUE}, creates convex hull polygon. Use 
if spatial data creating polygon are sparse or irregular.}

\item{epsg}{EPSG code. Manually set the epsg code, which will be applied to 
\code{spat} and \code{dat}. If epsg is not specified but is defined for 
\code{spat}, then the \code{spat} epsg will be applied to \code{dat}. In addition,
if epsg is not specified and epsg is not defined for \code{spat}, then a default
epsg value will be applied to \code{spat} and \code{dat} (\code{epsg = 4326}).
See \url{http://spatialreference.org/} to help identify optimal epsg number.}

\item{log_fun}{Logical, whether to log function call (for internal use).}
}
\value{
Returns primary dataset with new assignment column.
}
\description{
Assign each observation in the primary dataset to a fishery management or 
regulatory zone. This function is primarily called by other functions that 
require zone assignment but can also be used on its own.
}
\details{
Function uses the specified latitude and longitude from the primary 
  dataset to assign each row of the primary dataset to a zone. Zone polygons 
  are defined by the spatial dataset. Set \code{hull_polygon} to \code{TRUE} 
  if spatial data is sparse or irregular. Function is called by other functions 
  if a zone identifier does not exist in the primary dataset.
}
\examples{
\dontrun{
pollockMainDataTable <- 
     assignment_column(pollockMainDataTable, "pollock", spat = pollockNMFSSpatTable,
                       lon_dat = "LonLat_START_LON", lat_dat = "LonLat_START_LAT")
}
}
\keyword{polygon}
\keyword{zone,}
