% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox}
\alias{bbox}
\title{Compute bounding box for a dataframe with lon/lat columns.}
\usage{
bbox(dat, lon, lat, f = 0.05)
}
\arguments{
\item{dat}{Dataframe containing longitude/latitude columns.}

\item{lon}{Name of Longitude column.}

\item{lat}{Name of Latitude column.}

\item{f}{Number specifying the fraction by which to extend the range.}
}
\description{
Compute bounding box for a dataframe with lon/lat columns.
}
\keyword{internal}
