% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{bin_var}
\alias{bin_var}
\title{Creates numeric variables divided into equal sized groups}
\usage{
bin_var(dat, project, var, br, name = "bin", labs = NULL, ...)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{var}{Numeric variable in \code{dat} to bin into a factor.}

\item{br}{Numeric vector. If a single number, the range of \code{var} is divided into \code{br} even groups. If two or more values are given, \code{var} is divided into intervals.}

\item{name}{Variable name to return. Defaults to `bin`.}

\item{labs}{A character string of category labels.}

\item{...}{Additional arguments passed to \code{\link{cut}}.}
}
\value{
Returns the primary dataset with binned variable added.
}
\description{
Creates numeric variables divided into equal sized groups
}
\details{
Function adds a new factor variable, labeled by name, to the primary dataset. The numeric variable 
is divided into equal sized groups if the length of \code{br} is equal to one and into intervals if the 
length of \code{br} is greater than one.
}
\examples{
\dontrun{
 pollockMainDataTable <- bin_var(pollockMainDataTable, 'pollock', 'HAUL', 10, 'HAULCAT')
 pollockMainDataTable <- bin_var(pollockMainDataTable, 'pollock', 'HAUL', c(5,10), 'HAULCAT')
}
}
