% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{cache_check_table}
\alias{cache_check_table}
\title{Add check table to confidentiality cache list}
\usage{
cache_check_table(check, project)
}
\arguments{
\item{check}{Dataframe, check table to be added to confidentiality cache.}

\item{project}{Name of project.}
}
\description{
Add check table to confidentiality cache list
}
\keyword{internal}
