% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trip_centroid.R
\name{calc_trip_centroid}
\alias{calc_trip_centroid}
\title{Calculate trip centroid variable}
\usage{
calc_trip_centroid(dat, project, lon, lat, trip_id, weight_var = NULL)
}
\arguments{
\item{dat}{String or data frame. A string for the name of the main data table in the FishSET 
database (contains 'MainDataTable' in the name). Or a data frame of the main data table.}

\item{project}{String. Project name.}

\item{lon}{String. Column name in \code{dat} containing longitudinal data.}

\item{lat}{String. Column name in \code{dat} containing latitudinal data.}

\item{trip_id}{String. Column name that represents the unique trip identifier in \code{dat}.}

\item{weight_var}{String. Optional. Column name in \code{dat} to use for computing a 
weighted average centroid. If \code{NULL} (the default), an unweighted (simple)
average is calculated.}
}
\value{
Returns the original data frame (\code{dataset}) with two new columns added: 
  \code{cent_lon} (centroid longitude) and \code{cent_lat} (centroid latitude).
}
\description{
Calculate latitude and longitude variables (columns) containing the geographic
centroid of each trip
}
\details{
This function computes the average longitude and latitude for each unique trip, as 
  defined by the \code{trip_id} column. If \code{weight_var} is specified, the function 
  calculates the weighted centroid.
}
\examples{
\dontrun{
# Assuming 'pollockMainDataTable' is a data frame

pollockMainDataTable <- calc_trip_centroid(
  dat = pollockMainDataTable, 
  project = 'pollock', 
  lon = 'LonLat_START_LON',
  lat = 'LonLat_START_LAT',
  trip_id = "TRIP_ID",
  weight_var = NULL
)
}
}
