% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_to_fsdb.R
\name{centroid_to_fsdb}
\alias{centroid_to_fsdb}
\title{Save Primary Table's Centroid Columns to FishSET Database}
\usage{
centroid_to_fsdb(
  dat,
  spat = NULL,
  project,
  zoneID,
  cent.lon,
  cent.lat,
  type = "zone"
)
}
\arguments{
\item{dat}{Required, main data frame containing data on hauls or trips.
Table in FishSET database should contain the string \code{MainDataTable}.}

\item{spat}{Optional, a name to associate with the centroid table.}

\item{project}{Name of project.}

\item{zoneID}{Variable in \code{dat} that identifies the individual zones or
areas.}

\item{cent.lon}{Required, variable in \code{dat} that identifies the centroid
longitude of zones or areas.}

\item{cent.lat}{Required, variable in \code{dat} that identifies the centroid
latitude of  zones or areas.}

\item{type}{The type of centroid. Options include \code{"zone"} for zonal centroids
and \code{fish} for fishing centroids.}
}
\description{
Save the unique centroid values from the primary table to the FishSET Database.
Use this function if zone ID and centroid longitude/latitude are included in
the primary table.
}
\details{
In certain cases, the user may have the necessary spatial variables to run a
discrete choice model included in the primary table when uploaded to FishSET,
and does not need a spatial table to assign observations to zones or find centroids
(e.g. by using \code{\link[=create_centroid]{create_centroid()}}). However, a centroid table table must be
saved to the FishSET Database if a centroid option is used to define
alternative choice (see \code{\link[=create_alternative_choice]{create_alternative_choice()}}). \code{cent_to_fsdb()}
allows users to save a zonal or fishing centroid table provided they have
the required variables: a zone ID (\code{zoneID}), a centroid longitude (\code{cent.lon}),
and a centroid latitude (\code{cent.lat}) column.
}
