% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_class.R
\name{change_class}
\alias{change_class}
\title{Change variable data class}
\usage{
change_class(dat, project, x = NULL, new_class = NULL, save = FALSE)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Name of project.}

\item{x}{A character string of variable(s) in \code{dat} that will be changed 
to \code{new_class}. One ore more variables may be included. Default set to NULL.}

\item{new_class}{A character string of data classes that \code{x} should be 
changed to. Length of \code{new_class} should match the length of \code{x} 
unless all variables in \code{x} should be the same \code{new_class}.
Defaults to NULL. Options are "numeric", "factor", "date", "character". Must be in quotes.}

\item{save}{Logical. Should the data table be saved in the FishSET database, 
replacing the working data table in the database? Defaults to FALSE.}
}
\value{
Table with data class for each variable and the working data with modified 
  data class as specified.
}
\description{
View data class for each variable and call appropriate functions to change data class as needed.
}
\details{
Returns a table with data class for each variable in \code{dat} and 
  changes variable classes. To view variable classes run the function with default 
  settings, specifying only \code{dat} and \code{project}. If variable class 
  should be changed, run the function again, specifying the variable(s) \code{(x)} 
  to be changed and the new_class(es) \code{(new_class)}. Set \code{save} to 
  TRUE to save modified data table.
}
\examples{
\dontrun{
#View table without changing class or saving
change_class(pollockMainDataTable, "myproject")

#Change class for a single variable and save data table to FishSET database
change_class(pollockMainDataTable, "myproject", x = "HAUL", new_class = 'numeric', save=TRUE)

#Change class for multiple variables and save data table to FishSET database
change_class(pollockMainDataTable, "myproject", x = c("HAUL","DISEMBARKED_PORT"),
 new_class = c('numeric', 'factor'), save=TRUE)
}
}
