% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{check_and_suppress}
\alias{check_and_suppress}
\title{Check and suppress data}
\usage{
check_and_suppress(
  dat,
  output,
  project,
  v_id,
  value_var,
  group = NULL,
  rule,
  value,
  type = "code",
  names_to = "name",
  values_to = "value"
)
}
\arguments{
\item{dat}{The dataset used to create a summary table. This must include 
the vessel identifier column.}

\item{output}{The output table to be suppressed. If \code{output = NULL}, 
\code{dat} is used. 
@param project Name of project.}

\item{v_id}{String, the name of the vessel identifier column.}

\item{value_var}{String, the name(s) of the value variable(s).}

\item{group}{String, the name(s) of the grouping variable(s). This should 
include the `period` name if summarizing over time.}

\item{rule}{String, the confidentiality rule to apply. \code{rule = "n"} 
suppresses values containing fewer than n vessels. \code{rule = "k"} (the 
"majority allocation rule") suppresses values where a single vessel contains 
k percent or more of the total catch.}

\item{value}{The threshold for confidentiality. for \code{rule = "n"} must 
be an integer of at least 3. For \code{rule = "k"} any double value from
0 to 100.}

\item{type}{String, the value used to replace confidential data. \code{"code"} 
replaces values with \code{-999}, \code{"NA"} (with quotes) replaces with 
\code{NA}, and \code{"zero"} replaces with 0.}

\item{names_to}{String, the name for the column containing the names of value 
variables when `value_var` has two or more columns.}

\item{values_to}{String, the name for the column containing the values from the
variables listed in `names_to`.}
}
\description{
Check and suppress data
}
\keyword{internal}
