% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{check_conf_rc}
\alias{check_conf_rc}
\title{Check and suppress roll_catch output}
\usage{
check_conf_rc(dat, roll_tab, project, catch, date, group, k, full_dates, align)
}
\arguments{
\item{dat}{Dataset used to create \code{roll_tab} dataframe.}

\item{roll_tab}{Unsuppressed table from \code{roll_catch}.}

\item{project}{Name of project.}

\item{catch}{String, name of catch variable(s).}

\item{date}{String, name of date variable.}

\item{group}{String, name of group variable(s).}

\item{k}{Integer, width of window.}

\item{full_dates}{Vector of full dates.}

\item{align}{String, align argument for \code{rollapply()}.}
}
\description{
Check and suppress roll_catch output
}
\keyword{internal}
