% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{check_confid_par}
\alias{check_confid_par}
\title{Check confidentiality parameters}
\usage{
check_confid_par(rule, value)
}
\arguments{
\item{rule}{String, \code{"n"} for rule of n, \code{"k"} for n/k.}

\item{value}{Numeric, for \code{rule = "n"} must be an integer of at least 
2. For \code{rule = "k"} any numeric value from 0 to 100.}
}
\value{
\code{TRUE} if confidentiality parameters are valid, \code{FALSE} 
  if not.
}
\description{
Check confidentiality parameters
}
\keyword{internal}
