% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{check_confidentiality}
\alias{check_confidentiality}
\title{Create confidentiality check table}
\usage{
check_confidentiality(
  dataset,
  project,
  v_id,
  value_var,
  group = NULL,
  rule = c("n", "k"),
  value,
  names_to = "name",
  values_to = "value"
)
}
\arguments{
\item{dataset}{The dataset used to create a summary table. This must include 
the vessel identifier column.}

\item{project}{Name of project.}

\item{v_id}{String, the name of the vessel identifier column.}

\item{value_var}{String, the name(s) of the value variable(s).}

\item{group}{String, the name(s) of the grouping variable(s). This should 
include the `period` name if summarizing over time.}

\item{rule}{String, the confidentiality rule to apply. \code{rule = "n"} 
suppresses values containing fewer than n vessels. \code{rule = "k"} (the 
"majority allocation rule") suppresses values where a single vessel contains 
k percent or more of the total catch.}

\item{value}{The threshold for confidentiality. for \code{rule = "n"} must 
be an integer of at least 3. For \code{rule = "k"} any double value from
0 to 100.}

\item{names_to}{String, the name for the column containing the names of value 
variables when \code{value_var} has two or more columns.}

\item{values_to}{String, the name for the column containing the values from the
variables listed in \code{names_to}.}
}
\description{
This function checks for confidential values in a summary table and creates 
a table of suppression conditions, or "check table".
}
\keyword{internal}
