% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exp.R
\name{check_exp}
\alias{check_exp}
\title{Prepare expected catch matrix}
\usage{
check_exp(ec, ec_names)
}
\arguments{
\item{ec}{Expected catch list created by \code{\link{create_expectations}}.}

\item{ec_names}{The names of the expected catch matrices to include in the
model and how they are used (i.e. used together or separately). See the
\code{expectcatchmodels} argument in \code{\link{make_model_design}}.}
}
\value{
Returns a list containing the filtered expected catch list and specified
  matrices to include in model.
}
\description{
Checks that specified expected catch matrices exists and is formatted correctly
when \code{\link{make_model_design}} is called.
}
\details{
Checks that \code{ec_names} is specified properly by identifying 
  invalid options such as combining 'individual' or 'all' with any other option 
  or including matrices that don't exist in \code{ec}.
}
\keyword{internal}
