% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proj}
\alias{check_proj}
\title{Check for project folder. Create folders if required}
\usage{
check_proj(project = NULL)
}
\arguments{
\item{project}{Project name}
}
\description{
Check for project folder. Create folders if required
}
\keyword{internal}
