% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{check_spatdat}
\alias{check_spatdat}
\title{Check and correct spatial data format}
\usage{
check_spatdat(spatdat, lon = NULL, lat = NULL, id = NULL)
}
\arguments{
\item{spatdat}{Spatial data containing information on fishery management or 
regulatory zones.}

\item{lon}{Longitude variable in \code{spatdat}. This is required for csv files 
or if \code{spatdat} is a dataframe (i.e. is not a \code{sf} or \code{sp} object).}

\item{lat}{Latitude variable in \code{spatdat}. This is required for csv files 
or if \code{spatdat} is a dataframe (i.e. is not a \code{sf} or \code{sp} object).}

\item{id}{Polygon ID column. This is required for csv files or if \code{spatdat} 
is a dataframe (i.e. is not a \code{sf} or \code{sp} object).}
}
\description{
Converts spatial data to a \code{sf} object
}
\details{
This function checks whether \code{spatdat} is a \code{sf} object
  and attempts to convert it if not. It also applies \code{\link{clean_spat}}
  which fixes certain spatial issues such as invalid or empty polygons, 
  whether a projected CRS is used (converts to WGS84 if detected), and if 
  longitude should be shifted to Pacific view (0-360 format) to avoid 
  splitting the Alaska region during plotting.
}
