% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Model checklist}
\usage{
checklist(project, modDesignTab = NULL)
}
\arguments{
\item{project}{String, name of project.}

\item{modDesignTab}{The model design table. Only used in the FishSET app.}
}
\description{
Determines whether final data passes quality checks, that occurrence points
have been checked, that an alternative choice matrix has been created, and
if an expected catch/revenue matrix is required.
}
\keyword{internal}
