% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_directory}
\alias{choose_directory}
\title{Choose directory}
\usage{
choose_directory(
  method = select_directory_method(),
  title = "Identify location of FishSET folder and save output"
)
}
\arguments{
\item{method}{Method function}

\item{title}{Title to show}
}
\description{
Choose directory
}
\keyword{internal}
