% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{clean_list}
\alias{clean_list}
\title{Remove unnecessary nested lists}
\usage{
clean_list(l)
}
\arguments{
\item{l}{A list.}
}
\value{
a list.
}
\description{
Recursively removes unnecessary lists ("lame lists") from a list. A lame list
is a unnamed list containing a single list object (see \code{\link{lame_list}}).
}
\examples{
\dontrun{
clean_list(list(list(1:10)))
clean_list(list(A = list(1:10)))
clean_list(list(list(1:10), list(11:20)))
}
}
\seealso{
\code{\link{simplify_list}} \code{\link{lame_list}}
}
\keyword{internal}
