% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{clean_spat}
\alias{clean_spat}
\title{Clean spatial data}
\usage{
clean_spat(spat)
}
\arguments{
\item{spat}{Spatial data to check.}
}
\description{
Clean spatial data
}
\details{
\code{clean_spat} extracts polygons from "GEOMETRYCOLLECTION"
spatial features, removes non-polygons from the data, attempts to fix 
invalid geometries, and shifts longitude to Pacific view if any points 
are less than 0.
}
\keyword{internal}
