% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit.R
\name{clogit}
\alias{clogit}
\title{clogit}
\usage{
clogit(starts3, dat, otherdat, alts)
}
\arguments{
\item{starts3}{Starting values}

\item{dat}{Data matrix, see output from shiftSortX, alternatives with distance by column bind}

\item{otherdat}{Other data used in model (as list)}

\item{alts}{Number of alternative choices in model}
}
\value{
ld - negative log likelihood
}
\description{
Conditional logit likelihood
}
