% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_closure_scenario.R
\name{close_names}
\alias{close_names}
\title{Retrieve closure scenario names}
\usage{
close_names(project)
}
\arguments{
\item{project}{Name of project}
}
\description{
A helper function used to display the names of currently saved closure
scenarios.
}
\details{
To retrieve the complete closure scenario file, use 
\code{\link{get_closure_scenario}}.
}
