% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_log_call.R
\name{col_desc}
\alias{col_desc}
\title{Create a list of column names
Used to create a new metadata entry.}
\usage{
col_desc(dat)
}
\arguments{
\item{dat}{Dataframe}
}
\description{
Create a list of column names
Used to create a new metadata entry.
}
\keyword{internal}
