% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{colDescUI}
\alias{colDescUI}
\title{Create a text box for each column in data}
\usage{
colDescUI(id, nm, value = "")
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaCreateLoadServ}} or \code{\link{metaEditLoadServ}} and 
\code{\link{metaSaveServ}}.}

\item{nm}{Column name passed to \code{textAreaInput} id.}

\item{value}{Value passed to \code{textAreaInput}.}
}
\description{
Create a text box for each column in data
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates a text box for each column in a FishSET
  table.
}
\keyword{internal}
