% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{collapse_leaf}
\alias{collapse_leaf}
\title{Convert vector to scalar}
\usage{
collapse_leaf(x)
}
\arguments{
\item{x}{A vector or object in a list.}
}
\value{
a character vector of length 1.
}
\description{
Collapses a vector to a scalar. This is needed to convert an R list to
an html list.
}
\keyword{internal}
