% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{collapse_leaf_r}
\alias{collapse_leaf_r}
\title{Recursively apply \code{collapse_leaf()} to a list object}
\usage{
collapse_leaf_r(l)
}
\arguments{
\item{l}{A list}
}
\value{
A list.
}
\description{
Recursively apply \code{collapse_leaf()} to a list object
}
\keyword{internal}
