% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{column_check}
\alias{column_check}
\title{Check that column names exists}
\usage{
column_check(dat, cols)
}
\arguments{
\item{dat}{The data used in function.}

\item{cols}{String, the column names used in function.}
}
\description{
Check whether user supplied column names exist in the data.
}
\keyword{internal}
