% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_zone.R
\name{combine_zone}
\alias{combine_zone}
\title{Combine zone and closure area}
\usage{
combine_zone(spat, closure, grid.nm, closure.nm, recast = TRUE)
}
\arguments{
\item{spat}{Spatial file containing regulatory zones.}

\item{closure}{Closure file containing closure areas.}

\item{grid.nm}{Character, column name containing grid ID.}

\item{closure.nm}{Character, column name containing closure ID.}

\item{recast}{Logical, if \code{TRUE} \code{combined} is passed to 
\code{recast_multipoly}.}
}
\description{
Creates a new spatial dataset that merges regulatory zones with closure areas.
}
\details{
To combine zones with closure areas, this function performs the 
  following steps:
  \enumerate{
    \item Create the union of the closure area
    \item Take the difference between the closure union and the zone file
    \item Take the intersection of zone and the closure union 
    \item Combine the difference and intersection objects into one spatial 
      dataframe
    \item Assign new zone IDs to intersecting polygons
  } 
  The result is a single spatial dataset containing all polygons from both \code{spat}
  and \code{closure} with overlapping (intersecting) polygons receiving new
  IDs (see \code{\link{new_zone_id}}). This allows users to partially close
  regulatory zones during the model design stage.
}
\seealso{
\code{\link{recast_multipoly}} \code{\link{new_zone_id}}
}
