% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_closure_overlaps.R
\name{compute_closure_overlaps}
\alias{compute_closure_overlaps}
\title{Identify model zones overlapped by an uploaded spatial file}
\usage{
compute_closure_overlaps(
  uploaded_files,
  zones,
  overlap_threshold,
  return_shape = FALSE
)
}
\arguments{
\item{uploaded_files}{The data frame returned by a Shiny spatial `fileInput`.}

\item{zones}{An `sf` object with a `second_location_id` column.}

\item{overlap_threshold}{Minimum percentage of a zone that must be covered.}

\item{return_shape}{Whether to also return the cleaned uploaded spatial object.}
}
\value{
A character vector of selected `second_location_id` values, or a list
  containing the selected IDs and cleaned uploaded shape when `return_shape` is `TRUE`.
}
\description{
Identify model zones overlapped by an uploaded spatial file
}
\keyword{internal}
