% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{confid_cache_exists}
\alias{confid_cache_exists}
\title{Confidentialy cache exists}
\usage{
confid_cache_exists(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Returns \code{TRUE} if confidentiality cache file is found in the project
output folder.
}
\examples{
\dontrun{
confid_cache_exists("pollock")
}
}
