% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{contactUI}
\alias{contactUI}
\title{UIs for entering metadata contact info}
\usage{
contactUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaCreateLoadServ}} or \code{\link{metaEditLoadServ}} and 
\code{\link{metaSaveServ}}.}
}
\description{
UIs for entering metadata contact info
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates text boxes for contact info. These
  include person, organization, address, phone number, email, license,
  citation, and general purpose "other" category.
}
\keyword{internal}
