% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corr_plot}
\alias{corr_plot}
\title{Correlation plot}
\usage{
corr_plot(corr, project)
}
\arguments{
\item{corr}{Correlation matrix}

\item{project}{Name of project}
}
\description{
Correlation plot
}
\keyword{internal}
