% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dist_between_gui.R
\name{create_dist_between_for_gui}
\alias{create_dist_between_for_gui}
\title{Distance between two points}
\usage{
create_dist_between_for_gui(dat, start, end, units, portTable = NULL,
  gridfile = NULL, lon.dat = NULL, lat.dat = NULL, cat = NULL,
  lon.grid = NULL, lat.grid = NULL)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{start}{Starting location. Should be a port, lat/long location, or the centroid of zonal assignment.}

\item{end}{Ending location. Should be a port, lat/long location, or the centroid of the fishing zone or area.}

\item{units}{Unit of distance (miles, kilometers)}

\item{portTable}{Data table containing port data}

\item{gridfile}{Grid file}

\item{lon.dat}{Variable containing longitude from main dataset}

\item{lat.dat}{Variable containing latitude from main dataset}

\item{cat}{Variable in grid file designating column containing zones or areas.}

\item{lon.grid}{Variable containing longitude from main grid file}

\item{lat.grid}{Variable containing latitude from main grid file}
}
\description{
Creates a vector of distance between two points. The start and end points must be different vectors. 
If the start or ending points are from a port or the center of a fishing zone or area, 
then a prompt will appear asking for further parameters to be specified. 
If the starting or ending points are a port, then latitude and longitude are extracted from the port table 
stored in the fishset_db database.  In this case, PortTable must be specified.  If the starting or ending points 
are the center of the fishing zone or area, then the assignment_column function will be called to assign each observation
 to a zone. The find_centroid function will then be called to determine the centroid of each zone. Distance measurements will be between these centroids.
}
