% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_duration}
\alias{create_duration}
\title{Create duration of time variable}
\usage{
create_duration(
  dat,
  project,
  start,
  end,
  units = c("week", "day", "hour", "minute"),
  name = "create_duration"
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{start}{Date variable from \code{dat} indicating start of time period.}

\item{end}{Date variable from \code{dat} indicating end of time period.}

\item{units}{String, unit of time for calculating duration. Must be \code{"week"}, \code{"day"}, \code{"hour"}, or \code{"minute"}.}

\item{name}{String, name of created vector. Defaults to name of the function if not defined.}
}
\value{
Returns primary dataset with duration of time variable added.
}
\description{
Create duration of time variable based on start and ending dates in desired temporal units.
}
\details{
Calculates the duration of time between two temporal variables based on defined time unit.
The new variable is added to the dataset.
A duration of time variable is required for other functions, such as \code{\link{cpue}}.
}
\examples{
\dontrun{
pollockMainDataTable <- create_duration(pollockMainDataTable, 'pollock', 'TRIP_START', 'TRIP_END',
  units = 'minute', name = 'TripDur')
}
}
