% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logit_input.R
\name{create_logit_input}
\alias{create_logit_input}
\title{Creates a data matrix consistent with built-in model forms}
\usage{
create_logit_input(choice)
}
\arguments{
\item{choice}{A dataframe with a single vector of chosen locations with length = number of observations}
}
\value{
\code{dataCompile}: A large data matrix.Number of rows = the number of observations, number of cols = square of number of alternatives. 
   Each row contains a flattened identity matrix of size = number of alternatives x number of alternatives.
}
\description{
Creates a data matrix consistent with built-in model forms
}
\details{
Called in the discrete_fish_subroutine function
}
\keyword{internal}
