% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_input.R
\name{create_model_input}
\alias{create_model_input}
\title{Create model input data}
\usage{
create_model_input(
  project,
  x = NULL,
  mod.name = NULL,
  use.scalers = TRUE,
  scaler.func = NULL,
  expected.catch = NULL,
  exp.names = NULL
)
}
\arguments{
\item{project}{Project name}

\item{x}{Optional, Model input data}

\item{mod.name}{Optional, name of model. Include if building model data for 
specific defined model.}

\item{use.scalers}{Logical, should data be normalized? Defaults to TRUE. 
Rescaling factors are the mean of the numeric vector unless specified with 
\code{scaler}.}

\item{scaler.func}{Function to calculate rescaling factors.}

\item{expected.catch}{For conditional logit. Expected catch matrices to include}

\item{exp.names}{The names of expected catch matrices to use in model. Specified
in \code{\link{make_model_design}}.}
}
\value{
Returns a list of datacompile/choice matrix, dataCompile, distance, 
  otherdat, expname, choice, choice.table, and mod.name
}
\description{
Takes the structure of data compile, price (for EPM), distance, gridvarying, 
interaction terms
}
\keyword{internal}
