% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{create_proj_settings}
\alias{create_proj_settings}
\title{Create FishSET project settings}
\usage{
create_proj_settings(project)
}
\description{
This function creates the project settings file located in the project/doc folder.
}
\details{
The project settings file contains the confidentiality settings,
the file path to  the user output folder, plot size (width by height in 
inches), and table names used in the shiny app (for logging purposes).
}
\seealso{
\code{\link{edit_proj_settings}}
}
\keyword{internal}
