% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal_ID.R
\name{create_seasonal_ID}
\alias{create_seasonal_ID}
\title{Create fishery season identifier variable}
\usage{
create_seasonal_ID(
  dat,
  project,
  seasonal.dat,
  use.location = c(TRUE, FALSE),
  use.geartype = c(TRUE, FALSE),
  sp.col,
  target = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{seasonal.dat}{Table containing date of fishery season(s). Can be pulled from the FishSET database.}

\item{use.location}{Logical, should fishery season dates depend on fishery location? Column names containing
location in \code{dat} and \code{seasonal.dat} must match.}

\item{use.geartype}{Logical, should fishery season dates depend on gear type. Column names containing gear
type in \code{dat} and \code{seasonal.dat} must match.}

\item{sp.col}{Variable in \code{seasonal.dat} containing species names.}

\item{target}{Name of target species. If \code{target} is NULL, runs through fisheries in order listed in \code{seasonal.dat}}
}
\value{
Returns the primary dataset with the variable SeasonID, or a series of variables identifying by the individual
  fisheries included (seasonID*fishery).
}
\description{
Create fishery season identifier variable
}
\details{
Uses a table of fishery season dates to create fishery season identifier variables. Output is a SeasonID
variable and/or multiple SeasonID*fishery variables. If fishery season dates vary by location or gear type,
then \code{use.location} and \code{use.geartype} should be TRUE. \cr\cr
The function matches fishery season dates provided in \code{seasonal.dat} to the earliest date variable in \code{dat}.
The `seasonID` variable is a vector of fishery seasons whereas the `SeasonID*fishery` variables are 1/0 depending on whether the
fishery was open on the observed date. \cr\cr
If \code{target} is not defined, then each row of seasonID is defined as the earliest fishery listed in \code{seasonal.dat} for
which the fishery season date encompasses the date variable in the primary dataset. If \code{target} fishery is defined, then
`SeasonID` is defined by whether the target fishery is open on the date in the primary dataset or a different fishery. The
vector is filled with 'target' or 'other'.\cr\cr
`SeasonID*fishery` variables are a 1/0 seasonID vector for each fishery (labeled by seasonID and fishery) where 1
indicates the dates for a given row in the main data table fall within the fishery dates for that fishery.
}
\examples{
\dontrun{
pcodMainDataTable <- create_seasonal_ID("pcodMainDataTable", seasonal_dat,
  use.location = TRUE, use.geartype = TRUE, sp.col = "SPECIES", target = "POLLOCK"
)
}

}
