% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_trip_centroid}
\alias{create_trip_centroid}
\title{Calculate centroid of each trip}
\usage{
create_trip_centroid(dat, lon, lat, weight.var = NULL, ...)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{lon}{Vector containing longitudinal data.}

\item{lat}{Vector containing latitudinal data.}

\item{weight.var}{Variable for weighted average.}

\item{...}{Column(s) that identify the individual trip.}
}
\description{
Calculate centroid of each trip
}
\details{
Returns centroid of each trip
}
\examples{
\dontrun{
pollockMainDataTable <- create_trip_centroid('pollockMainDataTable', 
                              'LonLat_START_LON', 'LonLat_START_LAT', 
                              weight.var=NULL, 'DISEMBARKED_PORT','EMBARKED_PORT')
}
}
