% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_distance.R
\name{create_trip_distance}
\alias{create_trip_distance}
\title{Create haul level trip distance variable}
\usage{
create_trip_distance(
  dat,
  project,
  PortTable,
  trip_id,
  starting_port,
  starting_haul = c("Lon", "Lat"),
  ending_haul = c("Lon", "Lat"),
  ending_port,
  haul_order,
  name = "TripDistance",
  a = 6378137,
  f = 1/298.257223563
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{PortTable}{Port data frame. Contains columns: Port_Name, Port_Long, Port_Lat.
Table is generated using the \code{\link{load_port}} function and saved in the FishSET database as the project and PortTable,
for example 'pollockPortTable'.}

\item{trip_id}{Unique trip identifier in \code{dat}.}

\item{starting_port}{Variable in \code{dat} containing ports at the start of the trip.}

\item{starting_haul}{Character string, variables containing latitude and longitude at start of haul in \code{dat}.}

\item{ending_haul}{Character string, variables containing latitude and longitude at end of haul in \code{dat}.}

\item{ending_port}{Variable in \code{dat} containing ports at the end of the trip.}

\item{haul_order}{Variable in \code{dat} that identifies haul order within a trip. Can be time, coded variable, etc.}

\item{name}{String, name of created variable. Defaults to `TripDistance`.}

\item{a}{Numeric, major (equatorial) radius of the ellipsoid. The default value is for WGS84 ellipsoid.}

\item{f}{Numeric, ellipsoid flattening. The default value is for WGS84 ellipsoid.}
}
\value{
Returns the primary dataset with a trip distance variable added.
}
\description{
Create haul level trip distance variable
}
\details{
Summation of distance across a trip based on starting and ending ports and hauls in between.
The function uses \code{\link[geosphere]{distGeo}} from the geosphere package to calculate distances
 between hauls. Inputs are the trips, ports, and hauls from the primary dataset, and the latitude and
 longitude of ports from the \code{PortTable}. The ellipsoid arguments, \code{a} and \code{f}, are numeric
 and can be changed if an ellipsoid other than WGS84 is appropriate. See the geosphere R package for more details
 (\url{https://cran.r-project.org/web/packages/geosphere/geosphere.pdf}).
}
\examples{
\dontrun{
pcodMainDataTable <- create_trip_distance(pcodMainDataTable, "pcod", "pcodPortTable", 
  "TRIP_SEQ", "DISEMBARKED_PORT", c("LonLat_START_LON", "LonLat_START_LAT"),
  c("LonLat_END_LON", "LonLat_END_LAT"), "EMBARKED_PORT", "HAUL_SEQ", "TripDistance"
)
}

#
}
