% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_var_temp}
\alias{create_var_temp}
\title{Create temporal variables}
\usage{
create_var_temp(dataset, start, end, units = c("week", "day", "hour",
  "minute"), name = "create_var_temp")
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{start}{Variable indicating start of time period}

\item{end}{Variable indicating end of time period}

\item{units}{Units of time varibles. Must be weeks, days, hours, or minutes}

\item{name}{Name of created vector. Used in the logging function to reproduce work flow. Defaults to name of the function if not defined.}
}
\description{
Create temporal variables
}
\details{
Function for generating new or specialized variables. create_var_temp calculates the duration of time between two temporal variables based on defined time format. The new variable is added to the dataset.
}
