% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{current_db_table}
\alias{current_db_table}
\title{Retrieve name of the most recent table from a project}
\usage{
current_db_table(project, table)
}
\arguments{
\item{project}{Name of project.}

\item{table}{Name of table, e.g. "MainDataTable".}
}
\description{
Retrieve name of the most recent table from a project
}
\keyword{internal}
