% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{current_log}
\alias{current_log}
\title{Lists most recent log file}
\usage{
current_log(project)
}
\arguments{
\item{project}{Project name.}
}
\description{
Lists most recent log file
}
\details{
Prints the name of the most recent log file, but not the filepath.
}
\keyword{internal}
