% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{current_out}
\alias{current_out}
\title{Lists most recent output files}
\usage{
current_out(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Lists most recent output files
}
\details{
Prints the name of the most recent output files.
}
\examples{
\dontrun{
current_out("pollock")
}
}
\keyword{internal}
