% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_pull}
\alias{data_pull}
\title{Pull data from sqlite database}
\usage{
data_pull(dat, project)
}
\arguments{
\item{dat}{Data table name (character) or a data frame object}

\item{project}{Project name}
}
\description{
Pull data from sqlite database
}
\keyword{internal}
