% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_upload_helper}
\alias{data_upload_helper}
\title{Upload data from file, FishSET DB, or working environment}
\usage{
data_upload_helper(dat, type, ...)
}
\arguments{
\item{dat}{Reference to a dataframe. This can be a filepath, the name of an
existing FishSET table, or a dataframe object in the working environment.}

\item{type}{The type of data to upload. Options include \code{"main"}, 
\code{"port"}, \code{"grid"}, \code{"aux"}, and \code{"spat"}.}

\item{...}{Additional arguments passed to \code{\link{read_dat}}.}
}
\description{
Helper function that can read data from file, from FishSET DB, or a dataframe
in the working environment. Used for data upload functions: 
\code{load_maindata}, \code{load_port}, \code{load_aux}, \code{load_grid},
\code{load_spatial}.
}
\examples{
\dontrun{
dataset <- data_upload_helper(dat, type = "main")
}
}
