% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{data_verification}
\alias{data_verification}
\title{Check for common data quality issues that may be present in the data set.}
\usage{
data_verification(dataset)
}
\arguments{
\item{dataset}{Main dataframe containing data on hauls or trips.}
}
\value{
Returns statements as to whether data quality issues may exist.
}
\description{
Function tests for common data quality issues.
}
\details{
Checks that all columnn names in the dataset are unique, whether any columns in the dataset are empty, whether each row is a unique choice 
occurrence at the haul or trip level, and that either latitude and longitude or fishing area are included.
}
\examples{
\dontrun{ 
data_verification(MainDataTable)
}
}
