% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification_call.R
\name{data_verification_call}
\alias{data_verification_call}
\title{Checks for common issues with data}
\usage{
data_verification_call(dat)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}
}
\value{
Returns statements as to whether issues in the data may exist
}
\description{
Checks for common issues with data
}
\details{
Contains one function checks that all columnn names in the dataset are unique, whether any columns in the dataset are empty, 
whether each row is a unique choice occurrence at the haul or trip level, and that data for either lat/long or fishing area are included.
Main data table is not saved to fishset_db database if any tests fail.
}
\examples{
\dontrun{
data_verification_call(MainDataTable)
}
}
