% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{dataindex_update}
\alias{dataindex_update}
\title{Update dataindex file}
\usage{
dataindex_update(dataset, dataindex)
}
\arguments{
\item{dataset}{name main dataframe}

\item{dataindex}{Name of dataindex file saved in the database}
}
\description{
Update dataindex file
}
