% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVerification.R
\name{dataVerification}
\alias{dataVerification}
\title{Check for serveral issues that may be present in the data.}
\usage{
dataVerification(dataset, dataindex)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{dataindex}{dataframe that contains information on each column of the dataset}
}
\value{
Returns statements as to whether issues in the data may exist
}
\description{
Contains one function that tests several if statements. The function stops if a if statement does not pass.
}
\details{
dataVerification checks that all columnn names in the dataset are unique, whether specialized variables have been identified in the index 
dataset, whether any columns in the dataset are empty, if units are defined and recognized, whether each row is a unique choice occurrence at the haul
or trip level, and that data for either lat/long or fishing area are included.
}
