% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_check}
\alias{date_check}
\title{Parse date/date-time variable}
\usage{
date_check(dat, date)
}
\arguments{
\item{dat}{Dataframe containing variable to convert to date/date-time.}

\item{date}{Date variable name to convert.}
}
\value{
Checks whether variable can be converted to date or date-time. Returns
 dataframe with converted variable.
}
\description{
Parse date/date-time variable
}
\keyword{internal}
