% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_cols}
\alias{date_cols}
\title{Find columns that can be converted to Date class}
\usage{
date_cols(dat, out = "names")
}
\arguments{
\item{dat}{MainDataTable or dataframe to check.}

\item{out}{Whether to return the column \code{"names"} (the default) or a logical vector 
(\code{"logical"}).}
}
\description{
Find columns that can be converted to Date class
}
\examples{
\dontrun{
date_cols(pollockMainDataTable) # returns column names
date_cols(pollockMainDataTable, "logical")
}
}
\keyword{internal}
