% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_factorize}
\alias{date_factorize}
\title{Convert date variable of type character to ordered factor}
\usage{
date_factorize(dataset, date_col, date_code)
}
\arguments{
\item{dataset}{data frame containing date variable.}

\item{date_col}{date variable of type character to convert to ordered factor.}

\item{date_code}{date code used to format date variable.}
}
\description{
Convert date variable of type character to ordered factor
}
\keyword{internal}
