% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_parser}
\alias{date_parser}
\title{Parse date variable}
\usage{
date_parser(dates, args = NULL)
}
\arguments{
\item{dates}{Variable containing dates}
}
\description{
Parse date variable
}
\keyword{internal}
