% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_time_parser}
\alias{date_time_parser}
\title{Parse date-time variable}
\usage{
date_time_parser(dates)
}
\arguments{
\item{dates}{Variable containing date-times}
}
\description{
Parse date-time variable
}
\keyword{internal}
