% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_title}
\alias{date_title}
\title{Add date to ggplot title}
\usage{
date_title(plot, filter_date, date_value)
}
\arguments{
\item{plot}{ggplot2 plot object to add new title to.}

\item{filter_date}{The \code{filter_date} parameter in function.}

\item{date_value}{The values used to filter the data table used in plot.}
}
\value{
A plot with the year, month, or year-month included in tittle.
}
\description{
Add date to ggplot title
}
\keyword{internal}
