% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{dd2dms}
\alias{dd2dms}
\title{Convert DD to DMS}
\usage{
dd2dms(dd, NS = FALSE)
}
\arguments{
\item{dd}{numeric vector of decimal degrees.}

\item{NS}{logical, \code{TRUE} for north/south decimal degrees, \code{FALSE} 
for east/west decimal degrees.}
}
\description{
Convert decimal degree format to degree minutes seconds format.
}
\keyword{internal}
