% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Check and convert lat/lon to decimal degrees}
\usage{
degree(
  dat,
  project,
  lat = NULL,
  lon = NULL,
  latsign = FALSE,
  lonsign = FALSE,
  replace = TRUE
)
}
\arguments{
\item{dat}{Dataset containing latitude and longitude data.}

\item{project}{Project name.}

\item{lat}{Variable containing latitude data.}

\item{lon}{Variable containing longitude data.}

\item{latsign}{How should the sign value of \code{lat} be changed? Choices are \code{NULL}, no change, 
\code{"all"}, change all values, \code{"neg"}, convert all positive values to negative, or 
\code{"pos"}, convert all negative values to positive.}

\item{lonsign}{How should the sign value of \code{lon} be changed? Choices are \code{NULL}, no change, 
\code{"all"}, change all values, \code{"neg"}, convert all positive values to negative, or 
\code{"pos"}, convert all negative values to positive.}

\item{replace}{Logical, should \code{lat} and \code{lon} in \code{dat} be converted to decimal degrees? 
Defaults to TRUE. Set to FALSE if checking for compliance}
}
\value{
Returns the primary dataset with the latitudes and longitudes converted to decimal degrees. 
   Changing the sign, transforms all values in the variable.
}
\description{
Check that latitude and longitude are in decimal degrees and the variable sign is correct. Correct lat/lon if required.
}
\details{
First checks whether any variables containing 'lat' or 'lon' in their names are numeric. Returns 
  a message on results. To convert a variable to decimal degrees, identify the \code{lat} or \code{lon}
  variable(s) and set \code{replace} to TRUE. To change the sign, set \code{latsign} (for \code{lat}) 
  or \code{lonsign} (for \code{lon} to TRUE. FishSET requires that latitude and longitude 
  be in decimal degrees.
}
\examples{
\dontrun{
pollockMainDataTable <- degree(pollockMainDataTable, 'pollock', 'LatLon_START_LAT', 
      'LatLon_START_LON', latsign=FALSE, lonsign=FALSE, replace=TRUE)
}

}
