% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{delete_meta}
\alias{delete_meta}
\title{Delete table meta data or project meta file}
\usage{
delete_meta(project, tab.name = NULL, delete_file = FALSE)
}
\arguments{
\item{project}{Project name.}

\item{tab.name}{String, table name.}

\item{delete_file}{Logical, whether to delete project meta file.}
}
\description{
Delete table meta data or project meta file
}
