% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_models.R
\name{delete_models}
\alias{delete_models}
\title{Delete models from FishSET Database}
\usage{
delete_models(project, model.names, delete.nested = FALSE)
}
\arguments{
\item{project}{String, name of project.}

\item{model.names}{String, name of models to delete. Use \code{\link[=model_names]{model_names()}} to
see model names from the model design file.}

\item{delete.nested}{Logical, whether to delete a model containing nested
models. Defaults to \code{FALSE}.}
}
\description{
Delete models from the model design file (MDF) and the model output table
(MOT).
}
\details{
Nested models are conditional logit models that include more than
one expected catch/revenue model. For example, if a conditional logit model
named \code{'logit_c_mod1'} was saved to the MDF with the argument
\code{expectcatchmodels = list('exp1', 'recent', 'older')}, then \verb{'logit_c_mod1}
will include three separate models, each using a different expected catch
matrix. To delete all three models, enter \code{model.names = 'logit_c_mod1'} and
set \code{delete.nested = TRUE}. To delete one or more specific nested models, use
\code{model.names = 'logit_c_mod1.exp1'}, i.e. the original model name, a period, and
the name of the expected catch matrix used in the model.
}
\seealso{
\code{\link[=model_design_list]{model_design_list()}}, \code{\link[=model_out_view]{model_out_view()}}
}
