% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{dens_plot_helper}
\alias{dens_plot_helper}
\title{density_plot helper function}
\usage{
dens_plot_helper(
  dataset,
  var,
  group,
  date,
  facet_by,
  filter_date,
  date_value,
  type,
  bw,
  conv,
  tran,
  format_lab,
  scale,
  position,
  pages
)
}
\arguments{
\item{dataset}{Data used to create plot.}

\item{var}{String, variable passed from \code{density_plot}.}

\item{group}{String, grouping variable(s) passed from \code{density_plot}.}

\item{date}{String, date variable passed from \code{density_plot}.}

\item{facet_by}{String, facet variable(s) passed from \code{density_plot}.}

\item{filter_date}{String, date filter type passed from \code{density_plot}.}

\item{date_value}{Numeric, date filter value passed from \code{density_plot}.}

\item{type}{String, plot type(s) passed from \code{density_plot}.}

\item{bw}{Numeric, bandwidth passed from \code{density_plot}.}

\item{conv}{String, convert pounds to "tons" or "metric_tons".}

\item{tran}{String, scale transformation passed from \code{density_plot}.}

\item{format_lab}{String, label formatting option passed from \code{density_plot}.}

\item{scale}{Scale argument passed to \code{\link{facet_grid}}. Defaults to "fixed". 
Other options include "free_y", "free_x", and "free".}

\item{position}{String, plot position passed from \code{density_plot}.}

\item{pages}{String, single or multiple plots passed from \code{density_plot}.}
}
\description{
Creates and formats plots
}
\keyword{internal}
