% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_name}
\alias{deparse_name}
\title{Deparse a data table for log}
\usage{
deparse_name(dat)
}
\arguments{
\item{dat}{A dataframe object or string to deparse for \code{log_call} function.}
}
\description{
Deparse a data table for log
}
\keyword{internal}
