% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{dms_to_dd}
\alias{dms_to_dd}
\title{Convert decimal minutes/seconds to degrees}
\usage{
dms_to_dd(x)
}
\arguments{
\item{x}{Latitude or longitude vector.}
}
\description{
Convert decimal minutes/seconds to degrees
}
\examples{
\dontrun{
pollockMainDataTable$LonLat_START_LAT <- dms_to_dd(pollockMainDataTable$LonLat_START_LAT)
}
}
\keyword{internal}
