% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_matrix}
\alias{dummy_matrix}
\title{Create dummy matrix from a coded ID variable}
\usage{
dummy_matrix(dat, project, x)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{x}{Variable in \code{dat} used to generate dummy matrix.}
}
\description{
Create dummy matrix from a coded ID variable
}
\details{
Creates a dummy matrix of 1/0 with dimensions \emph{[(number of observations in dataset) x
(number of factors in x)]} where each column is a unique factor level. Values are 1 if the value in the
column matches the column factor level and 0 otherwise.
}
\examples{
\dontrun{
PortMatrix <- dummy_matrix(pollockMainDataTable, 'pollock', 'PORT_CODE')
}
}
