% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_num}
\alias{dummy_num}
\title{Create a dummy variable based on selected values}
\usage{
dummy_num(dat, var, value, opts = "more_less", name = "dummy_num")
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{var}{Variable in data frame to create dummy variable off of}

\item{value}{The set value will depend whether the data is a date, facotr, or numeric. If date, value should be a year, if factor, value should be a level
within the variable, if number value should be a single number or range of numbers [use c(1,5)]}

\item{opts}{Choices are x_y and more_less. x_y sets the selected year, factor, or numeric value (single or range) to 0 and all other values to 1.
more_less sets sets values before the set year or numberic value to 0 and all values greater than the year or value to 1. 
Default is set to 'more_less'.}

\item{name}{Name of created dummy variable. Used in the logging function to reproduce work flow. Defaults to name of the function if not defined.}
}
\description{
Create a dummy variable based on selected values
}
\details{
This function creates a dummy variable. How the dummy variable is created depends upon whether the variable the dummy variable should be created
from is a date, factor, or numeric variable. 
For date variables, the dummy variable is defined by a date (year) and may be either this year versus all other years (x_y) or before vs after this year (more_less).
Use this function to create a variable defining whether a not policy action had been implemented. 
For example, to create an ammendment 80 variable you would type: `dummy_num('MainDataTable', 'Haul_date', 2008, 'more_less', 'ammend80')` 
For factor variables, the only option is to compare selected levels against all others.
For example, to set a variable specifying whether fishers targeted pollock or something else type: `dummy_num('MainDataTable', 'GF_TARGET_FT', c('Pollock - bottom', 'Pollock - midwater'), 'x_y', 'pollock_target')`
For numeric variables, you can set a single value or a range of continuous values and contrast either the selected value(s) against all others (x_y) or less than the 
selected value versus more than the selected value (more_less). For more_less, the mean is used as the critical value is a range of values is proviced.
}
