% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_var}
\alias{dummy_var}
\title{Create a new dummy variable}
\usage{
dummy_var(dataset, DumFill = TRUE)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{DumFill}{Fill the dummy variable with TRUE or FALSE}
}
\description{
Create a new dummy variable
}
\details{
Function for generating new or specialized variables. dummy_var creates a dummy variable and dummy_matrix creates a dummy matrix.
}
