% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_map_gui.R
\name{edit_map_gui}
\alias{edit_map_gui}
\title{Edit spatial table interactively}
\usage{
edit_map_gui(spat, id, project)
}
\arguments{
\item{spat}{Spatial data table to edit.}

\item{id}{String, name of ID column in \code{spat}.}

\item{project}{Name of project.}
}
\description{
Add, delete, or edit a spatial feature interactively
}
