% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{edit_proj_settings}
\alias{edit_proj_settings}
\title{Edit project settings}
\usage{
edit_proj_settings(
  project,
  confid = NULL,
  user_out = NULL,
  tab_name = NULL,
  tab_type = NULL,
  plot_size = NULL,
  save_plot_rds = NULL
)
}
\arguments{
\item{project}{Name of project.}

\item{confid}{List containing new confidentiality settings. See 
\code{\link{set_confid_check}}.}

\item{user_out}{Folder directory containing FishSET output. see 
\code{\link{set_user_locoutput}}.}

\item{tab_name}{Name of table loaded into shiny app.}

\item{tab_type}{Table type. Options include "main", "port", "spat", "grid",
and "aux".}

\item{plot_size}{Plot size (width, height) in inches. Must be numeric.}

\item{save_plot_rds}{Logical, whether to save plot as an RDS file in the 
FishSETFolder ouput folder in addition to save as a PNG. This allows users 
to edit plots at a later time.}
}
\description{
Edit confidentiality settings, user output folder location, default plot 
size, and tables currently used in the shiny app.
}
\examples{
\dontrun{
edit_project_settings("pollock", plot_size = c(5, 4))
}
}
\seealso{
\code{\link{create_proj_settings}} \code{\link{get_proj_settings}}
}
\keyword{internal}
