% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{empty_meta}
\alias{empty_meta}
\title{Helper function
Determines if metadata fields are empty in \code{metadata_gui}.}
\usage{
empty_meta(input, col_nms)
}
\arguments{
\item{input}{Shiny input.}

\item{col_nms}{Column names.}
}
\description{
Helper function
Determines if metadata fields are empty in \code{metadata_gui}.
}
\keyword{internal}
