% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_log_call.R
\name{empty_meta_list}
\alias{empty_meta_list}
\title{Create an empty meta list
Used to create a new metadata entry.}
\usage{
empty_meta_list()
}
\description{
Create an empty meta list
Used to create a new metadata entry.
}
\keyword{internal}
