% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{empty_vars_filter}
\alias{empty_vars_filter}
\title{Check variables are not empty}
\usage{
empty_vars_filter(dat, project, remove = FALSE)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{remove}{Logical, whether to remove empty variables. Defaults to \code{FALSE}.}
}
\value{
Returns the dataset with empty variables removed if \code{remove = TRUE}.
}
\description{
Check for and remove empty variables from dataset. Empty variables are 
columns in the data that contain all NAs and/or empty strings.
}
\details{
Function checks for empty variables and prints an outcome message to 
  the console. If empty variables are present and \code{remove = TRUE}, then 
  empty variables will be removed from the dataset. Empty variables are 
  columns in the dataset that contain all NAs or empty strings.
}
\examples{
\dontrun{
# check for empty vars
empty_vars_filter(pollockMainDataTable)

# remove empty vars from data
mod.dat <- empty_vars_filter(pollockMainDataTable, 'pollock', remove = TRUE)
}
}
