% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_library}
\alias{ensure_library}
\title{Helper function to load packages that may be missing}
\usage{
ensure_library(lib.name)
}
\arguments{
\item{lib.name}{Library name required}
}
\description{
Helper function to load packages that may be missing
}
\keyword{internal}
