% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_normal.R
\name{epm_normal}
\alias{epm_normal}
\title{Expected profit model with normal catch function}
\usage{
epm_normal(starts3, dat, otherdat, alts, project, expname, mod.name)
}
\arguments{
\item{starts3}{Starting values as a numeric vector. The order of parameters in the vector is: \cr
c([\emph{catch-function params}], [\emph{travel-dist params}], 
[\emph{stdev}], [\emph{common scale param}]), \cr
where the length of catch-function parameters is the # of alternatives * # of
catch-function variables, length of travel-distance parameters is the # of
travel-distance variables, length of standard deviation defaults to 1 but alternative-
specific standard deviation values can be specified (length = # of alternatives), and the
common scale parameter is a single value.}

\item{dat}{Data matrix, see output from \code{\link{shift_sort_x}}, alternatives with
distance.}

\item{otherdat}{List that contains other data used in the model, see section 8.4.3 in the 
FishSET user manual for more details (link in the description above):
(1) \strong{'griddat'}: catch-function variables that interact with alternative-specific
catch-function parameters and do not vary across alternatives (e.g., vessel gross tonnage).
(2) \strong{'intdat'}: travel-distance variables that interact with travel-distance
parameters and the distance matrix and do not vary across alternatives.
(3) \strong{'prices'}: price in terms of $/landings units. This is typically a vector with
prices for each observation, but can be a single value representing price for the entire
dataset.}

\item{alts}{Number of alternative choices in model}

\item{project}{Name of project}

\item{expname}{Expected catch table (optional)}

\item{mod.name}{Name of model run for model result output table}
}
\value{
ld: negative log likelihood
}
\description{
Calculate the negative log-likelihood of the expected profit model (EPM) with a normal catch 
function. For more information on the EPM normal model see section 8.4.3 in the FishSET 
user manual.
\url{https://docs.google.com/document/d/1p8mK65uG8yp-HbzCeBgtO0q6DSpKV1Zyk_ucNskt5ug/edit#heading=h.mrt9b1ee2yb8}
}
\details{
This function is called in \code{\link{discretefish_subroutine}} when running an EPM model with
    a normal catch function.
}
