% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_predict.R
\name{epm_predict}
\alias{epm_predict}
\title{EPM predict}
\usage{
epm_predict(project, alts, modname, mod.type, price)
}
\arguments{
\item{project}{Name of project}

\item{alts}{Number of alternatives choices in model}

\item{modname}{Name of saved model to use}

\item{mod.type}{String. Options are weibull, lognormal, normal}

\item{price}{Numeric. Price.}
}
\value{
Returns probability of epm model by choice
}
\description{
Prediction component from epm models called in Policy3, under predict_model_tempNew.m
}
\keyword{internal}
