% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{exp_catch_names}
\alias{exp_catch_names}
\title{Return names of expected catch matrices}
\usage{
exp_catch_names(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Return the names of expected catch matrices saved to the FishSET database.
}
