% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_data}
\alias{expand_data}
\title{Add missing dates and variable combos to `MainDataTable`.}
\usage{
expand_data(
  dataset,
  project,
  date = NULL,
  value,
  sub_date = NULL,
  period = NULL,
  group = NULL,
  facet_by = NULL,
  fun = "sum"
)
}
\arguments{
\item{dataset}{Object containing `MainDataTable`.}

\item{project}{Name of project.}

\item{date}{String, name of date variable to find missing days.}

\item{value}{String, name of value variable to be aggregated by \code{agg_helper}.}

\item{sub_date}{String, name of date variable to subset by.}

\item{group}{String, name of grouping variable(s).}

\item{facet_by}{String, name of variable(s) to be facetted (split).}
}
\description{
Add missing dates and variable combos to `MainDataTable`.
}
\details{
This function expands the data to include missing periods/dates
  and combinations of grouping variables that will be used to aggregate the
  data. Only variables needed to aggregate the data are kept to minimize memory usage.
  If confidentiality checks are turned on, the vessel ID column is included
  as well.
}
\keyword{internal}
