% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{expected_catch_list}
\alias{expected_catch_list}
\title{Get Expected Catch List}
\usage{
expected_catch_list(project, name = NULL)
}
\arguments{
\item{project}{Name of project.}

\item{name}{Name of expected catch table from the FishSET database. The table 
name will contain the string "ExpectedCatch". If \code{NULL}, the default 
table is returned. Use \code{\link{tables_database}} to see a list of 
FishSET database tables by project.}
}
\description{
Returns the Expected Catch list from the FishSET database.
}
