% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_startparams.r
\name{explore_startparams}
\alias{explore_startparams}
\title{Explore starting value parameter space}
\usage{
explore_startparams(project, space, dev, startsr = NULL)
}
\arguments{
\item{project}{String, name of project.}

\item{space}{List of length 1 or length equal to the number of models to be evaluated.
\code{space} is the number of starting value permutations to test (the size of
the space to explore). The greater the \code{dev} argument, the larger the
\code{space} argument should be.}

\item{dev}{List of length 1 or length equal to the number of models to be evaluated.
\code{dev} refers to how far to deviate from the average parameter values when
exploring (random normal deviates). The less certain the average parameters are,
the greater the \code{dev} argument should be.}

\item{startsr}{Optional. List, average starting value parameters for
revenue/location-specific covariates then cost/distance. The best
guess at what the starting value parameters should be (e.g. all
ones). Specify starting value parameters for each model if values should be differetn than ones.
The number of starting value parameters should correspond to the likelihood and data that you want to
test.}
}
\value{
Returns three data frames. \cr
 \tabular{rlll}{
newstart: \tab Chosen starting values with smallest likelihood \cr
saveLLstarts: \tab Likelihood values for each starting value permutation \cr
savestarts: \tab Starting value permutations (corresponding to each saved likelihood value) \cr
    }
}
\description{
Shotgun method to find better parameter starting values by exploring starting value
     parameter space.
}
\details{
Function is used to identify better starting parameters when
  convergence is an issue. For more details on the likelihood functions or data, see \code{\link{make_model_design}}.
  Function calls the model design file and should be used after the \code{\link{make_model_design}}
  function is called. \cr
  If more than one model is defined in the model design file, then starting parameters
  must be defined for each model.
}
\examples{
\dontrun{
Example with only one model specified
results <- explore_startparams('myproject', 15, 3, rep(1,17))

Example with three models specified
results <- explore_startparams('myproject', space = list(15,10,100),
   dev=list(3,3,1), startsr=list(c(1,2,3), c(1,0, -1), c(0,0,.5)))

View results
results$startsOut
}
}
