% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_startparams.r
\name{explore_startparams_discrete}
\alias{explore_startparams_discrete}
\title{Explore starting value parameter space}
\usage{
explore_startparams_discrete(
  space,
  dev,
  breakearly = TRUE,
  max.iterations = 500,
  startsr = NULL,
  fr,
  d,
  otherdat,
  choice,
  project
)
}
\arguments{
\item{space}{List of length 1 or length equal to the number of models to be evaluated.
\code{space} is the number of starting value permutations to test (the size of
the space to explore). The greater the \code{dev} argument, the larger the
\code{space} argument should be.}

\item{dev}{List of length 1 or length equal to the number of models to be evaluated.
\code{dev} refers to how far to deviate from the average parameter values when
exploring (random normal deviates). The less certain the average parameters are,
the greater the \code{dev} argument should be.}

\item{breakearly}{Logical, should the function return the first set of inits that do not return INF or search the entire space 
and return the inits with the lowest LLoglikelihood.}

\item{max.iterations}{If \code{explorestarts = TRUE}, max.iterations indicates the
maximum number of iterations to run in search of valid starting parameter values.
If the maximum is reached before valid parameter values are found (i.e., likelihood 
= Inf) the loop will terminate and an error message will be reported for that model.}

\item{startsr}{Optional. List, average starting value parameters for
revenue/location-specific covariates then cost/distance. The best
guess at what the starting value parameters should be (e.g. all
ones). Specify starting value parameters for each model if values should be different than ones.
The number of starting value parameters should correspond to the likelihood and data that you want to
test.}

\item{fr}{Name of likelihood function to test.}

\item{d}{Data from \code{shift_sort_x}}

\item{otherdat}{Other data (as a list, corresponding to the likelihood function you want to test).}

\item{choice}{Data corresponding to actual zonal choice.}

\item{project}{Project name}
}
\value{
Returns three data frames. \cr
 \tabular{rlll}{
newstart: \tab Chosen starting values with smallest likelihood \cr
saveLLstarts: \tab Likelihood values for each starting value permutation \cr
savestarts: \tab Starting value permutations (corresponding to each saved likelihood value) \cr
    }
}
\description{
Shotgun method to find better parameter starting values by exploring starting value
     parameter space.
}
\details{
Function is used to identify better starting parameters when
  convergence is an issue. For more details on the likelihood functions or data, see \code{\link{make_model_design}}.
  Function calls the model design file and should be used after the \code{\link{make_model_design}}
  function is called. \cr
  If more than one model is defined in the model design file, then starting parameters
  must be defined for each model.
}
\keyword{internal}
