% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_nm_check}
\alias{file_nm_check}
\title{Check name or file name}
\usage{
file_nm_check(file_nm)
}
\arguments{
\item{file_nm}{String, the file name or name to check.}
}
\description{
Check a name or file name for illegal characters.
}
\keyword{internal}
