% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_dat}
\alias{filter_dat}
\title{Remove rows based on filter expressions defined in 'filterTable'}
\usage{
filter_dat(dat, project, exp, filterTable = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{exp}{How to filter. May be a row in the filter table generated by \code{\link{filter_table}} that contains a filter
expression or the filter expression to apply to the data. If the filter expression is supplied, it should take on the form of \code{"x<100"} or \code{"is.na(x) == F"}.}

\item{filterTable}{Name of filter table in FishSET database. Name should contain the phrase 'filterTable'.}
}
\value{
Filtered data frame
}
\description{
Remove rows based on filter expressions defined in 'filterTable'
}
\details{
Filter data frame based on a predefined filter expression from \code{\link{filter_table}} or a filter expression.
We recommend creating a filter table using \code{\link{filter_table}} so that filter expressions are stored and easily
accessed in the future.
}
\examples{
\dontrun{
newdat <- filter_dat(pcodMainDataTable, 'pollock', exp = 3, filterTable = 'pcodfilterTable01012011')
newdat <- filter_dat(pcodMainDataTable, 'pollock', exp = 'PERFORMANCE_Code==1', filteTable = NULL)
}

}
\keyword{filter,}
\keyword{subset}
