% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_data}
\alias{filter_data}
\title{Create dataframe containining customized filters}
\usage{
filter_data(dataset, x, exp, save.filter = FALSE, log.dat = TRUE)
}
\arguments{
\item{dataset}{dataframe or matrix over which to apply filter}

\item{x}{column in dataframe over which to filter will be applied}

\item{exp}{Filter expression. Should take on the form of 'x<100' or is.na(x)==F.}

\item{save.filter}{Whether to save the filterTable as a csv file}

\item{log.dat}{Whether to print filterTable to the log file}

\item{use.filter.Table}{TRUE or FALSE. If true then data is subsetted based on a filter in filterTable and exp is the row containing the filter.}
}
\value{
filter_data returns filterTable into the global environment. The data table will grow with each run of the function.

filter_dat applies user-definted filters. Output must be saved
}
\description{
Contains a function to to define and store filters
 Also contains a function to apply defined filters to dataframe
}
\details{
This function allows users to define and store data filters which can then be applied to the data. The filter dataframe can be saved and will be logged in the log file.
}
\examples{
Generate data for example: outliers inejcted into cars dataset.
cars1 <- cbind(rbind(cars[1:30, ], data.frame(speed=c(19,19,20,20,20), dist=c(190, 186, 210, 220, 218))), ID=rep(c('a','b','c','d','e'), 7))
filter_data(cars1, 'dist','dist>100')
filterTable

newdat <- filter_dat(cars1, exp=1, use.filter.Table=T)
newdat <- filter_dat(cars1, exp='dist>100', use.filter.Table=F)
}
\keyword{filter,}
\keyword{subset}
