% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outsample.R
\name{filter_outsample}
\alias{filter_outsample}
\title{Filter out-of-sample data for model predictions}
\usage{
filter_outsample(
  dat,
  project,
  mod.name,
  spatial_outsample = FALSE,
  zone.dat = NULL,
  spat = NULL,
  zone.spat = NULL,
  outsample_zones = NULL,
  lon.spat = NULL,
  lat.spat = NULL,
  use.scalers = FALSE,
  scaler.func = NULL
)
}
\arguments{
\item{dat}{Out-of-sample data}

\item{project}{Name of project}

\item{mod.name}{Name of saved model to use. Argument can be the name of the model or can pull the name 
of the saved "best" model. Leave \code{mod.name} empty to use the saved "best" model. If more than
one model is saved, \code{mod.name} should be the numeric indicator of which model to use.
Use \code{table_view("modelChosen", project)} to view a table of saved models.}

\item{spatial_outsample}{Logical, indicate whether the data are out-of-sample spatially or not. 
Note that models with zone-specific coefficients (e.g., zonal logit) cannot be used to predict data that are 
out-of-sample spatially. \code{spatial_outsample = FALSE} can represent data out-of-sample temporally or out-of-sample 
based on another variable (e.g., vessel tonnage, gear type, etc.)}

\item{zone.dat}{Variable in \code{dat}that identifies the individual areas or zones.}

\item{spat}{Required, data file or character. 
\code{spat} is a spatial data file containing information on fishery 
management or regulatory zones boundaries. Shape, json, geojson, and csv 
formats are supported. geojson is the preferred format. json files must be 
converted into geoson. This is done automatically when the file is loaded 
with \code{\link{read_dat}} with \code{is.map} set to true. \code{spat} 
cannot, at this time, be loaded from the FishSET database. \cr}

\item{zone.spat}{Variable in \code{spat} that identifies the individual areas or zones.}

\item{outsample_zones}{Vector of out-of-sample zones to filter \code{dat}. Only provided as input when running this function
in the main app.}

\item{lon.spat}{Required for csv files. Variable or list from \code{spat} 
containing longitude data. Leave as NULL if \code{spat} is a shape or json file.}

\item{lat.spat}{Required for csv files. Variable or list from \code{spat} 
containing latitude data.  Leave as NULL if \code{spat} is a shape or json file.}

\item{use.scalers}{Input for \code{create_model_input()}. Logical, should data be normalized? Defaults to \code{FALSE}. Rescaling factors are the mean of the 
numeric vector unless specified with \code{scaler.func}.}

\item{scaler.func}{Input for \code{create_model_input()}. Function to calculate rescaling factors.}
}
\value{
Returns probability of logit model by choice
}
\description{
Filter the out-of-sample dataset and prepare for predictions of fishing probability.
}
\details{
This function filters the out-of-sample data. If the data is out-of-sample spatially, then set \code{spatial_outsample = TRUE} and 
provide a spatial file (\code{spat}) and the zone id in the spatial file \code{zone.spat}. An interactive map is used for selecting out
of sample zones. If the data are not spatially out-of-sample, then just filter the data for the zones included in the selected model. Note that 
models with zone-specific coefficients (e.g., zonal logit) cannot predict spatial out-of-sample data. Upon successful execution of 
\code{filter_outsample()} the filtered dataset will be saved to an RDS file in the outputs folder. This function will overwrite the existing
RDS file each time it is run.
}
