% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_table}
\alias{filter_table}
\title{Define and store filter expressions}
\usage{
filter_table(dat, x, exp, project)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishet_db database should contain the string `MainDataTable`.}

\item{x}{Column in dat aframe over which filter will be applied}

\item{exp}{Filter expression. Should take on the form of `x<100` or `is.na(x)==F`.}

\item{project}{Name of project}
}
\value{
Filter expressions saved as a table into the global environment.
}
\description{
Define and store filter expressions
}
\details{
This function allows users to define and store data filter expressions which can then be applied to the data.  
The filter table will be saved in the SQLite database under the project name (\emph{project}), the date the table was created and the 
string \emph{filterTable}. The new filter functions are added each time the function is run and the table is also automatically updated in the 
fishet_db database. The function call will be logged in the log file.
}
\examples{
\dontrun{  
filter_data(MainDataTable, 'PERFORMANCE_Code','PERFORMANCE_Code==1', 'pcod') 
}

}
\keyword{filter,}
\keyword{subset}
