% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_catch}
\alias{find_catch}
\title{Find columns that may contain catch}
\usage{
find_catch(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may contain catch
}
\keyword{internal}
