% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Generate centroid of polygon of zone or area}
\usage{
find_centroid(use.grid, dataset, gridfile, lon.grid, lat.grid, lat.dat,
  lon.dat, cat, weight.var)
}
\arguments{
\item{use.grid}{TRUE/FALSE}

\item{dataset}{dataframe or matrix}

\item{gridfile}{name of gridded dataset}

\item{lon.grid}{Longitude of points from gridfile}

\item{lat.grid}{Latitude of points from gridfile}

\item{lat.dat}{Latitude of points from dataset}

\item{lon.dat}{Longitude of points from dataset}

\item{cat}{Categorical variable defining the individual areas or zones}

\item{weight.var}{Variable for weighted average}
}
\value{
Returns a dataframe with location of centroid
}
\description{
Generate centroid of polygon of zone or area
}
\details{
Functions returns the center of a zone or area based on set of latitude and longitudes for using in generating distance matrices.
 Function can also return a weighted centroid. Code works for dataframes and shape files. Lists can also be used as long as inputs are a list of latitudes, longitudes, and areas or zones.
 Includes assignment_column function. Assigns a zone from a gridded dataset to the main dataset.
}
\keyword{centroid,}
\keyword{polygon}
\keyword{zone,}
