% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_datetime}
\alias{find_datetime}
\title{Find columns that may be date/datetime}
\usage{
find_datetime(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may be date/datetime
}
\keyword{internal}
