% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_dev}
\alias{find_dev}
\title{Find how many standard deviations point \code{x} is from mean of \code{y}.}
\usage{
find_dev(x, y)
}
\arguments{
\item{x}{value to check}

\item{y}{data vector. Must be numeric}
}
\description{
Find how many standard deviations point \code{x} is from mean of \code{y}.
}
\keyword{internal}
