% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_duration}
\alias{find_duration}
\title{Find columns that may contain durations}
\usage{
find_duration(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may contain durations
}
\keyword{internal}
