% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fishing_centroid.R
\name{find_fishing_centroid}
\alias{find_fishing_centroid}
\title{Create fishing or weighted fishing centroid}
\usage{
find_fishing_centroid(
  dat,
  project = NULL,
  cat = "ZoneID",
  weight.var = NULL,
  lon.dat = NULL,
  lat.dat = NULL,
  spat,
  lon.spat = NULL,
  lat.spat = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Name of project}

\item{cat}{Variable in \code{dat} that identifies zonal assignments or the variable in \code{spat} that identifies the individual areas or zones. 
If \code{spat} is class sf, \code{cat} should be name of list containing information on zones.}

\item{weight.var}{Variable from \code{dat} for weighted average. If \code{weight.var} is defined, the centroid is defined by the latitude and 
longitude of fishing locations in each zone weighted by \code{weight.var}.}

\item{lon.dat}{Required. Longitude variable in \code{dat}.}

\item{lat.dat}{Required. Latitude variable in \code{dat}.}

\item{spat}{Spatial data containing information on fishery management or regulatory zones. Can be shape file, json, geojson, data frame, or list.
\code{spat} should be NULL if zone identifier variable exists in \code{dat}.}

\item{lon.spat}{Variable or list from \code{spat} containing longitude data. Required if \code{spat} is a csv file. 
Leave as NULL if zone identifier exists in \code{dat} or \code{spat} is a shape or json file.}

\item{lat.spat}{Variable or list from \code{spat} containing latitude data. Required if \code{spat} is a csv file. 
Leave as NULL if  zone identifier exists in \code{dat} \code{spat} is a shape or json file.}
}
\value{
Returns primary dataset with fishing centroid and, if \code{weight.var} is specified, the weighted fishing centroid.
}
\description{
Create fishing or weighted fishing centroid
}
\details{
Fishing centroid defines the centroid by mean latitude and longitude of fishing locations in each zone. 
    Weighted centroid defines the centroid by the mean latitude and longitude of fishing locations in each zone weighted by the \code{weight.var}.
    The fishing and weighted centroid variables can be used anywhere latitude/longitude variables appear.
    Each observation in \code{dat} must be assigned to a fishery or regulatory area/zone. 
    If the zone identifier exists in \code{dat} and is not called \code{'ZoneID'}, then \code{cat} 
    should be the variable name containing the zone identifier.
    If a zone identifier variable does not exist in \code{dat},   \code{spat} must be be specified and \code{cat} must be zone 
    identifier in \code{spat}. The \code{assignment_column} function will be run and a zone identifier variable added to \code{dat}.
}
\keyword{centroid,}
\keyword{zone}
