% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_lat}
\alias{find_lat}
\title{Find columns that may be latitude data}
\usage{
find_lat(dat)
}
\arguments{
\item{dat}{Data set to search over}
}
\description{
Find columns that may be latitude data
}
\keyword{internal}
