% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_lon}
\alias{find_lon}
\title{Find columns that may be longitude data}
\usage{
find_lon(dat)
}
\arguments{
\item{dat}{Data set to search over}
}
\description{
Find columns that may be longitude data
}
\keyword{internal}
