% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_lonlat}
\alias{find_lonlat}
\title{Find columns that may be longitude or latitude data}
\usage{
find_lonlat(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may be longitude or latitude data
}
\keyword{internal}
