% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_port}
\alias{find_port}
\title{Find columns that may contain ports}
\usage{
find_port(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may contain ports
}
\keyword{internal}
