% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_project}
\alias{find_project}
\title{Find project}
\usage{
find_project(dat, project = NULL)
}
\arguments{
\item{dat}{Data table name}

\item{project}{Project Name}
}
\description{
Find project
}
\keyword{internal}
