% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_value}
\alias{find_value}
\title{Find columns that may contain value}
\usage{
find_value(dat)
}
\arguments{
\item{dat}{Dataset to search over}
}
\description{
Find columns that may contain value
}
\keyword{internal}
