% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{fishset_tables}
\alias{fishset_tables}
\title{Show all SQL Tables in fishset_db for each project and table type}
\usage{
fishset_tables(project = NULL)
}
\arguments{
\item{project}{Project name. If NULL, tables from all available projects will
be displayed.}
}
\description{
Returns a data frame containing all tables along with their project names
and table type.
}
