% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_outsample_coefs.R
\name{format_outsample_coefs}
\alias{format_outsample_coefs}
\title{Reformat out-of-sample model coefficients}
\usage{
format_outsample_coefs(in_zones, out_zones, Eq, likelihood)
}
\arguments{
\item{in_zones}{Vector of zoneIDs in the in-sample dataset}

\item{out_zones}{Vector of zoneIDs in the out-of-sample dataset}

\item{Eq}{Tibble containing estimated model coefficients (including standard errors and t-values)}

\item{likelihood}{Character, name of the likelihood}
}
\value{
Return a list with (1) vector of coefficients (zones not in the out-of-sample dataset removed) and 
(2) flag indicating if the first alt (in-sample dataset) is not included in the out-of-sample dataset.
}
\description{
Reformat out-of-sample model coefficients by removing zones not included in the out-of-sample dataset
}
