% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{function_summary}
\alias{function_summary}
\title{Display summary of function calls}
\usage{
function_summary(project, date = NULL, type = "dat_load", show = "all")
}
\arguments{
\item{project}{Project name.}

\item{date}{Character string; the date of the log file ("%Y-%m-%d" format) to
retrieve. If \code{NULL} the most recent log is pulled.}

\item{type}{The type of function to display. "dat_load", "dat_quality", "dat_create",
"dat_exploration", "fleet", and "model".}

\item{show}{Whether to display \code{"all"} calls, the \code{"last"} (most recent) call, or
the \code{"first"} (oldest) function call from the log file.}
}
\description{
Display summary of function calls
}
\details{
Displays a list of functions by type and their arguments from a log file.
  If no date is entered the most recent log file is pulled.
}
\examples{
\dontrun{
function_summary()
}
}
\seealso{
\code{\link{filter_summary}}
}
