% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_closure_scenario.R
\name{get_closure_scenario}
\alias{get_closure_scenario}
\title{Retrieve closure scenario by project}
\usage{
get_closure_scenario(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Retrieve closure scenario by project
}
\examples{
\dontrun{
get_closure_scenario("pollock")
}
}
