% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{get_confid_cache}
\alias{get_confid_cache}
\title{Return cached confidentiality tables}
\usage{
get_confid_cache(project, show = "all")
}
\arguments{
\item{project}{Name of project}

\item{show}{Output \code{"all"} tables, \code{"last"} table, or \code{"first"}
table.}
}
\value{
A list of tables containing suppression conditions.
}
\description{
This function lists the confidentiality "check" tables used to suppress values.
}
\seealso{
\code{\link{reset_confid_cache}}
}
