% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{get_confid_check}
\alias{get_confid_check}
\title{Return the confidentiality settings}
\usage{
get_confid_check(project)
}
\arguments{
\item{project}{Name of project}
}
\value{
A list containing the confidentiality parameters: \code{check}, 
  \code{v_id}, \code{rule}, and \code{value}.
}
\description{
This function returns the confidentiality settings from project settings
file.
}
\seealso{
\code{\link{set_confid_check}} \code{\link{get_proj_settings}}
}
