% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_grid.R
\name{get_grid_log}
\alias{get_grid_log}
\title{Retrieve grid log file}
\usage{
get_grid_log(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Retrieves the grid log file for a project. The grid log shows which grid 
files are currently saved to the project data folder.
}
\details{
The grid log is a list containing information about the grid files 
  currently saved to the project data folder. Each grid entry contains three 
  fields: \code{grid_name}, \code{closure_name}, and \code{combined_areas}. 
  \code{grid_name} is the name of the original grid object. If the other two
  fields are empty, this means that the grid file has not been altered and 
  is the same as the original. \code{closure_name} is the name of a second 
  grid file containing closure areas that were combined with \code{grid_name}.
   \code{combined_areas} are the names/IDs of the closures areas from the 
   closure grid file that were combined with \code{grid_name}.
}
\examples{
\dontrun{
get_grid_log("pollock")
}
}
