% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_latest_projectfile}
\alias{get_latest_projectfile}
\title{Pull data from latest project file}
\usage{
get_latest_projectfile(project, mod.name)
}
\arguments{
\item{project}{Project name}

\item{mod.name}{Model name}
}
\description{
Pull data from latest project file
}
\examples{
\dontrun{
get_latest_projectfile("pollock", "logit_mod1")
}
}
