% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{get_proj_settings}
\alias{get_proj_settings}
\title{Retrieve project settings}
\usage{
get_proj_settings(project, format = FALSE)
}
\arguments{
\item{project}{Name of project.}

\item{format}{Logical, output project settings using \code{\link{pander}}. Useful
for markdown documents.}
}
\description{
Retrieve project settings
}
\details{
The project settings file includes confidentiality settings, the 
user output folder directory, and the default plot saving size.
}
