% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{get_user_locoutput}
\alias{get_user_locoutput}
\title{Print user folder directory}
\usage{
get_user_locoutput(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Print user folder directory
}
\details{
This function prints the local user directory saved in the project
 settings file. This directory path is used for inserting plots and tables 
 from a folder outside the FishSET package into the FishSET RMarkdown Template.
}
\keyword{internal}
