% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{globalcheck_view}
\alias{globalcheck_view}
\title{View error output from discrete choice model}
\usage{
globalcheck_view(table)
}
\arguments{
\item{table}{Table name in sqlite database. Should contain the phrase modelout.}
}
\description{
Returns error output from running the discretefish_subroutine function. The table parameter must be the full name of the table name in the fishet_db database.
}
\examples{
\dontrun{
globalcheck_view('pcodldglobalcheck20190604')
}
}
