% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_grid.R
\name{grid_lab_helper}
\alias{grid_lab_helper}
\title{Labeling function for saving grid files}
\usage{
grid_lab_helper(project, grid_info, grid_log = NULL, mod_type = "combine")
}
\arguments{
\item{project}{Name of project.}

\item{grid_info}{List containing grid information.}

\item{grid_log}{Optional, the grid log. if \code{NULL}, uses names from 
\code{grid_info}.}

\item{mod_type}{String, "combine" for combined map files or "edit" for 
edited map files.}
}
\description{
Labeling function for saving grid files
}
\keyword{internal}
