% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gridcheck}
\alias{gridcheck}
\title{Check that spatial data is a sf object. Convert if not.}
\usage{
gridcheck(
  spatialdat,
  catdat,
  londat = NULL,
  latdat = NULL,
  lon.grid = NULL,
  lat.grid = NULL
)
}
\arguments{
\item{spatialdat}{The spatial dataframe}

\item{catdat}{Variable that names polygons}

\item{londat}{Longitude data from primary dataset}

\item{latdat}{Latitude data from primary dataset}

\item{lon.grid}{Variable in spatialdat containing longitude data}

\item{lat.grid}{Variable in spatialdat containing latitude data}
}
\description{
Check that spatial data is a sf object. Convert if not.
}
\keyword{internal}
