% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{group_perc}
\alias{group_perc}
\title{Create a within-group percentage variable}
\usage{
group_perc(
  dat,
  project,
  id_group,
  group = NULL,
  value,
  name = "group_perc",
  create_group_ID = FALSE,
  drop_total_col = FALSE
)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{String, project name.}

\item{id_group}{String, primary grouping variable(s). Used to create the "total_value" 
variable which sums \code{value} by \code{id_group}. If \code{group = NULL}, then 
\code{value} is divided by "total_value".}

\item{group}{String, secondary grouping variable(s). Used to create the "group_total" 
variable which sums \code{value} by \code{id_group} and \code{group}. Percentage 
is calculated by dividing "group_total" by "total_value". Defaults to \code{NULL}.}

\item{value}{String, the value variable used to calculate percentage. Must be numeric.}

\item{name}{String, the name for the new variable. Defaults to "group_perc".}

\item{create_group_ID}{Logical, whether to create a group ID variable using \code{\link{ID_var}}.
Defaults to \code{FALSE}.}

\item{drop_total_col}{Logical, whether to remove the "total_value" and "group_total" variables
created to calculate percentage. Defaults to \code{FALSE}.}
}
\description{
Create a within-group percentage variable
}
\details{
\code{group_perc} creates a within-group percentage variable using a primary
  group ID (\code{id_group}) and secondary group (\code{group}). The total value of 
  \code{id_group} is stored in the "total_value" variable, and the within-group total
  stored in "group_total". The group percentage is calculated using these two function-created
  variables. "total_value" and "group_total" can be dropped by setting \code{drop_total_col = TRUE}.
  A group ID column can be created using the variables in\code{id_group} and \code{group} by setting 
  \code{create_group_ID = TRUE}.
}
\examples{
\dontrun{
group_perc(pollockMainDataTable, "pollock", id_group = "PERMIT", group = NULL, 
           value = "OFFICIAL_TOTAL_CATCH_MT")
           
group_perc(pollockMainDataTable, "pollock", id_group = "PERMIT",
           group = "DISEMBARKED_PORT", value = "HAUL")
}
}
