% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDVar.R
\name{IDvar}
\alias{IDvar}
\title{Create ID variable from another variable.}
\usage{
IDvar(dataset, newID, ...)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{newID}{name of new ID column}

\item{...}{Column(s) that define the variable(s) over which to make new ID}
}
\value{
Returns the dataframe with new ID variable included
}
\description{
Create ID variable from another variable.
}
\details{
Creates an variable to indicate distinct hauls or trips
}
\examples{
dat <- IDvar(MainDataTable, newID='PermiteID','GEAR_TYPE')
}
