% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{insert_plot}
\alias{insert_plot}
\title{Insert plot from user folder}
\usage{
insert_plot(out, project)
}
\arguments{
\item{out}{String, plot file name.}

\item{project}{Name of project.}
}
\description{
Insert plot from user folder
}
\examples{
\dontrun{
insert_plot("pollock_plot.png")
}
}
