% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Empty variable check}
\usage{
is_empty(x, trim = TRUE, ...)
}
\arguments{
\item{x}{x}

\item{trim}{defaults to true}

\item{...}{Additional arguments}
}
\description{
Empty variable check
}
\keyword{internal}
