% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{is_invalid_spat}
\alias{is_invalid_spat}
\title{Detect basic spatial issues}
\usage{
is_invalid_spat(spat)
}
\arguments{
\item{spat}{Spatial data to check.}
}
\value{
\code{TRUE} if a "GEOMETRYCOLLECTION" is found, if any spatial 
features are not "POLYGON" or "MULTIPOLYGON", if invalid geometries are 
found, if any empty geometries are detected, or if longitude needs to be
shifted to Pacific view.
}
\description{
Predicate function that returns \code{TRUE} if certain spatial issues are
found.
}
\keyword{internal}
