% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{is_leaf}
\alias{is_leaf}
\title{Is list object a leaf node}
\usage{
is_leaf(x)
}
\arguments{
\item{x}{An object in a list.}
}
\value{
\code{TRUE} if list object is a leaf node, \code{FALSE} if not.
}
\description{
Determines whether an object in a list is a leaf node.
}
\keyword{internal}
