% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{is_leaf_table}
\alias{is_leaf_table}
\title{Detect potential tables in list node}
\usage{
is_leaf_table(x)
}
\arguments{
\item{x}{An object in a list.}
}
\value{
\code{TRUE} if leaf node can be converted to dataframe, \code{FALSE} if not.
}
\description{
Determines whether an object in a list can be converted into a dataframe.
}
\keyword{internal}
