% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_value_empty}
\alias{is_value_empty}
\title{Empty value check}
\usage{
is_value_empty(x)
}
\arguments{
\item{x}{A value, input, or argument to check.}
}
\description{
Empty value check
}
\keyword{internal}
