% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{jitter_lonlat}
\alias{jitter_lonlat}
\title{Jitter longitude and latitude variables}
\usage{
jitter_lonlat(dat, project, lon, lat, factor = 1, amount = NULL)
}
\arguments{
\item{dat}{Primary data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Project name.}

\item{lon}{String, variable name containing longitude.}

\item{lat}{String, variable name containing latitude.}

\item{factor}{Numeric, see \code{\link[base]{jitter}} for details.}

\item{amount}{Numeric, see \code{\link[base]{jitter}} for details. 
Default (NULL): factor * d/5 where d is about the smallest difference between x values.}
}
\description{
Jitter longitude and latitude variables
}
\details{
This is one of the FishSET confidentiality functions. It "jitters" 
  longitude and latitude using the base R function \code{\link[base]{jitter}}.
}
\examples{
\dontrun{
jitter_lonlat(pollockMainDataTable, "pollock",
              lon = "LonLat_START_LON", lat = "LonLat_START_LAT")
}
}
