% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_zone.R
\name{lag_zone}
\alias{lag_zone}
\title{Lag zone variable}
\usage{
lag_zone(
  dat,
  project,
  spat = NULL,
  port,
  port_name,
  port_lon,
  port_lat,
  trip_id,
  haul_order,
  starting_port,
  zoneID_dat,
  zoneID_spat = NULL,
  name = "startingloc"
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Name of project}

\item{spat}{Spatial data. Required if \emph{ZoneID} does not exists in \code{dat}.
Shape, json, geojson, and csv formats are supported.}

\item{port}{Port data. Contains columns: Port_Name, Port_Long, Port_Lat. 
Table is generated using the \code{\link{load_port}} and saved in the FishSET 
database as the project and port table, for example 'pollockPortTable'.}

\item{port_name}{String indicating the column in port table that contains the port name}

\item{port_lon}{String indication the column in port table that contains port longitude}

\item{port_lat}{String indication the column in port table that contains port latitude}

\item{trip_id}{Variable in \code{dat} that identifies unique trips.}

\item{haul_order}{Variable in \code{dat} containing information on the order 
that hauls occur within a trip. Can be time, coded variable, etc.}

\item{starting_port}{Variable in \code{dat} to identify port at start of trip.}

\item{zoneID_dat}{Variable in \code{dat} that identifies the individual zones or 
areas.}

\item{zoneID_spat}{Variable in \code{spat} that identifies the individual zones or 
areas.}

\item{name}{String, name of created variable. Defaults to name of the function 
if not defined.}
}
\value{
Primary data set with starting location variable added.
}
\description{
Creates a lagged zone ID variable for haul data, where the lagged location of the first haul
is filled in with the port of departure. This variable is required for data with multiple sets
or hauls in a single trip and for the full information model with 
Dahl's correction (\code{\link{logit_correction}}).
}
\details{
Function creates the \code{startloc} vector that is required for multihaul data
  and the full information model with Dahl's correction \code{\link{logit_correction}}. 
  The vector is the lagged zone ID - zone when the decision of where to fish 
  next was made. Generally, the first zone of a trip is the port of departure. 
  The \code{\link{assignment_column}} function is called to assign starting port 
  locations and haul locations to zones. If ZoneID exists in \code{dat}, 
  \code{\link{assignment_column}} is not called and the following arguments are 
  not required: \code{spat, zoneID_spat}.
}
\examples{
\dontrun{
pcodMainDataTable <- lag_zone(pcodMainDataTable, 'pcod',
    map2, "pcodPortTable", "TRIP_SEQ", "HAUL_SEQ", "DISEMBARKED_PORT", 
 "START_LON", "START_LAT", "NMFS_AREA", "STARTING_LOC"
)
}
}
