% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{lame_list}
\alias{lame_list}
\title{Detect an unnecessary nested list}
\usage{
lame_list(l)
}
\arguments{
\item{l}{A list.}
}
\value{
\code{TRUE} if list object is a lame list, \code{FALSE} if not.
}
\description{
If a list object contains an unnamed list containing another list or a single 
object, it is a "lame list". For example: \code{list(list(1:10)))}. The
nested list in this example is unnecessary since it contains a single vector, 
and therefore can be removed. Pruning lame lists can help simplify a list object.
}
\examples{
\dontrun{
lame_list(list(list(1:10)))
lame_list(list(A = list(1:10)))
lame_list(list(list(1:10), list(11:20)))
}
}
\seealso{
\code{\link{clean_list}}
}
\keyword{internal}
