% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{layout.json.ed}
\alias{layout.json.ed}
\title{Log all actions and informative messages in an exterior file.}
\usage{
layout.json.ed(level, fun.name, dataset, x, msg = "")
}
\arguments{
\item{fun.name}{name of the function being applied}

\item{dataset}{dataframe or matrix}

\item{x}{column in dataframe}
}
\value{
Log file
}
\description{
Exterior file is dated to the date the code is run
 Requires the futile.library package and functions built from the package
}
\details{
Calls the futile.logger package. Calls to log messages are saved in the specified dated file.

Messages can be saves as informational (flog.info), as a record of actions taken (flog.trace), as error or warning messages (flog.error, flog.warn), as debug message (flog.debug).

The flog_func function is used within other functions to log actions taken. These trace messages are saved to an external file

The layout.json.ed writes the logged messages in a json format.
}
\keyword{logging}
