% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{list_depth}
\alias{list_depth}
\title{Find list depth}
\usage{
list_depth(l)
}
\arguments{
\item{l}{A list.}
}
\value{
A numeric vector of list depths.
}
\description{
Iterates \code{purrr::vec_depth()} over each object in a list.
}
\keyword{internal}
