% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{list_df}
\alias{list_df}
\title{Convert leaf node to a dataframe}
\usage{
list_df(l)
}
\arguments{
\item{l}{A list.}
}
\value{
A list.
}
\description{
Determines whether a leaf node contains an object that can be converted to
a dataframe.
}
\examples{
\dontrun{
list_df(list(A = list(X = 1:10, Y = letters[1:10])))
list_df(list(A = 1:10, B = "Text", C = c("text", "text"))) # no conversion
}
}
\seealso{
\code{\link{is_leaf_table}} \code{\link{simplify_list}}
}
\keyword{internal}
