% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_dirs}
\alias{list_dirs}
\title{Find directories}
\usage{
list_dirs(
  path = ".",
  pattern = NULL,
  all.dirs = FALSE,
  full.names = FALSE,
  ignore.case = FALSE
)
}
\arguments{
\item{path}{file path}

\item{pattern}{Use to define specific dirs to search for}

\item{all.dirs}{Logical}

\item{full.names}{Logical, Full name or directory name}

\item{ignore.case}{Logical}
}
\description{
Find directories
}
\keyword{internal}
