% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_html}
\alias{list_html}
\title{Create an HTML list item}
\usage{
list_html(l)
}
\arguments{
\item{l}{A list.}
}
\value{
An HTML string of a list item.
}
\description{
Converts a list object to an HTML list item (<li>). If an object is a dataframe
or matrix, it is converted to an HTML table first.
}
\keyword{internal}
