% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_html_r}
\alias{list_html_r}
\title{Create an un-ordered HTML list recursively}
\usage{
list_html_r(l)
}
\arguments{
\item{l}{A list.}
}
\value{
A list of HTML list item strings.
}
\description{
Applies \code{\link{list_html}} to every level of a list object.
}
\keyword{internal}
