% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{list_length}
\alias{list_length}
\title{Find list length}
\usage{
list_length(l)
}
\arguments{
\item{l}{A list.}
}
\value{
A numeric vector of list lengths.
}
\description{
Iterates \code{length()} over each object in a list.
}
\keyword{internal}
