% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{list_logs}
\alias{list_logs}
\title{Lists all logs files}
\usage{
list_logs(chron = FALSE)
}
\arguments{
\item{chron}{Logical, whether to display logs in chronological order (TRUE) or
reverse chronological order (FALSE).}
}
\description{
Lists all logs files
}
