% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{list_tables}
\alias{list_tables}
\title{Display tables in FishSET database by type for the defined project}
\usage{
list_tables(project, type = "main")
}
\arguments{
\item{project}{A project name to show main tables by.}

\item{type}{the type of fishset_db table to search for. Options include 
"main" (MainDataTable), "port" (PortTable), "spat" (SpatTable), "grid" 
(GridTable), "aux" (AuxTable) "ec" (ExpectedCatch),  "altc" (altmatrix), 
"info" (MainDataTableInfo), "gc" (ldglobalcheck), "fleet" (FleetTable), 
"filter" (FilterTable),  "model" (modelOut), "model_data" (modelinputdata), 
and "model_design" (modelDesignTable).}
}
\description{
Display tables in FishSET database by type for the defined project
}
\examples{
\dontrun{
list_tables(type = "main")
list_tables("pollock", "ec")
}

}
