% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_to_html}
\alias{list_to_html}
\title{Convert list to HTML}
\usage{
list_to_html(l)
}
\arguments{
\item{l}{A list.}
}
\value{
An un-ordered HTML list.
}
\description{
Converts a list to an un-ordered HTML list (<ul>).
}
\examples{
\dontrun{
  l <- list(A = 'a text', B = list(b1 = 'b1 text', b2 = 'b2 text'),
            C = list(data.frame(A = 1:10, B = letters[1:10])))
  list_to_html(l)
}
}
\seealso{
\code{\link{simplify_list}} \code{\link{collapse_leaf_r}} 
  \code{\link{list_html_r}}
}
\keyword{internal}
