% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_data}
\alias{load_data}
\title{Load data from FishSET database into the R environment}
\usage{
load_data(project, name = NULL)
}
\arguments{
\item{project}{String, name of project.}

\item{name}{Optional, name of table in FishSET database. Use this argument
if pulling raw or dated table (not the working table).}
}
\value{
Data loaded to working environment as the project and ‘MainDataTable’.
}
\description{
Load data from FishSET database into the R environment
}
\details{
Pulls the primary data table from the FishSET database and loads it into the working 
  environment as the project and MainDataTable. For example, if the project was pollock, 
  then data would be saved to the working environment as 'pollockMainDataTable'.
}
\examples{
\dontrun{
load_data('pollock')

load_data('pollock', 'pollockMainDataTable20190101')
}

}
