% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_grid}
\alias{load_grid}
\title{Save gridded data}
\usage{
load_grid(dataset, x, over_write = TRUE, project = NULL)
}
\arguments{
\item{x}{Name of gridded data frame to be saved}

\item{over_write}{TRUE/FALSE Save over previously saved file or not.}

\item{project}{Name of project for attaching to name of the table.}

\item{dat}{Main data frame. Table in fishset_db database should contain the string `MainDataTable`.}
}
\description{
Save gridded data
}
\details{
Grid data is an optional data frame that contains a variable that varies by the map grid (ex. sea surface temperature, wind speed). It can also vary by a second dimension (e.g., date/time). Both dimensions in the gridded data file need to be variables included in the main data frame. 
The grid locations (zones) must define the columns and the optional second dimension defines the rows. The row variable must have the exact name as the variable 
in the main data frame that it will be linked to. The data is saved in the fishset_db database as the raw data and the working data. In both cases, the table name is the `project`, if defined, and the file name `x`. Date is also attached to the name for the raw data.
}
\examples{
\dontrun{  
load_grid(dataset='pcodMainDataTable', x=SeaSurfaceTemp, over_write=TRUE, project='pcod') 
}
}
