% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gui_variables.R
\name{load_gui_variables}
\alias{load_gui_variables}
\title{load_gui_variables}
\usage{
load_gui_variables(project_name, folderpath)
}
\arguments{
\item{project_name}{The name of the current project.}

\item{folderpath}{The file path to the project's root folder.}
}
\value{
A list of saved variables if the file exists, otherwise returns NULL.
}
\description{
A helper function to load selected variables from an RDS file in the 
FishSET GUI.
}
