% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_maindata}
\alias{load_maindata}
\title{Import, parse, and save data to the FishSET database}
\usage{
load_maindata(dat, project, over_write = TRUE, compare = FALSE, y = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Should be the full path to the file or name of table in the FishSET database.
 Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{over_write}{Logical, If TRUE, saves data over previously saved data table in the FishSET database.}

\item{compare}{Logical, whether to compare new data frame to previously saved data frame \code{y}.}

\item{y}{Name of previously saved table in FishSET database. \code{y} must be defined if \code{compare} is TRUE.}
}
\description{
Import, parse, and save data to the FishSET database
}
\details{
Runs the \code{fishset_compare} function if \code{compare} is TRUE and calls the
 \code{\link{data_verification}} function to check for common data issues and that latitude
 and longitude are defined. The dataset is saved in the FishSET database as raw and working tables. 
 The table name is the \code{project} and the table type, 'MainDataTable'. Date is also attached to the name for the
 raw data. The main data is also loaded into the working environment as ‘projectMainDataTable’.
}
\examples{
\dontrun{
load_maindata(dat = Mydata, over_write = TRUE, project = 'pollock',
              compare = TRUE, y = 'MainDataTable01012011')
}

}
