% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_seasonal}
\alias{load_seasonal}
\title{Save auxilliary data}
\usage{
load_seasonal(x, y, compare, project)
}
\arguments{
\item{x}{name dataframe to be saved}

\item{y}{name of previously saved dataframe}

\item{compare}{TRUE/FALSE Compare new dataframe to previously saved dataframe before saving dataframe x to databases}

\item{project}{name of project for attaching to table}
}
\description{
Save auxilliary data
}
\details{
Runs a series of checks on the port data. If checks pass, runs the fishset_compare function and save the new dataframe x to the database.
}
