% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loc_data}
\alias{loc_data}
\title{Define source location for data folder
Returns the location of the data folder}
\usage{
loc_data(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Define source location for data folder
Returns the location of the data folder
}
\details{
if loc2 is not in the working environment, then the default location is use
}
\keyword{internal}
