% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loc_doc}
\alias{loc_doc}
\title{Define source location for doc folder
Returns the location of the doc folder}
\usage{
loc_doc(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Define source location for doc folder
Returns the location of the doc folder
}
\details{
if loc2 is not in the working environment, then the default location is use
}
\keyword{internal}
