% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loc_map}
\alias{loc_map}
\title{Define source location for MapViewer folder
Returns the location of the MapViewer folder}
\usage{
loc_map(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Define source location for MapViewer folder
Returns the location of the MapViewer folder
}
\details{
if loc2 is not in the working environment, then the default location is use
}
\examples{
\dontrun{
loc_map() # will return output folder location within the fishset package
loc2 <- getwd()
loc_map() #will return output folder location as within the working directory
}
}
\keyword{internal}
