% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loc_meta}
\alias{loc_meta}
\title{Define source location for meta file}
\usage{
loc_meta(project)
}
\arguments{
\item{project}{Project name.}
}
\description{
Define source location for meta file
}
\keyword{internal}
