% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locdatabase}
\alias{locdatabase}
\title{Returns the location of the database}
\usage{
locdatabase()
}
\description{
Returns the location of the database
}
\details{
if loc is not in the working directory, then the default location is use
}
\examples{
\dontrun{
locdatabase() # will return sqlite database location within the fishset package
loc <- getwd()
locdatabase() #will return sqlite database location as within the working directory
}
}
