% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loclog}
\alias{loclog}
\title{Returns the location of the log folder}
\usage{
loclog(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Returns the location of the log folder
}
\details{
if loc2 is not in the working environment, then the default location is use
}
\examples{
\dontrun{
loclog('pollock') # will return log folder location for project pollock within the fishset package
}
}
\keyword{internal}
