% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reset.R
\name{log_call}
\alias{log_call}
\title{Update function calls saved in log file}
\usage{
log_call(project, fun.name)
}
\arguments{
\item{project}{Name of project}

\item{fun.name}{Function name}
}
\description{
Update function calls saved in log file
}
\details{
update log file
}
\keyword{internal}
