% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_grid.R
\name{log_grid_info}
\alias{log_grid_info}
\title{Log grid file}
\usage{
log_grid_info(project, grid_info, mod_type = "combine")
}
\arguments{
\item{project}{Name of project.}

\item{grid_info}{List containing grid information.}

\item{mod_type}{String, "combine" for combined map files or "edit" for 
edited map files.}
}
\description{
Writes grid information to a JSON file located in the project data directory.
}
\details{
The grid log is a list containing information about the grid files 
  currently saved to the project data folder. Each grid entry contains three 
  fields: \code{grid_name}, \code{closure_name}, and \code{combined_areas}. 
  \code{grid_name} is the name of the original grid object. If the other two
  fields are empty, this means that the grid file has not been altered and 
  is the same as the original. \code{closure_name} is the name of a second 
  grid file containing closure areas that were combined with \code{grid_name}.
   \code{combined_areas} are the names/IDs of the closures areas from the 
   closure grid file that were combined with \code{grid_name}.
}
\seealso{
\code{\link{save_grid_cache}}
}
\keyword{internal}
