% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rerun.R
\name{log_rerun}
\alias{log_rerun}
\title{Console function for rerunning project log}
\usage{
log_rerun(
  log_file,
  dat = NULL,
  portTable = NULL,
  aux = NULL,
  gridfile = NULL,
  spat = NULL,
  ind = NULL,
  run = FALSE
)
}
\arguments{
\item{log_file}{String, name of the log file starting with the date (YYYY-MM-DD) and
ending in ".json".}

\item{dat}{String, new main data table to rerun log}

\item{portTable}{String, name of port table. Defualts to NULL.}

\item{aux}{String, name of auxiliary table. Defaults to NULL.}

\item{gridfile}{String, name of gridded data table. Defaults to NULL.}

\item{spat}{String, name of spatial data table. Defaults to NULL.}

\item{ind}{Numeric, indices of function calls to rerun.}

\item{run}{Logical, whether to run the logged function calls (TRUE) or simply
list all function calls (FALSE).}
}
\description{
Console function for rerunning project log
}
\examples{
\dontrun{
log_rerun("pollock_2020-10-23.json", run = TRUE) # reruns entire log with original data table
# runs log with new data table
log_rerun("pollock_2020-10-23.json", dat = "pollockMainDataTable", run = TRUE) 
}
}
\seealso{
\code{\link{log_rerun_gui}}
}
