% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rerun.R
\name{log_rerun_gui}
\alias{log_rerun_gui}
\title{Interactive function for rerunning project log}
\usage{
log_rerun_gui()
}
\description{
Interactive function for rerunning project log
}
\examples{
\dontrun{
log_rerun_gui()
}
}
\seealso{
\code{\link{log_rerun}}
}
