% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_avgcat.R
\name{logit_avgcat}
\alias{logit_avgcat}
\title{logit_avgcat}
\usage{
logit_avgcat(starts3, dat, otherdat, alts, project, expname, mod.name)
}
\arguments{
\item{starts3}{Starting values with dimensions equal to (number of alternatives - 1)*(number of grid-varying variables) + (number of interactions).
Recall the user must drop one alternative for identification. e.g. c([grid-varying variables -1], [interaction variables]).}

\item{dat}{Data matrix, see output from shift_sort_x, alternatives with distance by column bind}

\item{otherdat}{Other data used in model (as list). Any number of grid-varying variables (e.g. variables that affect catches across locations) or 
interaction variables (e.g. vessel characteristics that affect how much disutility is suffered by traveling a greater distance) are allowed. \cr \cr
However, the user must place these in `otherdat` as list objects named `griddat` and `intdat` respectively. Note the variables
within `griddat` and `intdat` have no naming restrictions. \cr \cr
Also note that `griddat` variables are dimension *(number of observations) x 1*, #' to be interacted with each alternative, 
while `intdat` variables are dimension *(number of observations) x 1*, to be interacted with the distance to each alternative. \cr \cr
If there are no other data, the user can set `griddat` as ones with dimension *(number of observations) x 
(number of alternatives)* and `intdat` variables as ones with dimension *(number of observations) x 1*.}

\item{alts}{Number of alternative choices in model}

\item{project}{Name of project}

\item{expname}{Expected catch table}

\item{mod.name}{Name of model run for model result output table}
}
\value{
ld - negative log likelihood
}
\description{
Average catch conditional logit procedure
}
