% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_correction.r
\name{logit_correction}
\alias{logit_correction}
\title{Full information model with Dahl's correction function}
\usage{
logit_correction(starts3, dat, otherdat, alts, project, expname, mod.name)
}
\arguments{
\item{starts3}{Starting values as a vector (num). For this likelihood, the order takes:
c([marginal utility from catch], [catch function parameters], [polynomial starting parameters], [cost (distance) parameters], [catch sigma]). \cr \cr
The number of polynomial interaction terms is currently set to 2, so given the chosen degree 'polyn' there should be (((polyn+1)*2) + 2)*kk polynomial
starting parameters, where kk equals the number of alternatives. The marginal utility from catch and catch sigma are of length 1 respectively.
The catch function and cost parameters are of length (# of catch variables)*kk and (# of cost variables) respectively.}

\item{dat}{Data matrix, see output from shift_sort_x, alternatives with distance.}

\item{otherdat}{Other data used in model (as list containing objects griddat, intdat, startloc, polyn, and distance). \cr \cr
For grid-specific variables (griddat) and cost variables to be interacted with distance (intdat), any number of variables are allowed, as a list of matrices. 
Note the variables (each as a matrix) within `griddat` and `intdat` have no naming restrictions. 
Also note that `griddat` variables are dimension *(number of observations) x (number of alternatives)*, 
while `intdat` variables are dimension *(number of observations) x 1*, to be interacted with the distance to each alternative.
Grid-specific variables may correspond to catches that vary by location, 
or interaction variables may be vessel characteristics that affect how much disutility is suffered by traveling a greater distance.
Note in this likelihood the grid-specific variables are the variables in the catch equation, and
each variable varies across observations but not for each location: they are grid-specific due to the location-specific coefficients. 
If there are no other data, the user can set `griddat` as ones with dimension *(number of observations) x (number of alternatives)*
and `intdat` variables as ones with dimension *(number of observations) x 1*. \cr \cr
The variable startloc is a matrix of dimension *(number of observations) x 1*, 
that corresponds to the starting location when the agent decides between alternatives. \cr \cr
The variable polyn is a vector of length 1 corresponding to the chosen polynomial degree. \cr \cr
The variable distance is a matrix of dimension *(number of observations) x (number of alternatives)* corresponding to the distance to each alternative.}

\item{alts}{Number of alternative choices in model as length 1 vector (num).}

\item{project}{Name of project}

\item{expname}{Expected catch table}

\item{mod.name}{Name of model run for model result output table}
}
\value{
ld: negative log likelihood
}
\description{
Full information model with Dahl's correction function
}
\examples{
\dontrun{
data(zi)
data(catch)
data(choice)
data(distance)
data(si)
data(startloc)

optimOpt <- c(1000,1.00000000000000e-08,1,0)

methodname <- "BFGS"

polyn <- 3
kk <- 4

si2 <- sample(1:5,dim(si)[1],replace=TRUE)
zi2 <- sample(1:10,dim(zi)[1],replace=TRUE)

otherdat <- list(griddat=list(si=as.matrix(cbind(si,si,si,si)),si2=as.matrix(cbind(si2,si2,si2,si2))),
			intdat=list(zi=as.matrix(zi),zi2=as.matrix(zi2)),startloc=as.matrix(startloc),polyn=polyn,distance=as.matrix(distance))

initparams <- c(3, 0.5, 0.4, 0.3, 0.2, 0.55, 0.45, 0.35, 0.25, rep(0, (((polyn+1)*2) + 2)*kk), -0.3,-0.4, 3)

func <- logit_correction

results <- discretefish_subroutine(catch,choice,distance,otherdat,initparams,optimOpt,func,methodname)
}
}
