% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{lonlat_to_centroid}
\alias{lonlat_to_centroid}
\title{Assign longitude and latitude points to zonal centroid}
\usage{
lonlat_to_centroid(dat, project, lon, lat, spat, zone)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Project name.}

\item{lon}{String, variable name containing longitude.}

\item{lat}{String, variable name containing latitude.}

\item{spat}{Spatial data table containing regulatory zones. This can be 
a "spatial feature" or sf object.}

\item{zone}{String, column name contain the assigned zone. Must be the same 
for both the spatial data table and MainDataTable.}
}
\description{
Assign longitude and latitude points to zonal centroid
}
\details{
This is one of the FishSET confidentiality functions. It replaces the 
  selected longitude and latitude columns with the zonal centroid derived 
  from a spatial data table.
}
\examples{
\dontrun{
lonlat_to_centroid(pollockMainDataTable, spatdat, 
                  lon = "LonLat_START_LON", lat = "LonLat_START_LAT",
                  zone = "NMFS_AREA")
}

}
