% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{main_mod}
\alias{main_mod}
\title{Modify the MainDataTableInfo table}
\usage{
main_mod(dat, x, new.unit = NULL, new.type = NULL, new.class = NULL)
}
\arguments{
\item{dat}{MainDataTableInfo table. Table in fishset_db database should contain the string `MainDataTableInfo`.}

\item{x}{Name of variable in the MainDataTableInfo table that is to be modified.}

\item{new.unit}{Units. Current units include fathoms, decimal degrees, dollars, lbs, metric tons, min, numeric, percent, WK, 'Y/N', yyyymmdd.}

\item{new.type}{General type. Current categories include Time, Flag, Code, Latitude, Code String, Other Numeric, Code Numeric}

\item{new.class}{New specialized variable category. For example, isCPUE, isLon, isLat, isValue, isZoneArea, isPort.}
}
\description{
Modify the MainDataTableInfo table
}
\details{
Modify the units `new.unit`, data format `new.type`, and specialized variable from old category `old.class` to new category `new.class` of the MainDataTableInfo table.
Updated MainDataTableInfo file is saved to the fishset_db database. It is advisable to use the working table and not the raw table as the modifications will automatically be saved over the input table name.
}
\examples{
\dontrun{  
main_mod(dataset='MainDataTableInfo01012011', x='DISEMBARKED_PORT', 
         new.unit='yyyymmdd', new.type='Other',  new.class='isPort') 
}
}
