% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spat_area.R
\name{make_polygon}
\alias{make_polygon}
\title{create a polygon}
\usage{
make_polygon(coord)
}
\arguments{
\item{coord}{Longitude and latitude coordinates forming a polygon. Can be a 
numeric vector of even length or a numeric matrix with two columns.}
}
\description{
create a polygon
}
\details{
If the coordinates create an open polygon the function automatically
closes it and gives a warning.
}
\keyword{internal}
