% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spat_area.R
\name{make_spat_area}
\alias{make_spat_area}
\title{Add an area/polygon to spatial data}
\usage{
make_spat_area(spat, project, coord, spat.id, new.id, combine)
}
\arguments{
\item{spat}{Spatial dataset to add polygon too.}

\item{project}{Name of project.}

\item{coord}{Longitude and latitude coordinates forming a polygon. Can be a 
numeric vector of even length or a numeric matrix with two columns.}

\item{spat.id}{The ID column in \code{spat}}

\item{new.id}{The ID for new polygon.}

\item{combine}{Whether to use \code{\link{combine_zone}}. This will turn 
the intersections between \code{poly} and \code{spat} into new polygons. 
Note that the new polygon IDs will be derived from \code{spat} and \code{new.id} 
will not be used.}
}
\description{
Add an area/polygon to spatial data
}
\details{
Adds an area/polygone to a spatial area
}
\seealso{
\code{\link{make_polygon}} \code{\link{add_polygon}}
}
