% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Map observed vessel locations}
\usage{
map_plot(dat, project, lat, lon, minmax = NULL, percshown = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in
FishSET database contains the string 'MainDataTable'.}

\item{project}{String, project name.}

\item{lat}{Variable in \code{dat} that defines latitude, in decimal degrees.}

\item{lon}{Variable in \code{dat} that defines longitude, in decimal degrees.}

\item{minmax}{Optional map extent argument, a vector (num) of length four
corresponding to c(minlat, maxlat, minlon, maxlon).}

\item{percshown}{Whole number, percent of points to show. Use this option 
if there are a lot of data points.}
}
\value{
mapout: ggplot2 object
}
\description{
Plot observed locations on a map. For large datasets, it is best 
  to plot a subset of points. Use \code{percshown} to randomly subset the 
  number of points. If the predefined map extent needs adjusting,
  use \code{minmax}.
}
\examples{
\dontrun{
map_plot(pollockMainDataTable, 'pollock', 'LonLat_START_LAT', 'LonLat_START_LON', percshown=10)
}
}
