% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dat.R
\name{merge_dat}
\alias{merge_dat}
\title{Merge datasets}
\usage{
merge_dat(dat, aux, project, mainKey, auxKey)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in the FishSET
database contains the string 'MainDataTable'.}

\item{aux}{Auxiliary data table to join to MainDataTable.}

\item{project}{String, name of project.}

\item{mainKey}{String, name of column(s) in MainDataTable to join by. The number
of columns must match "auxKey".}

\item{auxKey}{String, name of column(s) in the Auxiliary table to join by. The number
of columns must match "mainKey".}
}
\description{
Merge datasets
}
\examples{
\dontrun{

} 

}
