% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_expected_catch.R
\name{merge_expected_catch}
\alias{merge_expected_catch}
\title{Merge expected catch}
\usage{
merge_expected_catch(
  dat,
  project,
  zoneID,
  exp.name,
  new.name = NULL,
  ec.table = NULL,
  log_fun = TRUE
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET
database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{zoneID}{zoneID Variable in \code{dat} that identifies the individual zones or
areas.}

\item{exp.name}{Name(s) of expected catch matrix to merge into \code{dat}.}

\item{new.name}{Optional, new name for \code{exp.name}. These should be in the same
order as \code{exp.name}. \code{catch.formula} must use new names.}

\item{ec.table}{Optional, the name of a specific expected catch table to use.
Defaults to \code{projectnameExpectedCatch}.}

\item{log_fun}{For interal use. Whether to log the function call.}
}
\value{
Merges an expected catch matrix created using \code{\link[=create_expectations]{create_expectations()}}
to the primary dataset, \code{dat}.
}
\description{
Merge expected catch matrices to the primary dataset.
}
