% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_log_call.R
\name{meta_file_exists}
\alias{meta_file_exists}
\title{Check if meta file exists for a project}
\usage{
meta_file_exists(project)
}
\arguments{
\item{project}{Project name.}
}
\value{
TRUE if project meta file exists, FALSE if not.
}
\description{
Check if meta file exists for a project
}
