% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_log_call.R
\name{meta_list}
\alias{meta_list}
\title{Create a meta list with column description section
Used to create a new metadata entry.}
\usage{
meta_list(type, dat = NULL)
}
\arguments{
\item{type}{String}

\item{dat}{Dataset}
}
\description{
Create a meta list with column description section
Used to create a new metadata entry.
}
\keyword{internal}
