% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_log_call.R
\name{meta_log_call}
\alias{meta_log_call}
\title{Log metadata to project folder
Meta data is saved to the metadata `meta_log.json` file in a project's `doc/` directory.}
\usage{
meta_log_call(
  project,
  meta,
  dataset = NULL,
  tab_name,
  tab_type,
  overwrite = FALSE,
  raw = TRUE
)
}
\arguments{
\item{project}{Project name.}

\item{meta}{Meta data list to be saved.}

\item{dataset}{A data object. Used to create the column description section
of the metadata log entry. If \code{NULL}, the section is left empty. This
may be the case if saving the raw metadata. To add column descriptions, 
use the \code{meta_form} app.}

\item{tab_name}{The table name as it appears in the FishSET Database (e.g.
"projectMainDataTable" if the main table).}

\item{tab_type}{The table type. Options include "main", "spat" (spatial), 
"port", "grid" (gridded), and "aux" (auxiliary).}

\item{overwrite}{Logical, whether to save over existing metadata entry.}

\item{raw}{Logical, whether raw metadata file is being saved.}
}
\description{
Log metadata to project folder
Meta data is saved to the metadata `meta_log.json` file in a project's `doc/` directory.
}
\keyword{internal}
