% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{meta_tables}
\alias{meta_tables}
\title{Print meta tables by project and/or type}
\usage{
meta_tables(project, tab.type = NULL)
}
\arguments{
\item{project}{Name of project.}

\item{tab.type}{String, table type. Optional, used to filter output. Options 
include "main", "spat" (spatial), "port", "grid" (gridded), and "aux" 
(auxiliary).}
}
\description{
Print meta tables by project and/or type
}
