% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaColLayoutServ}
\alias{metaColLayoutServ}
\title{Arrange column UIs}
\usage{
metaColLayoutServ(id, cols)
}
\arguments{
\item{id}{An ID string that matches the corresponding UI function: 
\code{\link{metaColLayoutUI}}.}
}
\description{
Arrange column UIs
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It arranges column text boxes  in the main panel.
}
\keyword{internal}
