% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaColLayoutUI}
\alias{metaColLayoutUI}
\title{Arrange column UIs}
\usage{
metaColLayoutUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server function: 
\code{\link{metaColLayoutServ}}.}
}
\description{
Arrange column UIs
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It arranges column text boxes in the main panel.
}
\keyword{internal}
