% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaCreateServ}
\alias{metaCreateServ}
\title{Run metadata servers (Create section)}
\usage{
metaCreateServ(id, cols, meta)
}
\arguments{
\item{id}{An ID string. It is recommended to use an informative and unique 
string such as \code{"meta_create"}.}

\item{cols}{Reactive values object containing column names from a FishSET 
table.}

\item{meta}{Reactive values object to store metadata in.}
}
\description{
Run metadata servers (Create section)
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It runs the metadata server modules needed to save,
  edit, and delete metadata.
}
\keyword{internal}
