% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_gui.R
\name{metadata_gui}
\alias{metadata_gui}
\title{Import, create, and edit metadata}
\usage{
metadata_gui()
}
\description{
\code{metadata_gui} allows users to import metadata from various file types,
create and save new metadata, and edit metadata in a shiny application. 
Metadata is stored in the user's project folder.
}
\details{
The app has two tabs: "Create" and "Edit". The Create tab allows users
  to create new metadata for a selected FishSET table. When a table is loaded,
  the app creates several text boxes that the user can fill. There are 
  four metadata sections: About, Column Description, Contact Info, and Other. 
  
  \describe{
    \item{\strong{About}}{
      \itemize{
      \item \strong{Author} The author of the data.
      \item \strong{Date created} The date data was created.
      \item \strong{Date modified} The last data the data was modified. 
      \item \strong{Version} The current version of the data.
      \item \strong{Confidentiality} Whether the data contains confidential 
        information.
      }
    } 
    \item{\strong{Column Description}}{A text box for each column in the data.
      Include the data type, unit, and values (if categorical)}.
    \item{\strong{Contact Info}}{
      \itemize{
      \item \strong{Person} The primary contact.
      \item \strong{Organization} The primary contact's organization.
      \item \strong{Address} The primary contact's and/or organization's address.
      \item \strong{Phone} The primary contact's work phone number. 
      \item \strong{Email} The primary contact's work email. 
      }
    } 
    \item{\strong{Other}}{
      \itemize{
      \item \strong{License} License for data. 
      \item \strong{Citation} Citation for data.
      \item \strong{Other} Other relevant information.    
      }
    } 
  }
  Users can also import a metadata file from the Create tab, for example,  
  an XML, CSV, or JSON file. This gets saved as "raw" metadata and is separate 
  from the user-created metadata. To see a comprehensive list of accepted file types,
  see \code{\link{parse_meta}} and \code{\link{read_dat}}. To extract metadata
  from a data file (i.e. the data and metadata are both in the same file, but 
  the metadata is not contained within the data itself), use the \code{Reader
  parameters} text box to selectively parse the file (see \code{\link{parse_meta}}
  for details). 
  
  The Edit tab allows users to view, edit, and/or delete metadata saved to 
  FishSET.
}
\examples{
\dontrun{
metadata_gui()
}
}
