% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaDeleteServ}
\alias{metaDeleteServ}
\title{Server for deleting metadata}
\usage{
metaDeleteServ(id, cols, meta)
}
\arguments{
\item{id}{An ID string that matches the corresponding UI function: 
\code{\link{metaDeleteUI}}.}

\item{cols}{Reactive values object containing column names from a FishSET 
table.}

\item{meta}{Reactive values object to store metadata in.}
}
\description{
Server for deleting metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It allows users to delete metadata from the 
  project folder.
}
\keyword{internal}
