% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaDeleteUI}
\alias{metaDeleteUI}
\title{UI for deleting metadata}
\usage{
metaDeleteUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server function: 
\code{\link{metaDeleteServ}}.}
}
\description{
UI for deleting metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It allows users to delete metadata from the 
  project folder.
}
\keyword{internal}
