% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaEditProjServ}
\alias{metaEditProjServ}
\title{Project server for metadata gui (Edit section)}
\usage{
metaEditProjServ(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding UI function: 
\code{\link{metaProjUI}}.}
}
\description{
Project server for metadata gui (Edit section)
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates the UIs for project name and available
  project tables.
}
\keyword{internal}
