% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaEditSaveUI}
\alias{metaEditSaveUI}
\title{Load and Save buttons for metadata gui (Edit section)}
\usage{
metaEditSaveUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaEditLoadServ}} and \code{\link{metaSaveServ}}.}
}
\description{
Load and Save buttons for metadata gui (Edit section)
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates the buttons needed to load and save 
  metadata.
}
\keyword{internal}
