% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaOut}
\alias{metaOut}
\title{Output UI for metadata}
\usage{
metaOut(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaCreateLoadServ}} or \code{\link{metaEditLoadServ}} and 
\code{\link{metaSaveServ}}.}
}
\description{
Output UI for metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It outputs metadata to the main panel.
}
\keyword{internal}
