% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaProjUI}
\alias{metaProjUI}
\title{Project UI for metadata gui}
\usage{
metaProjUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server function: 
\code{\link{metaCreateProjServ}} or  \code{\link{metaEditProjServ}}.}
}
\description{
Project UI for metadata gui
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It creates the UIs for project name and available
  project tables.
}
\keyword{internal}
