% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaRawOut}
\alias{metaRawOut}
\title{Output UI for raw metadata}
\usage{
metaRawOut(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaRawUI}} and \code{\link{metaRawServ}}.}
}
\description{
Output UI for raw metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It outputs raw metadata to the main panel.
}
\keyword{internal}
