% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaRawServ}
\alias{metaRawServ}
\title{Server for handling raw metadata}
\usage{
metaRawServ(id, meta)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaRawUI}} and \code{\link{metaRawOut}}.}

\item{meta}{Reactive values object to store metadata in.}
}
\description{
Server for handling raw metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It imports and saves raw metadata.
}
\keyword{internal}
