% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaRawUI}
\alias{metaRawUI}
\title{UIs for handling raw metadata}
\usage{
metaRawUI(id)
}
\arguments{
\item{id}{An ID string that matches the corresponding server functions: 
\code{\link{metaRawServ}} and \code{\link{metaRawOut}}.}
}
\description{
UIs for handling raw metadata
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It allows users to import and save raw metadata
  from file.
}
\keyword{internal}
