% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_modules.R
\name{metaSaveServ}
\alias{metaSaveServ}
\title{Save server for metadata gui
Contains actions needed to save metadata to meta log.}
\usage{
metaSaveServ(id, cols)
}
\arguments{
\item{id}{An ID string. It is recommended to use an informative and unique 
string such as \code{"meta_create"} or \code{"meta_edit"}.}

\item{cols}{Reactive values object containing column names from a FishSET 
table.}
}
\description{
Save server for metadata gui
Contains actions needed to save metadata to meta log.
}
\details{
This module is used in the \code{\link{metadata_gui}} function 
  and in the FishSET app. It saves new or edited metadata to the project
  folder.
}
\keyword{internal}
