% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_logit_predict.R
\name{mixed_logit_predict}
\alias{mixed_logit_predict}
\title{Mixed Logit predict}
\usage{
mixed_logit_predict(project, logitEq, alts, mod.name, expected.catch = NULL)
}
\arguments{
\item{project}{Name of project}

\item{logitEq}{Parameter estimates}

\item{alts}{Number of alternatives choices in model}

\item{mod.name}{Name of saved model to use}

\item{expected.catch}{Required for conditional logit (\code{logit_c}) model. 
Name of expected catch table to use. 
 Can be the expected catch from the short-term scenario (\code{short}), the medium-term scenario (\code{med}), the 
 long-term scenario (\code{long}), or the user-defined temporal parameters (\code{user}).}
}
\value{
Returns probability of mixed logit model by choice
}
\description{
Prediction component from mixed logit models alled in Policy3, under predict_model_tempNew.m
}
\keyword{internal}
