% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design_gui.R
\name{model_design}
\alias{model_design}
\title{View and select models to run}
\usage{
model_design(dat, project)
}
\arguments{
\item{dat}{Name of table in database containing model measures of fit}

\item{project}{If FALSE, appends model selection to out.mod table. If true, overwrites new table with model selection columns. Set to TRUE if models are deleted or selected models are changed.}
}
\description{
View and select models to run
}
\details{
Opens an interactive session to specify data and parameters for models. Once all parameters have been chosen, select the 'save model and add new model' button
to save the selections. A new model can then be specified.
Click the 'run model' button to run the models.
}
\examples{
\dontrun{
model_design_gui(PollockData, 'pcod')
}
}
