% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_design_list}
\alias{model_design_list}
\title{Get Model Design List}
\usage{
model_design_list(project, name = NULL)
}
\arguments{
\item{project}{Name of project.}

\item{name}{Name of Model Design list in the FishSET database. 
The table name will contain the string "ModelInputData". If \code{NULL}, the 
default table is returned. Use \code{\link{tables_database}} to see a list 
of FishSET database tables by project.}
}
\description{
Returns the Model Design list from the FishSET database.
}
