% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design_outsample.R
\name{model_design_outsample}
\alias{model_design_outsample}
\title{Design out-of-sample model}
\usage{
model_design_outsample(
  project,
  mod.name,
  use.scalers = FALSE,
  scaler.func = NULL
)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Name of saved model to use. Argument can be the name of the model or can pull the name 
of the saved "best" model. Leave \code{mod.name} empty to use the saved "best" model. If more than
one model is saved, \code{mod.name} should be the numeric indicator of which model to use.
Use \code{table_view("modelChosen", project)} to view a table of saved models.}

\item{use.scalers}{Input for \code{create_model_input()}. Logical, should data be normalized? Defaults to \code{FALSE}. Rescaling factors are the mean of the 
numeric vector unless specified with \code{scaler.func}.}

\item{scaler.func}{Input for \code{create_model_input()}. Function to calculate rescaling factors.}
}
\description{
Use selected model design settings to create a model design for the out-of-sample data
}
