% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_fit}
\alias{model_fit}
\title{Load model comparison metrics to console for the defined project}
\usage{
model_fit(project)
}
\arguments{
\item{project}{String, name of project.}
}
\description{
Load model comparison metrics to console. Metrics are displayed for each 
model that was fun. Metrics produced by \code{\link{discretefish_subroutine}}.
}
\examples{
\dontrun{
model_fit('pollock')
}
}
