% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_names}
\alias{model_names}
\title{Return model names}
\usage{
model_names(project)
}
\arguments{
\item{project}{Name of project.}
}
\description{
Returns model names saved to to the model design file.
}
