% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{model_out_summary}
\alias{model_out_summary}
\title{Retrieve most recent summary of model output}
\usage{
model_out_summary(project, output = "print")
}
\arguments{
\item{project}{Name of project}

\item{output}{Output type. "print" returns formatted notes. "table" returns a 
dataframe. "print" is recommended for displaying summary table in a report.}
}
\description{
Retrieve most recent summary of model output
}
\examples{
\dontrun{
model_out_summary("pollock")
}
}
\keyword{internal}
