% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_out_view}
\alias{model_out_view}
\title{View discrete choice model output}
\usage{
model_out_view(table)
}
\arguments{
\item{table}{Table name in sqlite database. Should contain the phrase modelout.}
}
\description{
Returns output from running the discretefish_subroutine function. The table parameter must be the full name of the table name in the fishet_db database.
}
\examples{
\dontrun{
model_out_view('pcodmodelout20190604')
}
}
