% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_params}
\alias{model_params}
\title{Load model parameter estimates, standard errors, and t-statistic to console 
for the defined project}
\usage{
model_params(table, project)
}
\arguments{
\item{table}{Table name in FishSET database. Should contain the phrase 
'ModelOut'. Table name must be in quotes.}

\item{project}{Name of project}
}
\description{
Returns parameter estimates, standard errors, and t-statistic 
 from running the discretefish_subroutine function. The table parameter must 
 be the full name of the table name in the FishSET database.
}
\details{
Returns parameter estimates from running \code{\link{discretefish_subroutine}}. 
The table argument must be the full name of the table name in the FishSET 
database.
}
\examples{
\dontrun{
model_params('pcodModelOut20190604', 'pcod')
}

}
