% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_print}
\alias{msg_print}
\title{Display temporary file using message or print}
\usage{
msg_print(temp_file)
}
\arguments{
\item{temp_file}{temporary file to display. j}
}
\description{
If shiny is running, prints. Otherwise, uses message.
}
\keyword{internal}
