% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvgrnd}
\alias{mvgrnd}
\title{Multivariate normal random generator}
\usage{
mvgrnd(m, sigma, n)
}
\arguments{
\item{m}{mean vector}

\item{sigma}{variance-covariance matrix}

\item{n}{number of iterations}
}
\description{
Multivariate normal random generator
}
\keyword{internal}
