% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_filter.R
\name{na_filter}
\alias{na_filter}
\title{Filters NA's from variable}
\usage{
na_filter(dat, x, replace = F, remove = F, rep.value = NA,
  over_write = FALSE)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{x}{Column(s) in data framce in which to remove or replace NaNs. If multiple columns are passed use x=c().}

\item{replace}{TRUE/FALSE Replace NaNs in a vector? Defaults to FALSE.}

\item{remove}{TRUE/FALSE Remove all remove the entire row of the dataframe where NaN is present in a specified column? Defaults to FALSE.}

\item{rep.value}{Value to replace all NaNs in a column. Defaults to the mean value of the column.}

\item{over_write}{Over_write modified data set in fishset_db database?}
}
\value{
Returns the modified dataframe
}
\description{
Function to return a modified dataframe where NAs have been replaced or removed.
}
\details{
Function to return a modified dataframe where NAs have been replaced or removed. Modified data frame saved to fishset_db database.
}
\examples{
\dontrun{
nan_identify(MainDataTable)
mod.dat <- na_filter(MainDataTable, 'OFFICIAL_TOTAL_CATCH_MT') 
mod.dat <- na_filter(MainDataTable, 'OFFICIAL_TOTAL_CATCH_MT', replace=T) 
mod.dat <- na_filter(MainDataTable, 'OFFICIAL_TOTAL_CATCH_MT', replace=T, rep.value=0) 
mod.dat <- na_filter(MainDataTable, 'OFFICIAL_TOTAL_CATCH_MT', remove=T) 
}
}
\keyword{NA}
