% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_check}
\alias{name_check}
\title{Check for unique and syntatcic column names}
\usage{
name_check(dat, names, repair = FALSE)
}
\arguments{
\item{dat}{Dataset that will contain new columns.}

\item{names}{New names to be added to dataset.}

\item{repair}{Logical, whether to return repaired column names (\code{repair = TRUE})
or just check for unique column names (\code{repair = FALSE}).}
}
\description{
Used for creating new columns.
}
\details{
\code{name_check()} first checks to see if new column names are unique
  and returns an error if not. When \code{repair = TRUE}, \code{name_check()} 
  will check for unique column names and returns new column names that are
  unique and syntactic (see \code{\link[vctrs]{vec_as_names}} for details).
}
