% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_zone.R
\name{new_zone_id}
\alias{new_zone_id}
\title{Create new zone IDs}
\usage{
new_zone_id(combined, id, grid = NULL, closure = NULL, recast = TRUE)
}
\arguments{
\item{combined}{Combined version of grid and closure file.}

\item{id}{Character, the name of the zone ID column.}

\item{grid}{Grid file containing regulatory zones.}

\item{closure}{Closure file containing closure areas.}

\item{recast}{Logical, if \code{TRUE} \code{combined} is passed to 
\code{\link{recast_multipoly}}.}
}
\description{
Creates a new zone ID column for the combined zone file. Used when combining
regulatory zones with closure areas.
}
\details{
This function assigns a new zone ID if the intersection of \code{grid} 
  and \code{closure} creates new, non-contiguous polygons. The ID naming 
  convection is the original name of the regulatory zone followed by a 
  decimal and a number. For example, if a closure area bi-sects regulatory 
  zone A, the resulting zone IDs would be "A.1" and "A.2".
}
\seealso{
\code{\link{recast_multipoly}} \code{\link{combine_zone}}
}
\keyword{internal}
