% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{nfreq_table}
\alias{nfreq_table}
\title{Create one or more binned frequency tables}
\usage{
nfreq_table(
  dataset,
  var,
  group = NULL,
  bins = 30,
  type = "dens",
  v_id = NULL,
  format_lab = "decimal",
  format_tab = "wide"
)
}
\arguments{
\item{dataset}{Primary data containing information on hauls or trips. Table in FishSET
database should contain the string `MainDataTable`.}

\item{var}{String, name of numeric variable to bin.}

\item{group}{String, name of variable(s) to group \code{var} by.}

\item{bins}{Integer, the number of bins to create.}

\item{type}{String, the type of binned frequency table to create. \code{"freq"}
creates a frequency table, \code{"perc"} creates a relative frequency table,
and \code{"dens"} creates a density table.}

\item{v_id}{String, name of vessel ID column (used to detect confidential
information).}

\item{format_lab}{Formatting option for bin labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{format_tab}{Format table "wide" or "long"}
}
\description{
Create one or more binned frequency, relative frequency, or density table.
}
