% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_breaks}
\alias{num_breaks}
\title{Plot breaks for numeric variable}
\usage{
num_breaks(per)
}
\arguments{
\item{per}{period to set breaks for.}
}
\description{
Plot breaks for numeric variable
}
\keyword{internal}
