% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric_cols}
\alias{numeric_cols}
\title{Find numeric columns in data frame}
\usage{
numeric_cols(dat, out = "names")
}
\arguments{
\item{dat}{MainDataTable, dataframe, or list to check.}

\item{out}{Whether to return the column \code{"names"} (the default) or a logical vector 
(\code{"logical"}).}
}
\description{
Find numeric columns in data frame
}
\examples{
\dontrun{
numeric_cols(pollockMainDataTable)
}
}
\keyword{internal}
