% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_factor}
\alias{order_factor}
\title{Order variable}
\usage{
order_factor(dat, fac, val, rev = FALSE)
}
\arguments{
\item{dat}{dataframe containing variable to order.}

\item{fac}{variable name to order.}

\item{val}{value variable to order by.}

\item{rev}{logical, reverse order.}
}
\value{
Returns entire dataframe with ordered factor.
}
\description{
Order variable
}
\keyword{internal}
