% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{outlier_plot}
\alias{outlier_plot}
\title{Evaluate outliers by plotting the data}
\usage{
outlier_plot(dataset, x, outlier.mod = "none", x.dist = "normal",
  save.output = FALSE, remove = F)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{column in dataframe}

\item{outlier.mod}{defines method to ..
Choices include:
                none: No data points are removed
                5_95_quant: Removes data points outside the 5th and 95th quantiles
                25_75_quant: Removes data points outside the 25th and 75th quantiles
                mean_2SD: Removes data points outside +/- 2SD of the mean
                median_2SD: Removes data points outside +/- 2SD of the meadian
                mean_3SD: Removes data points outside +/- 3SD of the mean
                median_3SD: Removes data points outside +/- 3SD of the meadian}

\item{x.dist}{Distribution of the data. Choices include: normal, lognormal, exponential, weibull, poisson, negative binomial}

\item{save.output}{Save plots as pdf file. If TRUE, three plots are returned as pdf file. If false, plots are printed to the screen.}

\item{remove}{Save data with outliers removed. If TRUE, the revised data tables, based on outlier.mod, is returned.}
}
\value{
Plot of the data
}
\description{
Evaluate outliers by plotting the data
}
\keyword{outliers}
