% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{outlier_remove}
\alias{outlier_remove}
\title{Remove outliers from data table}
\usage{
outlier_remove(
  dat,
  project,
  x,
  dat.remove = "none",
  sd_val = NULL,
  over_write = FALSE
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{x}{Variable in \code{dat} containing potential outliers.}

\item{dat.remove}{Defines measure to subset the data. 
Users can use the predefined values (see below) or user-defined standard deviations from the mean.
 For user-defined values, \code{dat.remove} should be a numeric value. For example, \code{dat.remove=6} 
 would would result in value outside 6SD from the mean being class as outliers. 
 User-defined standard deviations from the mean can also be applied using \code{sd_val}.
 Predefined choices: 
 \code{"none"}, \code{"5_95_quant"}, \code{"25_75_quant"}, \code{"mean_2SD"}, 
 \code{"median_2SD"}, \code{"mean_3SD"}, \code{"median_3SD"}.}

\item{sd_val}{Optional. Number of standard deviations from mean defining outliers. 
Example, \code{sd_val=6} would mean values outside +/- 6 SD from the mean would be outliers.}

\item{over_write}{Logical, If TRUE, saves data over previously saved data 
table in the FishSET database.}
}
\value{
Returns the modified primary dataset. Modified dataset will be saved 
  to the FishSET database.
}
\description{
Remove outliers based on outlier measure.
}
\details{
The \code{dat.remove} choices are:
 \itemize{
 \item{numeric value: Remove data points outside +/- `x`SD of the mean}
 \item{none:        No data points are removed}
 \item{5_95_quant:  Removes data points outside the 5th and 95th quantiles}
 \item{25_75_quant: Removes data points outside the 25th and 75th quantiles}
 \item{mean_2SD:    Removes data points outside +/- 2SD of the mean}
 \item{median_2SD:  Removes data points outside +/- 2SD of the median}
 \item{mean_3SD:    Removes data points outside +/- 3SD of the mean}
 \item{median_3SD:  Removes data points outside +/- 3SD of the median}
 }
}
\examples{
\dontrun{
pollockMainDataTable <- outlier_remove(pollockMainDataTable, 'pollock', 'dist', 
   dat.remove = 'mean_2SD', save.output = TRUE)
}
}
