% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_data_name}
\alias{parse_data_name}
\title{Parse data name for logging}
\usage{
parse_data_name(dat, type, project)
}
\arguments{
\item{dat}{Data table to be parsed.}

\item{type}{String, the data type: "main", "aux", "grid", "port", or "spat".}

\item{project}{project name.}
}
\description{
Parse data name for logging
}
\details{
If called while the shiny app is running, the data table name is pulled from 
  the FishSET project settings file. Otherwise, the data table from the caller 
  environment is used.
}
\keyword{internal}
