% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_meta.R
\name{parse_meta_delim}
\alias{parse_meta_delim}
\title{Parse metadata from a delimited file}
\usage{
parse_meta_delim(file, sep = NULL, comment = "#", is_list = FALSE, ...)
}
\arguments{
\item{file}{String, the file path to metadata.}

\item{sep}{String, the field separator character. defaults to \code{comment = "#"}.}

\item{comment}{String, the comment character used to separate (or "comment-out") 
the metadata from the data. Only text that has been commented-out will be read.}

\item{is_list}{Logical, is metadata stored as a list (i.e. Field: value, Field: value)
format). If a colon (":") is used after the field name set this to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link{read.delim}}.}
}
\description{
Use this function if metadata is located in the same file as the data and 
the data is stored in a delimited file (csv, tsv).
}
\keyword{internal}
