% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_meta.R
\name{parse_meta_excel}
\alias{parse_meta_excel}
\title{Parse metadata from an excel file}
\usage{
parse_meta_excel(file, range = NULL, ...)
}
\arguments{
\item{file}{String, file path.}

\item{range}{String, the cell range to read from (e.g. "A1:C5"). See 
\code{\link[readxl]{read_excel}} for more details.}

\item{...}{Additional arguments passed to \code{\link[readxl]{read_excel}}.}
}
\description{
Use this function if metadata is located in the same file as the data.
Correct specification of the \code{range} parameter is key.
}
\keyword{internal}
