% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_meta.R
\name{parse_meta_json}
\alias{parse_meta_json}
\title{Parse metadata from json file}
\usage{
parse_meta_json(file, meta_ind = NULL, simplifyVector = TRUE, flatten = FALSE)
}
\arguments{
\item{file}{String, file path.}

\item{meta_ind}{Integer or string.}

\item{simplifyVector}{Logical, simplifies nested lists into vectors and 
data frames.}

\item{flatten}{Automatically flatten mested data frames into a single 
non-nested data frame.
See \code{\link[jsonlite]{fromJSON}}.}
}
\description{
Use this function if metadata is located in the same file as the data and 
the data is stored in a JSON file.
}
\keyword{internal}
