% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_meta.R
\name{parse_meta_xml}
\alias{parse_meta_xml}
\title{Parse metadata from xml file}
\usage{
parse_meta_xml(file, meta_ind = NULL)
}
\arguments{
\item{file}{String, file path.}

\item{meta_ind}{Integer or string.}
}
\description{
Use this function if metadata is located in the same file as the data and 
the data is stored in an xml file.
}
\keyword{internal}
