% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{parse_notes}
\alias{parse_notes}
\title{Selectively display a note section}
\usage{
parse_notes(project, date = NULL, section, output = "print")
}
\arguments{
\item{project}{The project name.}

\item{date}{Date to pull notes from. If NULL then the most recent version of notes
from the project are retrieved.}

\item{section}{The note section to display. Options include "upload" for Upload data,
"quality" for Data quality evaluation, "explore" for Data exploration, "fleet" for
Fleet functions, "analysis" for Simple analysis, "new_variable" for Create new variable,
"alt_choice" for Alternative choice, "models", and "bookmark".}

\item{output}{Output type. "print" returns formatted notes. "string" returns a 
character vector of the notes. "print" is recommended for displaying notes in a report.}
}
\description{
Selectively display a note section
}
\examples{
\dontrun{
parse_notes("pollock", type = "explore")
}
}
