% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseDeleteEvent}
\alias{parseDeleteEvent}
\title{Extracts the row id number from the id string}
\usage{
parseDeleteEvent(idstr)
}
\arguments{
\item{idstr}{the id string formated as id_INDEX}
}
\value{
INDEX from the id string id_INDEX
}
\description{
Extracts the row id number from the id string
}
\keyword{internal}
