% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perc_of_total}
\alias{perc_of_total}
\title{Calculate grouped percentages}
\usage{
perc_of_total(
  dat,
  value_var,
  group = NULL,
  drop = FALSE,
  val_type = "perc",
  output = "dataset"
)
}
\arguments{
\item{dat}{Data table to summarize.}

\item{value_var}{String, variable name(s) for calculating total.}

\item{group}{String, grouping variable(s) to group `value_var` by.}

\item{drop}{Logical, whether to drop the total column.}

\item{val_type}{String, whether to convert value output to percentage 
\code{"perc"} or proportion \code{"prop"}.}

\item{output}{String, whether to add new variables to dataset (\code{"dataset"})
or return a summary table (\code{"summary"})}
}
\description{
Calculate grouped percentages
}
\keyword{internal}
