% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{period_check}
\alias{period_check}
\title{Check for valid period}
\usage{
period_check(period, date)
}
\arguments{
\item{period}{String, name of period to create. This is used by fleet functions
to summarize data by period.}

\item{date}{String, name of date variable used to create period variable.}
}
\value{
A period code used to format \code{date}. Assign output to a variable
  when used inside a function.
}
\description{
Check for valid period
}
\keyword{internal}
