% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_resid_corr.R
\name{plot.fishset_spatial_resid}
\alias{plot.fishset_spatial_resid}
\title{Plot FishSET Spatial Residuals}
\usage{
\method{plot}{fishset_spatial_resid}(x, ...)
}
\arguments{
\item{x}{A \code{fishset_spatial_resid} object.}

\item{...}{Additional arguments.}
}
\description{
Plot FishSET Spatial Residuals
}
