% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_format}
\alias{plot_format}
\title{Import and format plots to notebook file}
\usage{
plot_format(x, project)
}
\arguments{
\item{x}{Name of plot saved in output}

\item{project}{Name of project}
}
\description{
Import and format plots to notebook file
}
\examples{
\dontrun{
plot_format("pollock_species_catch_2020-05-29.png")
plot_format(pull_output("pollock", "species_catch", type = "plot"))
}
}
\keyword{internal}
