% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_spat.R
\name{plot_spat}
\alias{plot_spat}
\title{Plot spatial dataset}
\usage{
plot_spat(spat)
}
\arguments{
\item{spat}{Spatial dataset to view. Must be an object of class \code{sf} 
or \code{sfc}.}
}
\description{
Simple plotting function for viewing spatial data.
}
\examples{
\dontrun{
plot_spat(pollockNMFSSpatTable)
}
}
