% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_metrics.R
\name{policy_metrics}
\alias{policy_metrics}
\title{Policy change metrics}
\usage{
policy_metrics(
  dat,
  project,
  tripID = "row",
  vesselID,
  catchID,
  datevar = NULL,
  price = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Name of project}

\item{tripID}{Trip identifier. Can be 'row' or the name or names of variables that define trips. 
If \code{tripID='row'} then each row of the  primary dataset is considered to be a unique trip}

\item{vesselID}{Vessel identifier. Variable name in primary dataset that contains unique vessel identifier.}

\item{catchID}{Name of variable in primary dataset that contains catch data.}

\item{datevar}{Name of variable containing date data. Used to split data into years.}

\item{price}{Name of variable containing data on revenue or price data.}
}
\value{
Tables containing basic metrics on effects of proposed zone closures.
}
\description{
Policy change metrics
}
\details{
The policy change metrics reflect the impact of proposed policies in the absence of changes in fisher behavior.
    Policy scenarios are defined using zone_closure() function. 
    Percent of vessels is calculated from the unique vessel identifiers grouped by year and zone. Trips are identified using the 
    tripID argument, otherwise each row is assumed to be a trip. If \code{price} is not defined then percent of revenue loss will 
    be reported as NA.
}
