% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PollockData}
\alias{PollockData}
\title{Pollock Data}
\format{
`PollockData` 
A dataframe with 36 columns and 839 rows:
\describe{
}
}
\usage{
data(PollockData)
}
\description{
Bering Sea pollock data from 2011
}
\keyword{datasets}
