% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_map.R
\name{predict_map}
\alias{predict_map}
\title{Map of predicted probabilities}
\usage{
predict_map(
  project,
  mod.name = NULL,
  policy.name = NULL,
  spat,
  zone.spat,
  plot_type = "dynamic",
  outsample = FALSE,
  outsample_pred = NULL
)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Name of model}

\item{policy.name}{Name of policy scenario}

\item{spat}{A spatial data file containing information on fishery management 
or regulatory zones boundaries. `sf` objects are recommended, but `sp` objects 
can be used as well. See [dat_to_sf()] to convert a spatial table read from 
a csv file to an `sf` object. To upload your spatial data to the FishSETFolder 
see [load_spatial()].}

\item{zone.spat}{Name of zone ID column in `spat`.}

\item{plot_type}{Character, \code{"dynamic"} for interactive leaflet plots (default) and 
\code{"static"} for ggplot.}

\item{outsample}{Logical, indicating if \code{predict_map()} is being used for creating map of 
out-of-sample predicted fishing probabilities \code{outsample = TRUE} or policy scenario 
\code{outsample = FALSE}.}

\item{outsample_pred}{A dataframe with fishing location and predicted probabilities for 
 out-of-sample data.
\code{outsample_pred = NULL} by default and when plotting policy scenarios.}
}
\value{
A map showing predicted probabilities
}
\description{
Create a map showing predicted probabilities by zone
}
\details{
This function requires that model and prediction output tables exist in the FishSET 
database when plotting policy scenario maps.
}
\examples{
\dontrun{

predict_map(project = "scallop", mod.name = "logit_c_mod1", policy.name = "closure_1", 
            spat = spat, zone.spat = "TEN_ID", plot_type = "static")

}
}
