% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outsample.R
\name{predict_outsample}
\alias{predict_outsample}
\title{Predict out-of-sample data}
\usage{
predict_outsample(
  project,
  mod.name,
  outsample.mod.name,
  use.scalers = FALSE,
  scaler.func = NULL
)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Name of saved model to use. Argument can be the name of the model or can pull the name 
of the saved "best" model. Leave \code{mod.name} empty to use the saved "best" model. If more than
one model is saved, \code{mod.name} should be the numeric indicator of which model to use.
Use \code{table_view("modelChosen", project)} to view a table of saved models.}

\item{outsample.mod.name}{Name of the saved out-of-sample model design.}

\item{use.scalers}{Input for \code{create_model_input()}. Logical, should data be normalized? Defaults to \code{FALSE}. Rescaling factors are the mean of the 
numeric vector unless specified with \code{scaler.func}.}

\item{scaler.func}{Input for \code{create_model_input()}. Function to calculate rescaling factors.}
}
\description{
Calculate predicted probabilities for out-of-sample dataset
}
\details{
This function automatically pulls model settings from the selected model and creates an alternative choice matrix, expected catch/revenue matrices, 
and model design for an out-of-sample dataset. This function requires that a filtered out-of-sample data file (.rds file) exists in the output folder.
Note: is that the out-of-sample functions only work with a single selected model at a time. To run out-of-sample functions on a new
out-of-sample dataset, start with load_outsample() if an entirely new dataset or filter_outsample().
}
\examples{
\dontrun{

model_design_outsample("scallop", "scallopModName")

}
}
