% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_probability.R
\name{predict_probability}
\alias{predict_probability}
\title{Subfunction of \code{\link{model_prediction}} for summing predictions with closures}
\usage{
predict_probability(
  probLogit,
  probDataModelIn,
  probDataModelOut,
  zoneClose,
  tacAllowed,
  zoneClosedFish
)
}
\arguments{
\item{probLogit}{Parameter estimates}

\item{probDataModelIn}{Parameter estimates for closed zones}

\item{probDataModelOut}{Parameter estiumates for open zones}

\item{zoneClose}{Sum of probabilitys with 0 fishing allowed}

\item{tacAllowed}{Decimal (percent) of TAC allowed}

\item{zoneClosedFish}{Index of zones that are closed to fishing}
}
\value{
List
  probPredict: Prediction by location of fishing
  sumPredctIn: Total probablilty of fishing inside of closers
  sumPredictOut: Total probablitily of fishing outside of closures
}
\description{
Subfunction of \code{\link{model_prediction}} for summing predictions with closures
}
\keyword{internal}
