% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pretty_lab}
\alias{pretty_lab}
\title{Format numbers in table}
\usage{
pretty_lab(tab, cols = "all", type = "pretty", ignore = NULL)
}
\arguments{
\item{tab}{Table to format.}

\item{cols}{Character string of columns to format. defaults to \code{"all"}
which will include all numeric variables in \code{tab}. If \code{ignore = TRUE}
then the columns listed in \code{cols} will be not be formatted and all 
other columns in \code{tab} will be formatted.}

\item{type}{The type of formatting to apply. \code{"pretty"} uses 
\code{\link[base]{prettyNum}} which uses commas (",") to mark big intervals.
\code{"scientific"} uses scientific notation. \code{"decimal"} simply rounds
to two decimal places.}

\item{ignore}{Logical, whether to exclude the columns listed in \code{cols}
and apply formatting to all other columns in \code{tab}.}
}
\description{
Format numeric columns.
}
