% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pretty_tab}
\alias{pretty_tab}
\title{Format table for R Markdown}
\usage{
pretty_tab(tab, full_width = FALSE)
}
\arguments{
\item{tab}{Table to format.}

\item{full_width}{Logical, whether table should fill out entire width of 
the page.}
}
\description{
Format table for R Markdown
}
