% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pretty_tab_sb}
\alias{pretty_tab_sb}
\title{Scroll box for R Markdown table}
\usage{
pretty_tab_sb(tab, width = "100\%", height = "500px", full_width = FALSE)
}
\arguments{
\item{tab}{Table to format.}

\item{width}{A character string indicating the width of the box. Can be
in pixels (e.g. "50px") or as a percentage (e.g. "50\%").}

\item{height}{A character string indicating the height of the box. Can be
in pixels (e.g. "50px") or as a percentage (e.g. "50\%").}

\item{full_width}{Logical, whether table should fill out entire width of 
the page.}
}
\description{
Allows tables to become scrollable. Useful for large tables.
}
