% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_loc.R
\name{previous_loc}
\alias{previous_loc}
\title{Create Previous Location/Area Variable}
\usage{
previous_loc(
  dat,
  spat,
  project,
  starting_port,
  v_id,
  tripID,
  haulID,
  zoneID = NULL,
  spatID = NULL,
  date = NULL,
  lon = NULL,
  lat = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET
database contains the string 'MainDataTable'.}

\item{spat}{A spatial data file containing information on fishery management
or regulatory zones boundaries. \code{sf} objects are recommended, but \code{sp} objects
can be used as well. See \code{\link[=dat_to_sf]{dat_to_sf()}} to convert a spatial table read from
a csv file to an \code{sf} object. To upload your spatial data to the FishSETFolder
see \code{\link[=load_spatial]{load_spatial()}}.}

\item{project}{String, name of project.}

\item{starting_port}{The name of the starting (or disembarking) port in \code{dat}.}

\item{v_id}{The name of the variable in \code{dat} that uniquely identifies vessels.}

\item{tripID}{Variable name in \code{dat} that uniquely identifies trips.}

\item{haulID}{Variable name in \code{dat} that uniquely identifies hauls.}

\item{zoneID}{Name of zone ID column in \code{dat}. Used to identify the previous
area. Required for previous area variable.}

\item{spatID}{Name of zone ID column in \code{spat}. \code{spat} is used to assign ports
to spatial areas. Required for previous area variable.}

\item{date}{Optional, a date variable to order hauls by.}

\item{lon}{Longitude variable from \code{dat}. Required for previous location
variable.}

\item{lat}{Latitude variable from \code{dat}. Required for previous location variable.}
}
\description{
Creates a variable of the previous port/zone (previous area) or the previous
longitude/latitude for a vessel.
}
\details{
\code{previous_loc()} can create a previous area or location variable.
"Previous area" is defined as the port or zone the vessel last visited. The first
area for each trip is the disembarking port (\code{starting_port}). If a
port is within a zone, the zone is returned. If a port is not within a zone,
the name of the port is returned. "Previous location" is defined as the
previous longitude and latitude of the vessel. The first set of coordinates
is the location of the port. Users must have a port table saved to the
FishSET database to use this function (see \code{\link[=load_port]{load_port()}}). This variable
can be used to define the distance matrix (see \code{\link[=create_alternative_choice]{create_alternative_choice()}}).
}
