% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_resid_corr.R
\name{print.fishset_spatial_resid}
\alias{print.fishset_spatial_resid}
\title{Print FishSET Spatial Residuals}
\usage{
\method{print}{fishset_spatial_resid}(x, ...)
}
\arguments{
\item{x}{A \code{fishset_spatial_resid} object.}

\item{...}{Additional arguments.}
}
\description{
Print FishSET Spatial Residuals
}
