% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{proj_settings_exists}
\alias{proj_settings_exists}
\title{Check if option file exists for a project}
\usage{
proj_settings_exists(project)
}
\arguments{
\item{project}{Project name.}
}
\value{
TRUE if project options file exists, FALSE if not.
}
\description{
Check if option file exists for a project
}
