% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_exists}
\alias{project_exists}
\title{Check if project exists}
\usage{
project_exists(project)
}
\arguments{
\item{project}{Project name.}
}
\description{
Check if project exists
}
\keyword{internal}
