% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{project_files}
\alias{project_files}
\title{List output files by project name}
\usage{
project_files(project)
}
\arguments{
\item{project}{Project name}
}
\description{
List output files by project name
}
\examples{
\dontrun{
project_files("pollock")
}
}
