% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{project_logs}
\alias{project_logs}
\title{List logs by project}
\usage{
project_logs(project, modified = FALSE)
}
\arguments{
\item{project}{Name of project.}

\item{modified}{Logical, whether to show modification date. Returns a data
frame.}
}
\description{
List logs by project
}
