% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{project_tables}
\alias{project_tables}
\title{Display database table names by project}
\usage{
project_tables(project, ...)
}
\arguments{
\item{project}{Name of project.}

\item{...}{String, additional characters to match by.}
}
\description{
Display database table names by project
}
\examples{
\dontrun{
project_tables("pollock")
project_tables("pollock", "main")
}
}
