% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{projects}
\alias{projects}
\title{Display projects names}
\usage{
projects()
}
\description{
Display projects names
}
\details{
Lists the unique project names currently in the FishSET Database.
}
\examples{
\dontrun{
projects()
} 

}
