% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_info_data}
\alias{pull_info_data}
\title{Pull the most recent data index file for given projet}
\usage{
pull_info_data(project)
}
\arguments{
\item{project}{Name of project, such as pollock}
}
\description{
Pull the most recent data index file for given projet
}
