% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pull_log}
\alias{pull_log}
\title{pull log file}
\usage{
pull_log(project, log_date = NULL)
}
\arguments{
\item{project}{Project name.}

\item{log_date}{Date of log to be pulled. If \code{NULL}, then most recent log file
is retrieved.}
}
\description{
pull log file
}
\keyword{internal}
