% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pull_meta}
\alias{pull_meta}
\title{Retrieve/display meta data by project}
\usage{
pull_meta(project, tab.name = NULL, tab.type = NULL, format = FALSE)
}
\arguments{
\item{project}{Project name.}

\item{tab.name}{String, table name. Optional, used to filter output to a specific
table.}

\item{tab.type}{String, table type. Optional, used to filter output. Options 
include "main", "spat" (spatial), "port", "grid" (gridded), and "aux" (auxiliary).}

\item{format}{Logical, whether to format output using \code{\link[pander]{pander}}.
Useful for displaying in reports.}
}
\description{
Retrieve/display meta data by project
}
