% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pull_notes}
\alias{pull_notes}
\title{Pull notes from output folder}
\usage{
pull_notes(project, date = NULL)
}
\arguments{
\item{project}{String, the project name.}

\item{date}{String, date to pull notes from. If NULL, most recent note file is 
retrieved.}
}
\description{
Pull notes from output folder
}
\details{
Notes are saved to the output folder by project name and date. If date is not
  specified then the most recent notes file with the project name is pulled. Notes are
  are also saved by FishSET app session; if more than one session occured in the same day, each 
  session's notes are pulled and listed in chronological order.
}
