% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pull_plot}
\alias{pull_plot}
\title{Import and format plots to notebook file}
\usage{
pull_plot(project, fun, date = NULL, conf = TRUE)
}
\arguments{
\item{project}{Project name.}

\item{fun}{String, the name of the function that created the plot.}

\item{date}{the date the plot was created. If NULL, then the most recent version
is retrieved.}

\item{conf}{Logical, whether to return suppressed confidential data. 
Unsuppressed output will be pulled if suppressed output is not available.}
}
\description{
Import and format plots to notebook file
}
\examples{
\dontrun{
pull_plot("pollock", "density_plot")
}
}
