% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pull_table}
\alias{pull_table}
\title{Import and format table to notebook file}
\usage{
pull_table(project, fun, date = NULL, conf = TRUE)
}
\arguments{
\item{project}{Project name.}

\item{fun}{String, the name of the function that created the table.}

\item{date}{the date the table was created. If NULL, then the most recent 
version is retrieved.}

\item{conf}{Logical, whether to return suppressed confidential data. 
Unsuppressed output will be pulled if suppressed output is not available.}
}
\description{
Import and format table to notebook file
}
\examples{
\dontrun{
pull_table("pollock", "vessel_count")
}

}
