% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qaqc_helper}
\alias{qaqc_helper}
\title{Helper function for testing data quality issues in MainDataTable}
\usage{
qaqc_helper(dat, fun, output = "logical")
}
\arguments{
\item{dat}{Dataframe to test for quality issues.}

\item{fun}{A function or custom function that returns a single logical value 
to apply to each column in \code{dat}. There are three quick options for 
common checks: \code{"NA"}, \code{"NaN"}, and \code{"Inf"}.}

\item{output}{\code{"logical"} returns a single logical value for each column in
\code{dat}. \code{"names"} returns the column names that evaluate to \code{TRUE}.}
}
\description{
Helper function for testing data quality issues in MainDataTable
}
\details{
Returns a vector of logical values (\code{output = "logical"}) or a vector of 
  column names where the condition evaluated by \code{fun} returns TRUE 
  (\code{output = "names"}).
}
\examples{
\dontrun{
 outputs names of columns containing NAs
    qaqc_helper(pollockMainDataTable, "NA", "names")
 outputs logical vector of columns containing strings
     qaqc_helper(pollockMainDataTable, is.character)
custom function example
     qaqc_helper(pollockMainDataTable, function(x) = all(is.na(x)), "names") 
}
}
\keyword{internal}
