% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quietly_test}
\alias{quietly_test}
\title{quietly_test
capture console messages if exist and print to shiny app}
\usage{
quietly_test(.f)
}
\arguments{
\item{.f}{function name}
}
\description{
quietly_test
capture console messages if exist and print to shiny app
}
\keyword{internal}
