% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{randomize_lonlat_zone}
\alias{randomize_lonlat_zone}
\title{Randomize latitude and longitude points by zone}
\usage{
randomize_lonlat_zone(dat, project, spat, lon, lat, zone)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Project name.}

\item{spat}{Spatial data table containing regulatory zones. This can be 
a "spatial feature" or sf object.}

\item{lon}{String, variable name containing longitude.}

\item{lat}{String, variable name containing latitude.}

\item{zone}{String, column name contain the assigned zone. Must be the same 
for both the spatial data table and MainDataTable.}
}
\description{
Randomize latitude and longitude points by zone
}
\details{
This is one of the FishSET confidentiality functions. It replaces 
  longitude and latitude values with randomly sampled coordinates from the
  regulatory zone the observation occurred in.
}
\examples{
\dontrun{
randomize_lonlat_zone(pollockMainDataTable, "pollock", spatdat, 
                   lon = "LonLat_START_LON", lat = "LonLat_START_LAT",
                   zone = "NMFS_AREA")
}
}
