% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{randomize_value_row}
\alias{randomize_value_row}
\title{Randomize value between rows}
\usage{
randomize_value_row(dat, project, value)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Project name.}

\item{value}{String, variable name to be randomly distributed between rows.}
}
\description{
Randomize value between rows
}
\details{
This is one of the FishSET confidentiality functions. It is useful
  for randomly assigning ID values between observations.
}
\examples{
\dontrun{
jitter_row(pollockMainDataTable, "PERMIT")
}
}
