% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{read_dat}
\alias{read_dat}
\title{Import data}
\usage{
read_dat(x, data.type = c("csv", "mat", "json", "shape", "txt", "spss",
  "stata", "R"))
}
\arguments{
\item{x}{name and directory of dataframe to be read in}

\item{data.type}{csv, mat, json, shape}
}
\description{
Import data
}
\details{
Uses the appropriate function to read in data based on data type. Supported data types include shaple, csv, json, matlab, R, spss, and stata files.
}
\examples{
\dontrun{ 
dat <- read_dat('nmfs_manage_simple.shp', 'shape')
}
}
