% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_zone.R
\name{recast_multipoly}
\alias{recast_multipoly}
\title{Recast multi-polygons}
\usage{
recast_multipoly(grid, closure, combined, id, inter)
}
\arguments{
\item{grid}{Grid file containing regulatory zones.}

\item{closure}{Closure file containing closure areas.}

\item{combined}{Combined version of grid and closure file.}

\item{id}{Character, the name of the zone ID column.}
}
\description{
Re-cast intersecting multi-polygons to polygons. Used when combining 
regulatory zones with closure areas.
}
\details{
This function is primarily used to extract polygons during 
  the ID re-labeling process performed by \code{\link{new_zone_id}}. If
  a polygon from \code{grid} intersected with \code{closure}, it is extracted
  from it's multi-polygon and given a unique ID.
}
\seealso{
\code{\link{combine_zone}} \code{\link{new_zone_id}}
}
\keyword{internal}
