% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{remove_model_design}
\alias{remove_model_design}
\title{Remove a model design from list in ModelInputData table}
\usage{
remove_model_design(project, names)
}
\arguments{
\item{project}{Name of project.}

\item{names}{Names of model designs to be deleted from the table}
}
\description{
Remove a model design from list in ModelInputData table
}
