% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{replace_sup_code}
\alias{replace_sup_code}
\title{Replace suppression code}
\usage{
replace_sup_code(output, code = NA)
}
\arguments{
\item{output}{Table containing suppressed values.}

\item{code}{The replacement suppression code. \code{code = NA} by default; 
this is ideal for plotting as ggplot automatically removes NAs.}
}
\description{
This function replaces the default suppression code in a table.
}
\details{
Suppressed values are represented as `-999` by default. This isn't ideal 
for plotting. NAs -- the default in `replace_sup_code()`-- are a better 
alternative for plots as they can easily be removed.
}
\examples{
\dontrun{
summary_tab <- replace_sup_code(summary_tab, code = NA)
}
}
