% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repmat}
\alias{repmat}
\title{R verions of matlabs repmat}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{matrix}

\item{m}{nrows}

\item{n}{n cols}
}
\description{
R verions of matlabs repmat
}
\keyword{internal}
