% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_grid_dat.R
\name{retrieve_map}
\alias{retrieve_map}
\title{Get stamen map}
\usage{
retrieve_map(grid, lon, lat)
}
\arguments{
\item{grid}{Gridded data table to visualize. Use string if visualizing a gridded data
table in the FishSET Database.}

\item{lon}{String, variable name containing longitude.}

\item{lat}{String, variable name containing latitude.}
}
\description{
Get stamen map
}
\details{
This is a wrapper for \code{\link[ggmap]{get_stamenmap}}
}
