% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_catch.R
\name{roll_catch}
\alias{roll_catch}
\title{Apply moving average function to catch}
\usage{
roll_catch(
  dat,
  project,
  catch,
  date,
  group = NULL,
  combine = FALSE,
  k = 10,
  fun = "mean",
  filter_date = NULL,
  date_value = NULL,
  filter_by = NULL,
  filter_value = NULL,
  filter_expr = NULL,
  facet_by = NULL,
  scale = "fixed",
  align = "center",
  conv = "none",
  tran = "identity",
  format_lab = "decimal",
  output = "tab_plot",
  ...
)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET database
should contain the string `MainDataTable`.}

\item{project}{Name of project.}

\item{catch}{Variable name or names containing catch data. Multiple variables can
entered as a vector.}

\item{date}{Date variable to aggregate by.}

\item{group}{Variable name or names to group by. Plot will display up to two grouping
variables.}

\item{combine}{Whether to combine variables listed in \code{group}. This is passed
to the "fill" or "color" aesthetic for plots.}

\item{k}{The width of the window.}

\item{fun}{The function to be applied to window. Defaults to \code{mean}.}

\item{filter_date}{The type of filter to apply to table. The "date_range" option will subset 
the data by two date values entered in \code{filter_val}. Other options include "year-day", 
"year-week", "year-month", "year", "month", "week", or "day". The argument filter_value must 
be provided.}

\item{date_value}{String containing a start and end date if using filter_date = "date_range", 
e.g. c("2011-01-01", "2011-03-15"). If filter_date = "period" or "year-period", use integers 
(4 digits if year, 1-2 if day, month, or week). Use a list if using a two-part filter, e.g. "year-week",
with the format \code{list(year, period)} or a vector if using a single period, \code{c(period)}. 
For example, \code{list(2011:2013, 5:7)} will filter the data table from weeks 5 through 7 for 
years 2011-2013 if filter_date = "year-week".\code{c(2:5)} will filter the data
February through May when filter_date = "month".}

\item{filter_by}{String, variable name to filter by.}

\item{filter_value}{A vector of values to filter `MainDataTable` by using the variable 
in \code{filter_by}.}

\item{filter_expr}{String, a valid R expression to filter `MainDataTable` by using the variable 
in \code{filter_by}.}

\item{facet_by}{Variable name to facet by. This can be a variable that exists in
the dataset, or a variable created by \code{roll_catch()} such as \code{"year"},
\code{"month"}, or \code{"species"} if more than one variable is entered in \code{catch}.}

\item{scale}{Scale argument passed to \code{\link{facet_grid}}.
Options include \code{"free"}, \code{"free_x"}, \code{"free_y"}. Defaults to
\code{"fixed"}.}

\item{align}{Indicates whether results of window should be left-aligned (\code{"left"}),
right-aligned (\code{"right"}), or centered (\code{"center"}). Defaults to
\code{"center"}.}

\item{conv}{Convert catch variable to \code{"tons"}, \code{"metric_tons"}, or
by using a function. Defaults to \code{FALSE}.}

\item{tran}{A function to transform the y-axis. Options include log, log2, log10, sqrt.}

\item{format_lab}{Formatting option for y-axis labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{output}{Whether to display \code{"plot"}, \code{"table"}, or both. Defaults
to both (\code{"tab_plot"}).}

\item{...}{Additional arguments passed to \code{\link[zoo]{rollapply}}}
}
\description{
Apply moving average function to catch
}
\examples{
\dontrun{
roll_catch(pollockMainDataTable, project = "pollock", catch = "LBS_270_POLLOCK_LBS",
  date = "FISHING_START_DATE", group = "GEAR_TYPE", k = 15
)

roll_catch(pollockMainDataTable, project = "pollock", catch = c("LBS_270_POLLOCK_LBS", 
 "LBS_110_PACIFIC_COD_LBS"), date = "FISHING_START_DATE", group = "GEAR_TYPE", k = 5, 
 filter_date = "month", date_value = 4:6, facet_by = "month", conv = "tons"
)
}

}
