% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_catch.R
\name{roll_catch_plot}
\alias{roll_catch_plot}
\title{Create and format \code{roll_catch} plot}
\usage{
roll_catch_plot(
  roll_tab,
  catch,
  date,
  group,
  facet_by,
  fun,
  k,
  conv,
  tran,
  format_lab,
  scale
)
}
\arguments{
\item{roll_tab}{Table containing rolling summary of catch.}

\item{catch}{String, name of catch variable(s).}

\item{date}{String, name of date column.}

\item{group}{String, name of grouping variable(s).}

\item{facet_by}{String, name of facet variable(s).}

\item{fun}{String, name of summary function.}

\item{k}{Numeric, the width of the window.}

\item{conv}{String, convert pounds to "tons" or "metric_tons".}

\item{tran}{A function to transform the y-axis. Options include log, log2, 
log10, sqrt.}

\item{format_lab}{Formatting option for y-axis labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{scale}{Scale argument passed to \code{\link{facet_grid}}.Options include 
\code{"free"}, \code{"free_x"}, \code{"free_y"}. Defaults to \code{"fixed"}.}
}
\description{
Create and format \code{roll_catch} plot
}
\keyword{internal}
