% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{roll_cc}
\alias{roll_cc}
\title{Apply window_cc}
\usage{
roll_cc(dat, k, align)
}
\arguments{
\item{dat}{Dataframe used to calculate rolling catch.}

\item{k}{Window width.}

\item{align}{Window alignment.}
}
\description{
Used by \code{roll_catch()} to count unique vessels within a rolling window 
for multiple columns of catch.
}
\seealso{
\code{\link{roll_catch}} \code{\link{window_cc}}
}
\keyword{internal}
