% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{run_confid_check}
\alias{run_confid_check}
\title{Check whether confidentiality rules should be applied}
\usage{
run_confid_check(project)
}
\arguments{
\item{project}{Name of project.}
}
\value{
\code{TRUE} if confidentiality settings exists and \code{check = TRUE}, 
  \code{FALSE} if settings do not exists yet or \code{check = FALSE}.
}
\description{
Check whether confidentiality rules should be applied
}
\keyword{internal}
