% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_closure_scenario.R
\name{save_closure_scenario}
\alias{save_closure_scenario}
\title{Save unique closure scenarios}
\usage{
save_closure_scenario(project, c_list)
}
\arguments{
\item{project}{Name of project.}

\item{c_list}{List of closure scenarios to check and save.}
}
\description{
Saves closure scenarios in the project's internal SQLite database.
}
\keyword{internal}
