% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_grid.R
\name{save_grid_cache}
\alias{save_grid_cache}
\title{Save grid file to project data directory}
\usage{
save_grid_cache(project, grid_list, grid_info, mod_type = "combine")
}
\arguments{
\item{project}{Name of project.}

\item{grid_list}{List containing grid files.}

\item{grid_info}{List containing grid information.}

\item{mod_type}{String, "combine" for combined map files or "edit" for 
edited map files.}
}
\description{
Save grid file to project data directory
}
\details{
This function references the grid log to determine whether a grid
 file should be saved. If a grid file is unique it is saved, otherwise no
 action is taken.
}
\seealso{
\code{\link{unique_grid}} \code{\link{log_grid_info}}
}
\keyword{internal}
