% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_nplot}
\alias{save_nplot}
\title{Save table to output folder}
\usage{
save_nplot(project, func_name, plot_list, id = "num", ...)
}
\arguments{
\item{project}{name of project.}

\item{func_name}{Name of function used to create plot.}

\item{...}{addition arguments passed to \code{\link[ggplot2]{ggsave}}.}
}
\description{
Save table to output folder
}
\examples{
\dontrun{
save_nplot(project, "species_catch", plot_list)
}
}
\keyword{internal}
