% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_ntable}
\alias{save_ntable}
\title{Save list of tables to output folder}
\usage{
save_ntable(table, project, func_name, id = "num", ...)
}
\arguments{
\item{table}{List containing tables to save.}

\item{project}{project name.}

\item{func_name}{Name of function used to create table.}

\item{id}{String, id to append to function name. Options include "seq" to 
save by list entry number or "name" to save by list entry name.}

\item{...}{addition arguments passed to \code{\link{write.csv}}.}
}
\description{
Save list of tables to output folder
}
\examples{
\dontrun{
save_ntable(tab_list, project, "species_catch")
}
}
\keyword{internal}
