% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save plot to output folder}
\usage{
save_plot(project, func_name, ...)
}
\arguments{
\item{project}{name of project.}

\item{func_name}{Name of function used to create plot.}

\item{...}{addition arguments passed to \code{\link[ggplot2]{ggsave}}.}
}
\description{
Save plot to output folder
}
\examples{
\dontrun{
save_plot(project, "species_catch")
}
}
\keyword{internal}
