% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raw_meta.R
\name{save_raw_meta}
\alias{save_raw_meta}
\title{Save a meta data file to project folder}
\usage{
save_raw_meta(
  file,
  project,
  dataset = NULL,
  tab.name = NULL,
  tab.type,
  parse = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{file}{String, file path.}

\item{project}{Project name.}

\item{dataset}{Optional, the data.frame associated with the meta data. Used 
to add column names to meta file.}

\item{tab.name}{The table name as it appears in the FishSET Database (e.g.
"projectMainDataTable" if the main table).}

\item{tab.type}{The table type. Options include "main", "spat" (spatial), 
"port", "grid" (gridded), and "aux" (auxiliary).}

\item{parse}{Logical, whether to parse meta data from a data file. See 
\code{\link{parse_meta}}.}

\item{overwrite}{Logical, whether to overwrite existing meta table entry.}

\item{...}{Additional arguments passed to \code{\link{parse_meta}}.}
}
\description{
Raw (i.e. original or pre-existing) meta data can be saved to the project folder. 
To add additional meta data (e.g. column descriptions), see ...
}
\seealso{
\code{\link{parse_meta}}
}
