% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_table}
\alias{save_table}
\title{Save table to output folder}
\usage{
save_table(table, project, func_name, ...)
}
\arguments{
\item{table}{table name.}

\item{project}{project name.}

\item{func_name}{function name.}

\item{...}{addition arguments passed to \code{\link{write.csv}}.}
}
\description{
Save table to output folder
}
\examples{
\dontrun{
save_table(count, project, "species_catch")
}
}
\keyword{internal}
