% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scallop_ports}
\alias{scallop_ports}
\title{Ports from the NE scallop fishery}
\format{
A data frame (tibble) with 40 observations and 3 variables.\cr
\cr
[,1] Port names \cr
[,2] Longitude \cr
[,3] Latitude \cr
}
\source{
NEED TO ADD SOURCE DESCRIPTION
}
\usage{
scallop_ports
}
\description{
A dataset containing the names and lat/lon coordinates of ports used in the US northeast scallop fishery.
}
\keyword{datasets}
