% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal_ID.R
\name{seasonalID}
\alias{seasonalID}
\title{Create single binary fishery season identifier variable}
\usage{
seasonalID(
  dat,
  project,
  seasonal.dat = NULL,
  start,
  end,
  overlap = FALSE,
  name = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{seasonal.dat}{Data table containing date of fishery season(s). Data table can be pulled from the FishSET database.
Leave \code{seasonal.dat} as NULL if supplying start and end dates with \code{start} and \code{end} arguments.}

\item{start}{Date, supplied as a string (example: start='2011/04/22', start='04222011'), or 
variable in \code{seasonal.dat} which identifies start date of fishery season}

\item{end}{DDate, supplied as a string (example: start='2011/04/22', start='04222011'), or 
variable in \code{seasonal.dat} which identifies end date of fishery season}

\item{overlap}{Logical. Should trip or haul dates that start before or end after the fishery season date but starts or ends within the fishery season
dates be included? FALSE indicates to inlude only hauls/trips that fall completely within the bounds of a fishery season date. Defaults to FALSE.}

\item{name}{String  Seasonal identifier name}
}
\value{
Returns a binary variable of within (1) or outside (0) the fishery season.
}
\description{
Create single binary fishery season identifier variable
}
\details{
Uses a supplied dates or a table of fishery season dates to create fishery season identifier variables. Output is a binary
variable called \code{name} or `SeasonID` if \code{name} is not supplied. \cr\cr
For each row \code{dat}, the function matches fishery season dates provided in \code{seasonal.dat} to the earliest date variable in \code{dat}.
}
\examples{
\dontrun{
#Example using a table stored in the FishSET database
pcodMainDataTable <- season_ID("pcodMainDataTable", 'pcod', seasonal_dat='seasonTable', 
     start='SeasonStart', end='SeasonEnd', name='2001A')
#Example using manually entered dates
pcodMainDataTable <- season_ID("pcodMainDataTable", 'pcod', seasonal.dat=NULL, 
    start='04152011', end='06302011', name='2001A')
}

}
