% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_checks_gui.R
\name{select_checks}
\alias{select_checks}
\title{View and select which validation checks to run}
\usage{
select_checks(dat, project)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{project}{Name of project. Parameter is used to generate meaningful table names in fishset_db database.}
}
\description{
View and select which validation checks to run
}
\details{
Opens an interactive table that allows uses to select which data validation checks to run by clicking check boxes.
}
