% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_directory_method}
\alias{select_directory_method}
\title{Search for method to select directory}
\usage{
select_directory_method()
}
\description{
Search for method to select directory
}
\details{
Tries out a sequence of potential methods for selecting a directory to find one that works 
   The fallback default method if nothing else works is to get user input from the console.
}
\keyword{internal}
