% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{select_model}
\title{select_model
View and select models}
\usage{
select_model(table, overwrite.table)
}
\arguments{
\item{table}{Name of table in database containing model measures of fit}

\item{overwrite.table}{Append model selection to modelChosen table or delete existing table and save new table}
}
\description{
select_model
View and select models
}
