% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vars_gui.R
\name{select_vars}
\alias{select_vars}
\title{View and select which variables to include}
\usage{
select_vars(dat, project)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{project}{Name of project. Parameter is used to generate meaningful table names in fishset_db database.}
}
\description{
View and select which variables to include
}
\details{
Opens an interactive table that allows uses to select which variables to included by clicking check boxes. 
Data should be loaded into the fishset_db database before running this function. Select variables that will be used to generate further variables, such as rates or cpue, and variables to be included in models.
}
\examples{
\dontrun{
select_vars('pcodMainDataTable', pcod)
}
}
