% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_closure_scenario.R
\name{serialize_table}
\alias{serialize_table}
\title{Serialize an object into a FishSET project database table}
\usage{
serialize_table(table, object, project)
}
\arguments{
\item{table}{Database table name.}

\item{object}{R object to serialize.}

\item{project}{Project name.}
}
\description{
Serialize an object into a FishSET project database table
}
\keyword{internal}
