% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{set_confid_check}
\alias{set_confid_check}
\title{Set confidentiality parameters}
\usage{
set_confid_check(project, check = TRUE, v_id = NULL, rule = "n", value = NULL)
}
\arguments{
\item{project}{Name of project.}

\item{check}{Logical, whether to check for confidentiality.}

\item{v_id}{String, the column name containing the vessel identifier.}

\item{rule}{String, the confidentiality rule to apply. See "Details" below. 
\code{rule = "n"} suppresses values containing fewer than n vessels. 
\code{rule = "k"} suppresses values where a single vessel contains k percent 
or more of the total catch.}

\item{value}{The threshold for confidentiality. for \code{rule = "n"} must 
be an integer of at least 2. For \code{rule = "k"} any numeric value from 
0 to 100.}
}
\description{
This function specifics whether to check for confidentiality and which 
rule should be applied.
}
\details{
\code{rule = "n"} counts the number of vessel in each strata 
  and suppresses values where fewer than n vessels are present. For
  \code{rule = "k"}, or the "Majority allocation rule", each vessel's share of 
  catch is calculated by strata. If any vessel's total catch share is greater 
  than or equal to k percent the value is suppressed.
}
\examples{
\dontrun{
set_confid_check("pollock", check = TRUE, v_id = "PERMIT", rule = "n", value = 3L)
}
}
