% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_settings.R
\name{set_user_locoutput}
\alias{set_user_locoutput}
\title{Set user folder directory}
\usage{
set_user_locoutput(loc_dir, project)
}
\arguments{
\item{loc_dir}{Local user directory}

\item{project}{Name of project.}
}
\description{
Set user folder directory
}
\details{
This function saves the local user directory to the project settings file
 with a valid folder directory. This directory path is used for inserting plots 
 and tables from a folder outside the FishSET package into the FishSET RMarkdown 
 Template.
}
\seealso{
\code{\link{insert_plot}} \code{\link{insert_table}} \code{\link{get_proj_settings}}
}
