% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{shift_long}
\alias{shift_long}
\title{Shift longitude predicate
Detects whether longitude coordinates should be re-centered to Pacific view}
\usage{
shift_long(spat)
}
\arguments{
\item{spat}{Spatial data to check.}
}
\value{
\code{TRUE} if longitude should be shifted to pacific view (i.e. 
any longitude values < 0).
}
\description{
Shift longitude predicate
Detects whether longitude coordinates should be re-centered to Pacific view
}
\keyword{internal}
