% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftSortX.R
\name{shiftSortX}
\alias{shiftSortX}
\title{shiftSortX}
\usage{
shiftSortX(x, ch, y, distance, alts, ab)
}
\arguments{
\item{x}{Matrix of choice possibilites from create_logit_input}

\item{ch}{Data corresponding to actual zonal choice}

\item{y}{Data corresponding to actual catch}

\item{distance}{Data corresponding to distance}

\item{alts}{Number of alternative choices in model}

\item{ab}{Number of cost parameters + number of alts}
}
\value{
d - matrix of choice possibilites and distance
}
\description{
Shifts choices so that the chosen zone will be automatically the first one
}
\details{
Called in zonalSubroutine n
}
