% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_expectation.R
\name{short_expectations}
\alias{short_expectations}
\title{Short expectations}
\usage{
short_expectations(
  dat,
  project,
  catch,
  price,
  defineGroup,
  temp.var,
  temporal,
  calc.method,
  lag.method,
  empty.catch,
  empty.expectation,
  dummy.exp
)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in FishSET database should contain the string `MainDataTable`.}

\item{project}{Name of project. Used to pull working alternative choice matrix from FishSET database.}

\item{catch}{Variable containing catch data.}

\item{price}{Variable containing price/value data. Used in calculating expected revenue. Leave null if calculating expected catch. Multiplied against catch to generated revenue.}

\item{defineGroup}{If empty, data is treated as a fleet}

\item{temp.var}{Variable containing temporal data}

\item{temporal}{Daily (Daily time line) or sequential (sequential order)}

\item{calc.method}{Select standard average (standardAverage), simple lag regression of means (simpleLag), or weights of regressed groups (weights)}

\item{lag.method}{Use regression over entire group (simple) or for grouped time periods (grouped)}

\item{empty.catch}{Replace empty catch with NA, 0, mean of all catch (allCatch), or mean of grouped catch(groupCatch)}

\item{empty.expectation}{Do not replace (NULL) or replace with 0.0001 or 0}

\item{dummy.exp}{Logical, should dummy variable be outputted. If true, output dummy variable for originally missing value.}
}
\value{
Expected catch matrix. Saved to database via create_expectations
}
\description{
Short expectations
}
\keyword{internal}
