% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_welfare.R
\name{sim_welfare}
\alias{sim_welfare}
\title{Welfare simulator}
\usage{
sim_welfare(project, X, alts, beta_j, scenario)
}
\arguments{
\item{project}{Name of project}

\item{X}{Independent variables from RUM}

\item{alts}{Number of alternative zones}

\item{beta_j}{Output from mvrgrnd}

\item{scenario}{Scenario list created in \code{\link{welfare_predict}}}
}
\description{
Welfare simulator
}
\details{
Subfunction of \code{\link{welfare_predict}}
}
\keyword{internal}
