% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{simplify_list}
\alias{simplify_list}
\title{Simplify a list}
\usage{
simplify_list(l, format = FALSE)
}
\arguments{
\item{l}{A list.}

\item{format}{Logical, whether to print list using pandoc markdown.}
}
\value{
A list
}
\description{
Cleans a list and converts applicable leaf nodes to dataframes.
}
\examples{
\dontrun{
simplify_list(list(A = 1:10, B = 11:20))

simplify_list(list(A = list(X = 1:10, Y = letters[1:10])))

simplify_list(list(A = 1:10, B = "Text", C = c("text", "text")))
}
}
\seealso{
\code{\link{clean_list}} \code{\link{list_df}}  \code{\link[pander]{pander}}
}
\keyword{internal}
