% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{Calculate skewness}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{variable of interest}

\item{na.rm}{set to FALSE}
}
\description{
Calculate skewness
}
\keyword{internal}
