% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spars}
\alias{spars}
\title{Helper function for sparsity functions}
\usage{
spars(x, dname)
}
\arguments{
\item{x}{Name of table}

\item{dname}{time period}
}
\description{
Helper function for sparsity functions
}
\keyword{internal}
