% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparstable.R
\name{sparsetable}
\alias{sparsetable}
\title{Sparsity table}
\usage{
sparsetable(dat, timevar, zonevar, var)
}
\arguments{
\item{dat}{Main data frame over which to apply function.}

\item{timevar}{Variable containing temporal data}

\item{zonevar}{Variable contianing zone observation assigned to}

\item{var}{Variable containing catch data}
}
\description{
Function to create sparsity table
}
