% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparstable.R
\name{sparsplot}
\alias{sparsplot}
\title{Sparsity plot}
\usage{
sparsplot(x)
}
\arguments{
\item{x}{output from spartstable function}
}
\description{
Sparsity plot
}
\details{
Returns a plot of sparsity values over time. Requires sparsity table generated by sparsetable function.
}
