% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_qaqc.R
\name{spat_qaqc_gui}
\alias{spat_qaqc_gui}
\title{GUI for spatial data checks}
\usage{
spat_qaqc_gui(dataset, project, spatdat, checks = NULL)
}
\arguments{
\item{dataset}{Primary data containing information on hauls or trips. Table in 
FishSET database contains the string 'MainDataTable'.}

\item{project}{Name of project.}

\item{spatdat}{Spatial data containing information on fishery management or 
regulatory zones. See \code{\link{read_dat}} for details on importing 
spatial data.}

\item{checks}{(Optional) A list of spatial data quality checks outputted by
\code{spatial_qaqc}.}
}
\description{
Runs the spatial checks performed by \code{\link{spatial_qaqc}} in a shiny
application.
}
\seealso{
\code{\link{spatial_qaqc}}
}
