% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{spatial_hist}
\alias{spatial_hist}
\title{Histogram of latitude and longitude by grouping variable}
\usage{
spatial_hist(dat, group)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{group}{Vector containing grouping categories}
}
\value{
histogram of latitude and longitude by grouping variable
}
\description{
Histogram of latitude and longitude by grouping variable
}
\examples{
\dontrun{
spatial_summary(MainDataTable, 'GEAR_TYPE')
}
}
