% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{spatial_summary}
\alias{spatial_summary}
\title{spatial summary statistics}
\usage{
spatial_summary(
  dat,
  stat.var = c("length", "no_unique_obs", "perc_total", "mean", "median", "min", "max",
    "sum"),
  variable,
  gridfile,
  lon.grid,
  lat.grid,
  lon.dat,
  lat.dat,
  cat
)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{stat.var}{Options are length, no_unique_obs, perc_total, mean, median, min, max, and sum}

\item{variable}{Vector to summarize over date and zone}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{lon.grid}{Column containing longitude data in gridfile.}

\item{lat.grid}{lColumn containing latitude data in gridfile.}

\item{lon.dat}{Column containing longitude data in main data frame.}

\item{lat.dat}{Column containing latitude data in main data frame.}

\item{cat}{Column in gridfile that identifies the individual areas or zones. If gridfile is class sf, `cat` should be name of list containing information on zones.}
}
\description{
Returns plot of selected variable by against date and zone.
}
\details{
\tabular{rlll}{
length: \tab Number of observations \cr
no_unique_obs: \tab Number of unique observations \cr 
perc_total: \tab Percent of total observations \cr 
mean: \tab Mean \cr
median: \tab  Median \cr 
min: \tab  Minimum\cr
max: \tab Maximum \cr
sum: \tab Sum \cr
}
}
