% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dat.R
\name{split_check}
\alias{split_check}
\title{Perfom quality check for \code{split_dat}}
\usage{
split_check(main, split, key, arg)
}
\arguments{
\item{main}{String, column names from MainDataTable.}

\item{split}{String, column names from "split_by" argument.}

\item{key}{String, column names from "key" argument.}

\item{arg}{String, whether "aux" or "split_by" argument was used.}
}
\description{
Perfom quality check for \code{split_dat}
}
\keyword{internal}
