% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SST.R
\docType{data}
\name{SST}
\alias{SST}
\title{Sea Surface Temperature data}
\format{
A data.frame with 7203 rows and 5 variables:
\describe{
\item{time}{Date and time temperature was captured}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{analysed_sst}{analysed sea surface temperature}
\item{sea_ice_fraction}{sea ice area fraction}
}
}
\source{
\url{https://upwell.pfeg.noaa.gov/erddap/info/nasa_jpl_844f_dc54_9c41/index.html}
}
\usage{
SST
}
\description{
Remote Sensing Systems (\url{http://www.remss.com}) sea surface temperature (SST). 
The data includes one week of SST for Bering Sea / Aleutian Islands, 2011.
}
\keyword{datasets}
