% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub_date_check}
\alias{sub_date_check}
\title{Check subset date variable}
\usage{
sub_date_check(sub_date, date, filter_date, group, facet_by)
}
\arguments{
\item{sub_date}{String, name of date column to subset by.}

\item{date}{String, name of date column used in creating period variables.}

\item{filter_date}{The type of date filter to apply to the data.}

\item{group}{String, name of group variable(s). Many fleet function allow 
users to create a year, month, or week variable to group by. If grouping 
by period and \code{sub_date} and \code{date} are null, the function is 
stopped.}

\item{facet_by}{String, name of facetting variable(s). Many fleet function allow 
users to create a year, month, or week variable to facet by. If splitting 
by period and \code{sub_date} and \code{date} are null, the function is 
stopped.}
}
\value{
\code{sub_date}. When used in a function, assign output to \code{sub_date}.
}
\description{
Check subset date variable
}
\keyword{internal}
