% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_date}
\alias{subset_date}
\title{Subset dataset by date value/range}
\usage{
subset_date(dataset, date, filter, value)
}
\arguments{
\item{dataset}{`MainDataTable` to filter.}

\item{date}{String, name of date variable to subset by.}

\item{filter}{String, filter type.}

\item{value}{A range of dates if \code{filter = "date_range"}, or integer if
using a period filter.}
}
\description{
Subset dataset by date value/range
}
\keyword{internal}
