% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_var}
\alias{subset_var}
\title{Subset `MainDataTable` by variable}
\usage{
subset_var(dataset, filter_by, filter_value = NULL, filter_expr = NULL)
}
\arguments{
\item{dataset}{dataset `MainDataTable` to filter.}

\item{filter_by}{String, name of variable used to subset `MainDataTable`.}

\item{filter_value}{Vector of values to subset by (inclusive).}

\item{filter_expr}{String, a valid R expression used to subset `MainDataTable`.}
}
\description{
Subset `MainDataTable` by variable
}
\keyword{internal}
