% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_catch.R
\name{sum_catch}
\alias{sum_catch}
\title{Aggregate species catch by vessel ID}
\usage{
sum_catch(
  dat,
  project,
  catch,
  v_id,
  species = NULL,
  exp,
  val = c("raw", "per"),
  output = c("logical", "table")
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. 
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{catch}{Variable name containing catch weight. A vector of variable names can be given 
if data frame is in wide format (i.e. each species has its own catch weight column).}

\item{v_id}{Variable name containing unique vessel identifier.}

\item{species}{Variable name containing species name if data frame is in long format.}

\item{exp}{An expression wrapped in quotes which will be evaluated after the summary
table has been created.}

\item{val}{Whether to output raw count ('raw') or proportion ('per') of catch. If data frame
is in wide format, the proportion of catch is calculated based on the variables provided
to the \code{catch} argument.}

\item{output}{String, options are ‘logical’ and ‘table’. If 'logical', function returns a vector of logical values based on 
the \code{exp} argument. This can be used to subset tables or assign values. The option 'table' returns a summary table.}
}
\value{
A summary table or vector of logical values based on the \code{exp} argument.
}
\description{
\code{sum_catch} aggregates species catch by vessel ID and outputs
  a logical vector or table.
}
\examples{
\dontrun{
sum_catch(pollockMainDataTable, 'myProject', 'LBS_270_POLLOCK_LBS',
     'LBS_270_POLLOCK_LBS > 3000', val = 'raw', out = 'table')
sum_catch(pollockMainDataTable, 'myProject', 'catch', 
    "species == 'cod' & catch > .5", val = 'per', out = 'logical')
}

}
