% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{View summary statistics for a dataframe}
\usage{
summary_stats(dataset)
}
\arguments{
\item{dataset}{dataframe or matrix over which to apply function}
}
\description{
View summary statistics for a dataframe
}
\details{
This function print summary statistics. It prints the number of unique observations for each column along with standard summary statistics
}
\keyword{statistics}
\keyword{summary}
