% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summary_table}
\alias{summary_table}
\title{Display dataset summary table}
\usage{
summary_table(project, output = "print")
}
\arguments{
\item{project}{Name of project.}

\item{output}{Output type. "print" returns formatted notes. "table" returns a 
dataframe. "print" is recommended for displaying summary table in a report.}
}
\description{
Display dataset summary table
}
\details{
Displays the most recent table created by \code{\link{summary_stats}}
as a dataframe. Can be used in console or notebook.
}
\keyword{internal}
