% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{suppress_table}
\alias{suppress_table}
\title{Suppress confidential values in summary table}
\usage{
suppress_table(
  check,
  output,
  value_var,
  group,
  rule,
  type = "code",
  as_vector = FALSE
)
}
\arguments{
\item{check}{The check table containing suppression conditions.}

\item{output}{The summary table to be edited based on check table.}

\item{value_var}{String, value variable name(s).}

\item{group}{String, grouping variable name(s). This includes
`period` and `facet_by` from summary function.}

\item{rule}{String, the confidentiality rule to apply. \code{rule = "n"} 
suppresses values containing fewer than n vessels. \code{rule = "k"} (the 
"majority allocation rule") suppresses values where a single vessel contains 
k percent or more of the total catch.}

\item{type}{String, the value used to replace confidential data. \code{"code"} 
replaces values with \code{-999}, \code{"NA"} (with quotes) replaces with 
\code{NA}, and \code{"zero"} replaces with 0.}

\item{as_vector}{Logical, whether to return the suppressed values as a vector.
If \code{as_vector == FALSE} the output table is returned.}
}
\description{
This function suppresses values in a summary table based on suppression 
conditions found in the check table (see \code{link{check_confidentiality}})
}
\keyword{internal}
