% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_exists}
\alias{table_exists}
\title{Check if table exists in the fishset_db database}
\usage{
table_exists(table)
}
\arguments{
\item{table}{Name of table in fishset_db database}
}
\value{
True or False
}
\description{
Function utilizes sql functions to check if a table exists in the SQLite fishset_db database.
}
\examples{
\dontrun{  
table_exists('MainDataTable') 
}
}
