% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_fields}
\alias{table_fields}
\title{View fields in the selected table in the fishset_db database}
\usage{
table_fields(table)
}
\arguments{
\item{table}{name of table in sql database}
}
\description{
View fields in the selected table in the fishset_db database
}
