% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{table_format}
\alias{table_format}
\title{Import and format saved tables to notebook file}
\usage{
table_format(x, project)
}
\arguments{
\item{x}{Name of table saved in output}

\item{project}{project name}
}
\description{
Import and format saved tables to notebook file
}
\examples{
\dontrun{
table_format("pollock_species_catch_2020-05-29.csv", 'pollock')
table_format(pull_output("pollock", "species_catch", type = "table"), 'pollock')
}
}
\seealso{
\code{\link{pull_output}}
}
\keyword{internal}
