% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info_verification.R
\name{table_info_verification}
\alias{table_info_verification}
\title{Contains one function that tests several if statements. The function stops if a if statement does not pass.}
\usage{
table_info_verification(dataset, dataindex)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{dataindex}{dataframe that contains information on each column of the dataset}
}
\value{
Returns statements as to whether issues in the data may exist
}
\description{
Contains one function that tests several if statements. The function stops if a if statement does not pass.
}
\details{
checks whether specialized variables have been identified in the index dataset and if units are defined and recognized
}
