% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_remove}
\alias{table_remove}
\title{Remove table from FishSET database}
\usage{
table_remove(table, project)
}
\arguments{
\item{table}{String, name of table in FishSET database. Table name must be in quotes.}

\item{project}{Name of project}
}
\description{
Wrapper for \code{\link[DBI]{dbRemoveTable}}. Remove a table from the FishSET database.
}
\details{
Function utilizes sql functions to remove tables from the FishSET database.
}
\examples{
\dontrun{
table_remove('pollockMainDataTable')
}
}
