% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{table_type}
\alias{table_type}
\title{Detect table type}
\usage{
table_type(tab)
}
\arguments{
\item{tab}{FishSET table.}
}
\description{
Detect table type
}
\examples{
\dontrun{
table_type(pollockMainDataTable) # returns "main"
}
}
\keyword{internal}
