% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{tables_database}
\alias{tables_database}
\title{View names of tables in the fishset_db database}
\usage{
tables_database()
}
\description{
Function utilizes sql functions to view names of tables in the SQLite fishset_db database.
}
\examples{
\dontrun{  
tables_database() 
}
}
