% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot.R
\name{temp_plot}
\alias{temp_plot}
\title{Plot variable by month/year}
\usage{
temp_plot(
  dat,
  project,
  var.select,
  len.fun = "length",
  agg.fun = "mean",
  date.var = NULL,
  alpha = 0.5,
  pages = "single",
  text.size = 8
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{var.select}{Variable in \code{dat} to plot against a date variable.}

\item{len.fun}{Method, \code{"length"} returns the number of observations, 
\code{"unique"} returns the number of unique observations, \code{"percent"}
returns the percentage of total observations.}

\item{agg.fun}{Method to aggregate \code{var.select} by date. Choices are 
\code{"mean"}, \code{"median"}, \code{"min"}, \code{"max"}, or \code{"sum"}.}

\item{date.var}{Date variable in \code{dat}. Defaults to first date variable 
in \code{dat} set if not defined.}

\item{alpha}{The opaqueness of each data point in scatterplot. 0 is total 
transparency and 1 is total opaqueness.  Defaults to .5.}

\item{pages}{Whether to output plots on a single page (\code{"single"}, the 
default) or multiple pages (\code{"multi"}).}

\item{text.size}{Text size of x-axes.}
}
\value{
Returns plot to R console and saves output to the Output folder.
}
\description{
Returns three plots showing the variable of interest against time 
  (as month or month/year). Plots are raw points by date, number of observations 
  by date, and measures of a representative observation by date.
}
\details{
Returns three plots showing the variable of interest against time (as month 
or month/year). Plots are raw points by time, number of observations by time, 
and aggregated variable of interest by time.
}
\examples{
\dontrun{
temp_plot(pollockMainDataTable, project='pollock', 
          var.select = 'OFFICIAL_TOTAL_CATCH_MT', len.fun = 'percent', 
          agg.fun = 'mean', date.var = 'HAUL_DATE')
          
temp_plot(pollockMainDataTable, project='pollock', 
          var.select = 'OFFICIAL_TOTAL_CATCH_MT', len.fun = 'length',
          agg.fun = 'max')
}

}
\keyword{exploration}
\keyword{plot,}
\keyword{temporal,}
