% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenMNSQR}
\alias{tenMNSQR}
\title{Northeast Ten Minute Squares}
\format{
`tenMNSQR`
A simple feature COLLECTION with 5267 features and 9 fields:
\describe{
  \item{AREA}{}
  \item{PERIMETER}{}
  \item{TEN_}{}
  \item{TEN_ID}{}
  \item{LL}{}
  \item{LAT}{}
  \item{LON}{}
  \item{TEMP}{}
  \item{LOC}{}
}
}
\usage{
tenMNSQR
}
\description{
Northeast Ten Minute Squares
}
\keyword{datasets}
