% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_filepath}
\alias{text_filepath}
\title{Create a filepath for a .txt document in the output folder}
\usage{
text_filepath(project, fun_name)
}
\arguments{
\item{project}{Name of project.}

\item{fun_name}{Name of function.}
}
\value{
Useful for saving messages generated in functions.
}
\description{
Create a filepath for a .txt document in the output folder
}
\examples{
\dontrun{
cat("message", file = text_filepath("my_project", "qaqc_output"))
}
}
\keyword{internal}
