% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{to_html_table}
\alias{to_html_table}
\title{Convert dataframe/matrix to html table}
\usage{
to_html_table(x, rownames = FALSE, ...)
}
\arguments{
\item{x}{A vector or list object.}

\item{rownames}{Logical, whether to show rownames.}

\item{...}{Arguments to pass to \code{\link{shiny::renderTable}}}
}
\value{
Returns an HTML table or (if the object is not a matrix or dataframe)
  the original object unmodified.
}
\description{
Convert dataframe/matrix to html table
}
\keyword{internal}
