% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_space}
\alias{trim_space}
\title{trim space function}
\usage{
trim_space(
  x,
  what = c("both", "leading", "trailing", "none"),
  space.regex = "[:space:]",
  ...
)
}
\arguments{
\item{x}{variable of interest}

\item{what}{Choices are both, leading, trailing, none}

\item{space.regex}{Default set to [:space:]}

\item{...}{Additional arguments}
}
\description{
trim space function
}
\keyword{internal}
