% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_dur_out.R
\name{trip_dur_out}
\alias{trip_dur_out}
\title{Trip duration table and plot}
\usage{
trip_dur_out(
  dat,
  project,
  start,
  end,
  units = "days",
  vpue = NULL,
  group = NULL,
  combine = TRUE,
  haul_count = TRUE,
  sub_date = NULL,
  filter_date = NULL,
  date_value = NULL,
  filter_by = NULL,
  filter_value = NULL,
  filter_expr = NULL,
  facet_by = NULL,
  type = "hist",
  bins = 30,
  density = TRUE,
  scale = "fixed",
  tran = "identity",
  format_lab = "decimal",
  pages = "single",
  remove_neg = FALSE,
  output = "tab_plot",
  tripID = NULL,
  fun.time = NULL,
  fun.numeric = NULL
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET
database should contain the string `MainDataTable`.}

\item{project}{String, name of project.}

\item{start}{Date variable containing the start of vessel trip.}

\item{end}{Date variable containing the end of vessel trip.}

\item{units}{Time unit, defaults to \code{"days"}. Options include \code{"secs"},
\code{"mins"}, \code{"hours"}, \code{"days"}, or \code{"weeks"}.}

\item{vpue}{Optional, numeric variable in \code{dat} for calculating value per 
unit effort (VPUE).}

\item{group}{Optional, string names of variables to group by. By default, 
grouping variables are combined unless \code{combine = FALSE} and 
\code{type = "freq_poly"} (frequency polygon). \code{combine = TRUE} will not 
work when \code{type = "hist"} (histogram). Frequency polygon plots can use up 
to two grouping variables if \code{combine = FALSE}: the first variable is 
assigned to the "color" aesthetic and second to the "linetype" aesthetic.}

\item{combine}{Logical, whether to combine the variables listed in \code{group} 
for plot.}

\item{haul_count}{Logical, whether to include hauls per trip in table and/or plot 
(this can only be used if collapsing data to trip level using \code{tripID}. If 
data is already at trip level, add your haul frequency variable to \code{vpue}).}

\item{sub_date}{Date variable used for subsetting, grouping, or splitting by date.}

\item{filter_date}{The type of filter to apply to `MainDataTable`. To filter by a 
range of dates, use \code{filter_date = "date_range"}. To filter by a given
period, use "year-day", "year-week", "year-month", "year", "month", "week", 
or "day". The argument \code{date_value} must be provided.}

\item{date_value}{This argument is paired with \code{filter_date}. To filter
by date range, set \code{filter_date = "date_range"} and enter a  start- and 
end-date into \code{date_value} as a string: 
\code{date_value = c("2011-01-01", "2011-03-15")}. 

To filter by period (e.g. "year", "year-month"), use integers (4 digits if year, 1-2 
digits if referencing a day, month, or week). Use a vector if filtering by 
a single period: \code{date_filter = "month"} and \code{date_value = c(1, 3, 5)}. 
This would filter the data to January, March, and May. 

Use a list if using a year-period type filter, e.g. "year-week", with the 
format: \code{list(year, period)}. For example, \code{filter_date = "year-month"}
and \code{date_value = list(2011:2013, 5:7)} will filter the data table from 
May through July for years 2011-2013.}

\item{filter_by}{String, variable name to filter `MainDataTable` by. the argument 
\code{filter_value} must be provided.}

\item{filter_value}{A vector of values to filter `MainDataTable` by using the 
variable in \code{filter_by}. For example, if \code{filter_by = "GEAR_TYPE"}, 
\code{filter_value = 1} will include only observations with a gear type of 1.}

\item{filter_expr}{String, a valid R expression to filter `MainDataTable` by.}

\item{facet_by}{Variable name to facet by. Facetting by \code{"year"}, \code{"month"}, 
or \code{"week"} provided a date variable is added to \code{sub_date}.}

\item{type}{The type of plot. Options include histogram (\code{"hist"}, the 
default) and frequency polygon (\code{"freq_poly"}).}

\item{bins}{The number of bins used in histogram/freqency polygon.}

\item{density}{Logical, whether densities or frequencies are used for histogram. 
Defaults  to \code{TRUE}.}

\item{scale}{Scale argument passed to \code{\link{facet_grid}}. Defaults to 
\code{"fixed"}. Other options include \code{"free_y"}, \code{"free_x"}, and 
\code{"free_xy"}.}

\item{tran}{Transformation to be applied to the x-axis. A few options include 
\code{"log"}, \code{"log10"}, and \code{"sqrt"}. See 
\code{\link[ggplot2]{scale_continuous}} for a complete list.}

\item{format_lab}{Formatting option for x-axis labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{pages}{Whether to output plots on a single page (\code{"single"}, the 
default) or multiple pages (\code{"multi"}).}

\item{remove_neg}{Logical, whether to remove negative trip durations from the 
plot and table.}

\item{output}{Options include 'table', 'plot', or 'tab_plot' (both table and 
plot, the default).}

\item{tripID}{Column(s) that identify the individual trip.}

\item{fun.time}{How to collapse temporal data. For example, \code{min}, 
\code{mean}, \code{max}. Cannot be \code{sum} for temporal variables.}

\item{fun.numeric}{How to collapse numeric or temporal data. For example, 
\code{min}, \code{mean}, \code{max}, \code{sum}. Defaults to \code{mean}.}
}
\value{
\code{trip_dur_out()} calculates vessel trip duration given a start and end date,
  converts trip duration to the desired unit of time (e.g. weeks, days, or hours),
  and returns a table and/or plot. There is an option for calculating vpue (value 
  per unit of effort) as well. The data can be filtered by date and/or by a variable.
  \code{filter_date} specifies the type of date filter to apply--by date-range or by
  period. \code{date_value} should contain the values to filter the data by. To filter 
  by a variable, enter its name as a string in \code{filter_by} and include the values 
  to filter by in \code{filter_value}. If multiple grouping variables are given then 
  they are combined into one variable  unless \code{combine = FALSE} and 
  \code{type = "freq_poly"}. No more than three grouping variables is recommended 
  if \code{pages = "single"}. Any variable in the dataset can be used for faceting, 
  but "year", "month", and "week" are also available. Distribution plots can be 
  combined on a single page or printed individually with \code{pages}.
}
\description{
Display trip duration and value per unit effort
}
\examples{
\dontrun{
trip_dur_out(pollockMainDataTable,
  start = "FISHING_START_DATE", end = "HAUL_DATE",
  units = "days", vpue = "OFFICIAL_TOTAL_CATCH", output = "plot",
  tripID = c("PERMIT", "TRIP_SEQ"), fun.numeric = sum, fun.time = min
)

}
}
\seealso{
\code{\link{haul_to_trip}}
}
