% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_dur_out.R
\name{trip_duration_plot}
\alias{trip_duration_plot}
\title{Trip duration plot helper}
\usage{
trip_duration_plot(
  trip_tab,
  trp_nms,
  vpue,
  group,
  facet_by,
  units,
  type,
  dens,
  bins,
  tran,
  format_lab,
  scale,
  combine,
  pages
)
}
\arguments{
\item{trip_tab}{Dataframe passed on from \code{trip_dur_out}.}

\item{trp_nms}{Column names of trip length and vpue variables.}

\item{vpue}{Column names of vpue variable(s).}

\item{group}{Column names of grouping variable(s).}

\item{facet_by}{Column names of facet variable(s).}

\item{units}{Units of trip length.}

\item{type}{Whether histogram or frequency polygon is plotted.}

\item{dens}{Logical, whether to calculate density \code{TRUE} or frequency
\code{FALSE}.}

\item{bins}{Numeric, the number of bins to sort variables by.}

\item{tran}{Transformation to be applied to the x-axis.}

\item{format_lab}{Formatting option for x-axis labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{scale}{Scale argument passed to \code{\link{facet_grid}}.}

\item{combine}{Logical, whether grouping variables should be combined.}

\item{pages}{Whether to output plots on a single page (\code{"single"}, the 
default) or multiple pages (\code{"multi"}).}
}
\description{
Creates and formats plots for \code{trip_dur_out}.
}
\keyword{internal}
