% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_closure_scenario.R
\name{unique_closure}
\alias{unique_closure}
\title{Check for unique closure scenarios}
\usage{
unique_closure(project, c_list, ind = TRUE)
}
\arguments{
\item{project}{Name of project}

\item{c_list}{List of closure scenarios to check.}

\item{ind}{Logical, whether to return an index of unique closure scenarios
from \code{c_list} or a single TRUE/FALSE value indicating that one or more
closure scenarios are unique.}
}
\description{
Check for unique closure scenarios
}
\keyword{internal}
