% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{unique_filter}
\alias{unique_filter}
\title{Check rows are unique}
\usage{
unique_filter(dat, project, remove = FALSE)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{remove}{Logical, if TRUE removes non-unique rows. Defaults to FALSE.}
}
\value{
Returns the modified primary dataset with non-unique rows removed if \code{remove} is TRUE.
}
\description{
Check for and remove non-unique rows from primary dataset.
}
\details{
Output is determined by \code{remove}. If \code{remove} is TRUE then
non-unique rows are removed. If \code{remove} is FALSE then only a statement is
returned regarding the number of rows that are not unique.
}
\examples{
\dontrun{
unique_filter(pollockMainDataTable)
mod.dat <- unique_filter(pollockMainDataTable, remove=TRUE)
}
}
