% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_grid.R
\name{unique_grid}
\alias{unique_grid}
\title{Check for unique grid files}
\usage{
unique_grid(project, grid_info, ind = TRUE)
}
\arguments{
\item{project}{Name of project.}

\item{grid_info}{List of grid characteristics used to determine whether a
grid should be saved to the project data folder.}

\item{ind}{Logical, whether to return an index of unique grid or return a 
single logical value.}
}
\value{
\code{TRUE} if a grid is unique (i.e. has no matches in the current
  grid log).
}
\description{
This function determines whether a grid file should be saved to the project
data folder based on the values in \code{grid_info}. If a match is found, 
indicating that an identical grid file has already been saved, it is deleted.
If no match is found the grid file is saved.
}
\keyword{internal}
