% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_rows}
\alias{unique_rows}
\title{Return unique rows}
\usage{
unique_rows(dat)
}
\arguments{
\item{dat}{Data to check for non-unique rows.}
}
\description{
Check whether all rows are in a dataset are unique. If not, returns only the
unique rows.
}
\examples{
\dontrun{
dat <- unique_rows(dat)
}
}
\keyword{internal}
