% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{unserialize_table}
\alias{unserialize_table}
\title{Unserialize special tables in FishSET DB}
\usage{
unserialize_table(table, project)
}
\arguments{
\item{table}{The name of the special table to unserialize. Special tables
include alternative choice matrix output, expected catch matrix output,
the model data list, and the prediction output.}

\item{project}{Name of project.}
}
\description{
Unserialize special tables in FishSET DB
}
\keyword{internal}
