% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_folderpath}
\alias{update_folderpath}
\title{Update FishSETFolder location}
\usage{
update_folderpath()
}
\description{
Select the location of the FishSET folder. This can be helpful if 
switching between different FishSET folders or if 'folderpath' is 
inaccurate.
}
\keyword{internal}
