% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{user_locoutput}
\alias{user_locoutput}
\title{Print user folder directory}
\usage{
user_locoutput()
}
\description{
Print user folder directory
}
\details{
This function looks for an object named "locuser" with a valid folder directory. 
 if it doesn't find one, it asks user to set a valid directory. This is used for
 inserting plots and tables from a folder outside the FishSET package.
}
\keyword{internal}
