% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_count.R
\name{vessel_count}
\alias{vessel_count}
\title{Vessel count}
\usage{
vessel_count(
  dat,
  project,
  v,
  t,
  period = "month",
  group = NULL,
  position_grp = "stack",
  output = c("table", "plot"),
  ...
)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db 
database should contain the string `MainDataTable`.}

\item{project}{name of project.}

\item{v}{vessel ID variable to count.}

\item{t}{time variable containing dates to aggregate by.}

\item{period}{Time period to count by. Options include "year", "year_abv", 
"month", "month_abv", "month_num", "weeks" (weeks in the year), 
"weekday", "weekday_abv", "weekday_num", "day" (day of the month), 
and "day_of_year".}

\item{group}{factor variable to group count by.}

\item{position_grp}{Positioning of bar plot. Options include "stack", "dodge", 
and "fill".}

\item{output}{table or plot.}

\item{...}{other arguments passed on to \code{\link{aggregate}}.}
}
\value{
Table or plot of the number of unique vessels within a time period.
}
\description{
Generates a table or plot of the number of unique vessels by time period
}
