% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vgsub}
\alias{vgsub}
\title{vgsub function}
\usage{
vgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{pattern}

\item{replacement}{replacement}

\item{x}{x}

\item{...}{Additional arguments}
}
\description{
vgsub function
}
