% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{view_fleet_table}
\alias{view_fleet_table}
\title{View the most recent fleet table by project}
\usage{
view_fleet_table(project)
}
\arguments{
\item{project}{The name of project.}
}
\description{
View the most recent fleet table by project
}
\examples{
\dontrun{
view_fleet_table("pollock")
}
}
