% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_spat.R
\name{view_lon_lat}
\alias{view_lon_lat}
\title{View}
\usage{
view_lon_lat(dat, lon, lat, id = NULL, crs = 4326)
}
\arguments{
\item{dat}{Data containing \code{lon} and \code{lat} columns.}

\item{lon}{Name of Longitude column.}

\item{lat}{Name of Lattitude column.}

\item{id}{Optional, name of an ID variable that is paired with \code{lon} and 
\code{lat} columns}

\item{crs}{Optional, coordinate reference system to use. Defaults to EPSG code
4326 (WGS 84).}
}
\description{
View
}
