% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{view_model_design}
\alias{view_model_design}
\title{View model design file in database}
\usage{
view_model_design(project, date = NULL)
}
\arguments{
\item{project}{Project name.}

\item{date}{String, date model design file was created.}
}
\description{
View model design file in database
}
