% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{week_labeller}
\alias{week_labeller}
\title{X-axis scale labeller for weekly data}
\usage{
week_labeller(breaks, year)
}
\arguments{
\item{breaks}{scale label breaks}

\item{year}{vector of years from summary table}
}
\description{
X-axis scale labeller for weekly data
}
\keyword{internal}
