% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_effort.R
\name{weekly_effort}
\alias{weekly_effort}
\title{Summarize average CPUE by week}
\usage{
weekly_effort(
  dat,
  project,
  cpue,
  date,
  group = NULL,
  sub_date = NULL,
  filter_date = NULL,
  date_value = NULL,
  filter_by = NULL,
  filter_value = NULL,
  filter_expr = NULL,
  facet_by = NULL,
  conv = "none",
  tran = "identity",
  format_lab = "decimal",
  combine = FALSE,
  scale = "fixed",
  output = "tab_plot",
  format_tab = "wide"
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{cpue}{Variable(s) in \code{dat} containing catch per unit effort.}

\item{date}{A variable in \code{dat} containing dates to aggregate by.}

\item{group}{Grouping variable name(s). Up to two grouping variables are available.
For plotting, if a single CPUE column is entered the first grouping variable is 
passed to the "color" aesthetic and the second to "linetype". If multiple CPUE
columns are entered, a new variable named "species" is created and passed to "fill", 
the first group variable to "linetype", and second is dropped.}

\item{sub_date}{Date variable used for subsetting, grouping, or splitting by date.}

\item{filter_date}{The type of filter to apply to `MainDataTable`. To filter by a 
range of dates, use \code{filter_date = "date_range"}. To filter by a given period, use
"year-day", "year-week", "year-month", "year", "month", "week", or "day". 
The argument \code{date_value} must be provided.}

\item{date_value}{This argument is paired with \code{filter_date}. If 
\code{filter_date = "date_range"}, enter a string containing a start- and end-date, 
e.g. \code{date_value = c("2011-01-01", "2011-03-15")}. If filtering by period 
(e.g. "year", "year-month"), use integers (4 digits if year, 1-2 digits if 
referencing a day, month, or week). Use a list if using a year-period type filter, 
e.g. "year-week", with the format: \code{list(year, period)}. Use a vector if 
using a single period (e.g. "month"): \code{c(period)}. For example, 
\code{date_value = list(2011:2013, 5:7)} will filter the data table from May 
through July for years 2011-2013 if \code{filter_date = "year-month"}.
\code{date_value = c(2:5)} will filter the data from February through May when 
\code{filter_date = "month"}.}

\item{filter_by}{String, variable name to filter `MainDataTable` by. the argument 
\code{filter_value} must be provided.}

\item{filter_value}{A vector of values to filter `MainDataTable` by using the 
variable in \code{filter_by}. For example, if \code{filter_by = "GEAR_TYPE"}, 
\code{filter_value = 1} will include only observations with a gear type of 1.}

\item{filter_expr}{String, a valid R expression to filter `MainDataTable` by 
using the variable in \code{filter_by}.}

\item{facet_by}{Variable name to facet by. Accepts up to two variables. Facetting 
by \code{"year"} is available if a date variable is added to \code{sub_date}. 
Facetting by \code{"species"} is available if multiple cpue columns are included 
in \code{"cpue"}. The first variable is facetted by row and the second by column.}

\item{conv}{Convert catch variable to \code{"tons"}, \code{"metric_tons"}, or 
by using a function entered as a string. Defaults to \code{"none"} for no conversion.}

\item{tran}{A function to transform the y-axis. Options include log, log2, log10, 
sqrt.}

\item{format_lab}{Formatting option for y-axis labels. Options include 
\code{"decimal"} or \code{"scientific"}.}

\item{combine}{Whether to combine variables listed in \code{group}. This is passed
to the "color" aesthetic for plots.}

\item{scale}{Scale argument passed to \code{\link[ggplot2]{facet_grid}}. Defaults 
to \code{"fixed"}.}

\item{output}{Whether to display \code{"plot"}, \code{"table"}. Defaults 
to both (\code{"tab_plot"}).}

\item{format_tab}{How table output should be formatted. Options include 'wide' 
(the default) and 'long'.}
}
\value{
\code{weekly_effort()} calculates mean CPUE by week. This function doesn't 
  calculate CPUE; the CPUE variable must be created in advance (see \code{\link{cpue}}).
  When multiple CPUE variables are entered, a new column named "species" 
  is created and used to group values in plots. 
  The "species" column can also be used to split (or facet) the plot. For table output,
  the "species" column will be kept if \code{format_tab = "long"}, i.e. a column of
  species names ("species") and a column containing the mean CPUE ("mean_cpue"). When 
  \code{format_tab = "wide"}, each CPUE variable is given its own value column.  
  The data can be filtered by date and/or by a variable. \code{filter_date} 
  specifies the type of date filter to apply--by date-range or by period. 
  \code{date_value} should contain the values to filter the data by. To filter 
  by a variable, enter its name as a string in \code{filter_by} and include the 
  values to filter by in \code{filter_value}. Up to two grouping variables can 
  be entered. Grouping variables can be merged into one variable using \code{combine}; 
  in this case any number of variables can be joined, but no more than three is 
  recommended. For faceting, any variable (including ones listed in \code{group}) 
  can be used, but "year" and "species" are also available. Facetting by "year" requires a date 
  variable be added to \code{sub_date}. Currently, combined variables cannot be 
  faceted. A list containing a table and plot are printed to the console and 
  viewer by default.
}
\description{
Summarize average CPUE by week
}
\examples{
\dontrun{
weekly_effort(pollockMainDataTable, "CPUE", "DATE_FISHING_BEGAN", filter_date = "year", 
              date_value = 2011, output = "table")
}
}
