% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welfare_outputs.R
\name{welfare_outputs}
\alias{welfare_outputs}
\title{Welfare plots and tables}
\usage{
welfare_outputs(project, closures, betadraws = 1000)
}
\arguments{
\item{project}{Name of project}

\item{closures}{Closure scenarios}

\item{betadraws}{Integer indicating the numer of times to run the welfare simulation. Default value is
\code{betadraws = 1000}}
}
\description{
Generate plots and tables for welfare simulations
}
\details{
Returns a list with (1) plot showing welfare loss/gain for all scenarios in dollars, (2) plot showing welfare
  loss/gain as percentage, (3) dataframe with welfare summary stats in dollars, (4) dataframe with welfare summary
  stats as percentages, and (5) dataframe with welfare details such as number of trips, mean loss per trip, and mean of 
  the total welfare loss across all trips.
}
