% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windLease}
\alias{windLease}
\title{Northeast wind closure areas}
\format{
`windLease`
Simple features collection with 32 features and 1 field:
\describe{
  \item{NAME}{Name of wind lease.}
}
}
\usage{
windLease
}
\description{
Northeast wind closure areas
}
\keyword{datasets}
