% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidentiality.R
\name{window_cc}
\alias{window_cc}
\title{Unique values window function}
\usage{
window_cc(x, k)
}
\arguments{
\item{x}{List of vessel IDs to count}

\item{k}{Window width.}
}
\description{
Used by \code{roll_catch()} to count unique vessels within a rolling window.
}
\seealso{
\code{\link{roll_catch}} \code{\link{roll_cc}}
}
\keyword{internal}
