% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_plot.R
\name{xy_plot}
\alias{xy_plot}
\title{Evaluate relationship of two variables in a plot format}
\usage{
xy_plot(dat, project, var1, var2, regress = FALSE)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{var1}{First variable in \code{dat}.}

\item{var2}{Second variable in \code{dat}.}

\item{regress}{Logical, if TRUE, returns plot with fitted linear regression line. Defaults to FALSE.}
}
\value{
Returns plot output to R console and saves plot to Output folder.
}
\description{
Evaluate relationship of two variables in a plot format. 
  Plots first variable against second variable.

Plot of var1 against var 2
}
\examples{
\dontrun{
xy_plot(pollockMainDataTable, 'OFFICIAL_TOTAL_CATCH_MT', 'HAUL', regress=TRUE)
}
}
\keyword{plot}
\keyword{xy}
