% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_closure.R
\name{zone_closure}
\alias{zone_closure}
\title{Define zone closure scenarios}
\usage{
zone_closure(
  project,
  spatname,
  mainname = NULL,
  zone_spat,
  zone_main = NULL,
  lon_spat = NULL,
  lat_spat = NULL,
  epsg = NULL
)
}
\arguments{
\item{project}{Required, name of project.}

\item{spatname}{Required, data file or character. 
\code{spat} is a spatial data file containing information on fishery 
management or regulatory zone boundaries. Shape, json, geojson, and csv 
formats are supported. geojson is the preferred format. json files must be 
converted into geojson. This is done automatically when the file is loaded 
with \code{\link{read_dat}} with \code{is.map} set to true. \code{spat} 
cannot, at this time, be loaded from the FishSET database. \cr}

\item{mainname}{Optional, data file or character. The main dataset containing
non-spatial records that can be used to select zones based on existing variables.}

\item{zone_spat}{Variable in \code{spatname} that identifies the individual areas or zones.}

\item{zone_main}{Variable in \code{mainname} that identifies the individual areas or zones.
Required if \code{mainname} is provided.}

\item{lon_spat}{Required for csv files. Variable or list from \code{spat} 
containing longitude data. Leave as NULL if \code{spat} is a shape or json file.}

\item{lat_spat}{Required for csv files. Variable or list from \code{spat} 
containing latitude data.  Leave as NULL if \code{spat} is a shape or json file.}

\item{epsg}{EPSG number. Set the epsg to ensure that \code{spat} has the correct projections. 
If epsg is not specified but is defined for \code{spat}. 
See \url{http://spatialreference.org/} to help identify the optimal epsg number.}
}
\value{
A Shiny application object.
}
\description{
Define zone closure scenarios
}
\details{
Define zone closure scenarios via an interactive app. 
  Users can define scenarios by clicking on one or more zones on the map, 
  adjusting the allowable TAC percentages in the table, and entering a unique scenario name. 
  Clicking 'Add closure' instantly saves the scenario to the project database. 
  These saved choices are later called in the policy scenario function.
}
