% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumarray.R
\name{accumarray}
\alias{accumarray}
\title{Construct array with accumulation}
\usage{
accumarray(subs, val, sz = NULL, func = sum, fillval = 0)
}
\arguments{
\item{subs}{}

\item{val}{}
}
\value{
an array
}
\description{
Construct array with accumulation
}
\details{
Based on the accumarray function in Matlab. Modified from the pracma package. Called in the createAlternativeChoice function.
}
\examples{
accumarray(C, C)
}
