% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vars.R
\name{add_vars}
\alias{add_vars}
\title{Add removed variables back into dataset - non-interactive version}
\usage{
add_vars(working_dat, raw_dat, vars, project)
}
\arguments{
\item{working_dat}{Primary data containing information on hauls or trips. Table in FishSET database contains the string `MainDataTable`.}

\item{raw_dat}{Unmodified raw version of the primary dataset. Should be a character specifying a table from the FishSET database
containing the string ‘MainDataTable’ and date table was created.}

\item{vars}{Character string, variables from \code{raw_dat} to add back into \code{working_dat}.}

\item{project}{Character, name of project. Parameter is used to generate meaningful table names in FishSET database.}
}
\description{
Add columns that have been removed from the primary dataset back into the primary dataset.
}
\details{
Add variables back into the dataset that were removed.
The removed variables are obtained from the \code{raw_dat} and merged into the working data based on a row identifier.
The row identifier is created when a variable is removed using the \code{\link{select_vars}} function.
The row identifier is used to match the raw data variables to \code{working_dat}.
}
\examples{
\dontrun{
add_vars(pcodMainDataTable, "pcodMainDataTable20200410", "pollock")
}

}
