% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_column.R
\name{assignment_column}
\alias{assignment_column}
\title{Assign observations to zone or area}
\usage{
assignment_column(dat, gridfile, hull.polygon = c(TRUE, FALSE), lon.grid,
  lat.grid, lon.dat, lat.dat, cat, closest.pt = c(TRUE, FALSE),
  epsg = NULL)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{hull.polygon}{T/F If TRUE, creates convex hull polygon. Use if spatial data creating polygon are sparse or irregular.}

\item{lon.grid}{Column containing longitude data in gridfile.}

\item{lat.grid}{lColumn containing latitude data in gridfile.}

\item{lon.dat}{Column containing longitude data in main data frame.}

\item{lat.dat}{Column containing latitude data in main data frame.}

\item{cat}{Column in gridfile that identifies the individual areas or zones. If gridfile is class sf, `cat` should be name of list containing information on zones.}

\item{closest.pt}{TRUE/FALSE If true, zone ID identified as the closest polygon to the point.}

\item{epsg}{EPSG number. See \url{http://spatialreference.org/} to help identify optimal epsg number.}
}
\value{
Main data frame with new assignment column labeled zoneID.
}
\description{
Assign observations to zone or area
}
\details{
Function is called by the \code{\link{create_alternative_choice}} function to assign each observation to zones defined by a spatial data set.
Converts point data from gridfile into polygons and then finds which polygon each observation in the main data frame is within. Use hull.polygon=T if data is sparse or irregular.
}
\keyword{polygon}
\keyword{zone,}
