% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp.R
\name{calc_exp}
\alias{calc_exp}
\title{Calculate expected catch/revenue}
\usage{
calc_exp(
  dataset,
  catch,
  price = NULL,
  defineGroup = NULL,
  temp.var = NULL,
  temporal = "daily",
  calc.method = "standardAverage",
  lag.method = "simple",
  empty.catch = NULL,
  empty.expectation = NULL,
  temp.window = 7,
  temp.lag = 0,
  year.lag = 0,
  dummy.exp = FALSE,
  Alt
)
}
\arguments{
\item{dataset}{Data}

\item{catch}{catch}

\item{price}{price}

\item{defineGroup}{define group}

\item{temp.var}{temporal variable}

\item{temporal}{temporal}

\item{calc.method}{calculation method}

\item{lag.method}{lag method}

\item{empty.catch}{empty catch}

\item{empty.expectation}{empty expectation}

\item{temp.window}{temporal window}

\item{temp.lag}{temporal lag}

\item{year.lag}{year lag}

\item{dummy.exp}{dummy matrix}

\item{Alt}{Alternative choice list}
}
\value{
Returns a list containing the expected catch/revenue matrix and 
  dummy matrix (if \code{dummy.exp = TRUE}).
}
\description{
Calculate expected catch/revenue
}
\keyword{internal}
