% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_lm.R
\name{catch_lm}
\alias{catch_lm}
\title{Linear Model for Catch}
\usage{
catch_lm(
  dat,
  project,
  catch.formula,
  zoneID = NULL,
  exp.name = NULL,
  new.name = NULL,
  output = "ec matrix"
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET
database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{catch.formula}{A formula object specifying the linear model. See
\code{\link[stats:lm]{stats::lm()}}.}

\item{zoneID}{zone ID Variable in \code{dat} that identifies the individual zones or
areas.}

\item{exp.name}{Name(s) of expected catch matrix to merge into \code{dat}.}

\item{new.name}{Optional, new name for \code{exp.name}. These should be in the same
order as \code{exp.name}. \code{catch.formula} must use new names.}

\item{output}{Whether to output \code{dat} with a single variable added (\code{'variable'})
or}
}
\description{
First stage regression model for catch. Can be used to
}
