% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_data.R
\name{check_model_data}
\alias{check_model_data}
\title{Check data for data quality issues that will affect modeling functions.}
\usage{
check_model_data(dat, uniqueID, save.name, save.file = TRUE)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{uniqueID}{Variable that identifies unique occurrences.}

\item{save.name}{Name for verified and saved data.}

\item{save.file}{TRUE/FALSE Save to fishset_db SQLite database? Defaults to TRUE}
}
\value{
Returns statements as to data quality issues in the data may exist.
}
\description{
Check data for data quality issues that will affect modeling functions.
}
\details{
Checks data to be used for modeling for presence of NAs, NaNs and Inf and that each row is a unique choice occurrence. 
   Model functions may fail or return inaccurate results if any of these issues exist.
   The verified data will not save if any of these issues are in the data set. 
   If data passes all tests, the verified data can then be saved as SQL table or csv file. 
   When saving to SQL, the modified table will be saved and the previous, unmodified version of the table 
   (if it exists), will be saved with the prefix 'raw'.
}
\examples{
\dontrun{ 
check_model_data(MainDataTable, uniqueID='uniqueID_Code', save.name='modelData')
}
}
