% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_out.R
\name{corr_plot}
\alias{corr_plot}
\title{Correlation plot}
\usage{
corr_plot(corr, p.val, show_coef, project)
}
\arguments{
\item{corr}{Correlation matrix}

\item{p.val}{Correlation matrix of p values.}

\item{show_coef}{Whether to show correlation coefficients.}

\item{project}{Name of project}
}
\description{
Correlation plot
}
\keyword{internal}
