% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVariables.R
\name{cpue}
\alias{cpue}
\title{Create variable or matrix}
\usage{
cpue(dataset, xWeight, xTime)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{dataindex}{dataframe that contains information on each column of the dataset}

\item{...}{Column(s) that define the variable(s) over which to make new ID}

\item{x}{Variable to create}

\item{DumFill}{Fill the dummy variable with TRUE or FALSE}
}
\value{
Returns the dataframe with new Dummy variable included
}
\description{
Create variable or matrix
}
\details{
Function for generating specialized variables.
}
