% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{cpue}
\alias{cpue}
\title{Create catch per unit effort variable}
\usage{
cpue(dataset, xWeight, xTime)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{xWeight}{Weight variable}

\item{xTime}{Time variables. Must be weeks, days, hours, or minutes}
}
\description{
Create catch per unit effort variable
}
\details{
Function for generating new or specialized variables. cpue function create catch per unit effort variable.
}
