% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternative_choice.R
\name{create_alternative_choice}
\alias{create_alternative_choice}
\title{Create alternative choice matrix}
\usage{
create_alternative_choice(
  dat,
  gridfile,
  case = c("Centroid", "Port", "Other"),
  min.haul,
  alt_var,
  occasion,
  dist.unit = "miles",
  lon.dat,
  lat.dat,
  lon.grid,
  lat.grid,
  cat,
  hull.polygon = c(TRUE, FALSE),
  closest.pt = FALSE,
  project,
  griddedDat = NULL,
  weight.var = NULL
)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{case}{Centroid='Centroid of Zonal Assignment', Port, Other}

\item{min.haul}{Minimum number of hauls. Removes zones with fewer hauls than min.haul. For instance, include only zones with at least 100 hauls.}

\item{alt_var}{Identifies how to find lat/lon for starting point (must have a lat/lon associated with it)}

\item{occasion}{Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment'}

\item{dist.unit}{Distance returned as meters, kilometers, or miles. Defaults to miles.}

\item{lon.dat}{Longitude variable in dataset}

\item{lat.dat}{Latitude variable in dataset}

\item{lon.grid}{Longitude variable in gridfile}

\item{lat.grid}{Latitude variable in gridfile}

\item{cat}{Variable defining zones or areas. Must be defined for dataset or gridfile.}

\item{hull.polygon}{Used in assignment_column function. Creates polygon using convex hull method.}

\item{closest.pt}{TRUE/FALSE If true, zone ID identified as the closest polygon to the point. Called in assignment_column function.}

\item{project}{Name of project. Used for naming table saved to database}

\item{griddedDat}{Data frame containing a variable that varies by the map grid. First column should match a column in the dataset. The remaining columns should match the zone IDs in the gridfile.}

\item{weight.var}{variable weighted centroids}
}
\value{
Alternative choice matrix\cr
\tabular{rlll}{                                                                                                                                                                                                           
        dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that zone is to be included in the model\cr
        greaterNZ: \tab zone which pass numofNecessary test\cr
        numOfNecessary: \tab Minimum number of hauls necessary within a zone for zone to be included\cr
        altChoiceUnits: \tab set to miles\cr
        altChoiceType: \tab set to distance\cr
        alt_var: \tab Identifies how to find lat/lon for starting point\cr
        occasion: \tab Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment' \cr
        zoneRow: \tab zones and choices array\cr
        int: \tab Centroid for each zone. Generated from find_centroid function
        }
}
\description{
Creates a list containing information on how alternative fishing choices should be defined.
}
\details{
Functions returns alternative choice matrix. Function must be called before the create_expectations, make_model_design, or discretefish_subroutine functions can be called.
Function identifies which zones are to be included, based on number of hauls per trip, and assigns each observation to a zone.
}
