% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternative_choice.R
\name{create_alternative_choice}
\alias{create_alternative_choice}
\title{Define alternative fishing choice}
\usage{
create_alternative_choice(
  dat,
  project,
  occasion = "zonal centroid",
  occasion_var = NULL,
  alt_var = "zonal centroid",
  dist.unit = "miles",
  min.haul = 0,
  zoneID,
  zone.cent.name = NULL,
  fish.cent.name = NULL,
  spatname = NULL,
  spatID = NULL,
  outsample = FALSE
)
}
\arguments{
\item{dat}{Required, Primary data frame containing data on hauls or trips.
Table in FishSET database should contain the string \code{MainDataTable}.}

\item{project}{Required, name of project.}

\item{occasion}{String, determines the starting point when calculating the
distance matrix. Options are \code{"zonal centroid"}, \code{"fishing centroid"},
\code{"port"}, or \code{"lon-lat"}. See \code{occasion_var} for requirements.}

\item{occasion_var}{Identifies an ID column or set of lon-lat variables needed
to create the distance matrix. Possible options depend on the value of
\code{occasion}:
\describe{
\item{Centroid}{When \verb{occasion = zonal/fishing centroid} the possible
options are \code{NULL}, the name of a zone ID variable, or a set coordinate
variables (in Lon-Lat order).
\describe{
\item{NULL}{This will merge centroid lon-lat data to the primary table
using the column enter in \code{zoneID}. A centroid table must be saved
to the FishSET Database.}
\item{Zone ID}{This option specifies the zone ID variable to merge the
centroid table to. For example, a column containing the previous zonal
area. A centroid table must be saved to the FishSET Database.}
\item{Lon-Lat}{A string vector of length two containing the longitude
and latitude of an existing set centroid variables in \code{dat}.}
}
}
\item{Port}{When \code{occasion = port} the possible options include the name
of a port ID variable or a set of lon-lat variables describing the
location of the port. A value of \code{NULL} will return an error.
\describe{
\item{Port ID}{The name of a port ID variable in \code{dat} that will be
used to join the port table to the primary table. A port table
is required (see \code{\link[=load_port]{load_port()}}) which contains the port name and
the longitude and latitude of each port.}
\item{Lon-Lat}{A string vector of length two containing a port's
longitude and latitude in \code{dat}.}
}
}
\item{Lon-Lat}{When \code{occasion = lon-lat}, \code{occasion_var} must contain a
string vector of length two containing the longitude and latitude of a
vessel's location in the \code{dat}. For example, the current or previous
haul location.}
}}

\item{alt_var}{Determines the alternative choices used to calculate the distance
matrix. \code{alt_var} may be the centroid of zonal assignment (\code{"zonal centroid"}),
\code{"fishing centroid"}, or the closest point in fishing zone
(\code{"nearest point"}). The centroid options require that the appropriate
centroid table has been saved to the project's FishSET Database. See
\code{\link[=create_centroid]{create_centroid()}} to create and save centroids. List existing centroid
tables  by running \code{list_tables("project", type = "centroid")}.}

\item{dist.unit}{String, how distance measure should be returned. Choices are
\code{"meters"} or \code{"m"}, \code{"kilometers"} or \code{"km"}, \code{"miles"}, or \code{"nmiles"}
(nautical miles). Defaults to \code{"miles"}.}

\item{min.haul}{Required, numeric, minimum number of hauls. Zones with fewer
hauls than the \code{min.haul} value will not be included in model data.}

\item{zoneID}{Variable in \code{dat} that identifies the individual zones or
areas.}

\item{zone.cent.name}{The name of the zonal centroid table to use when
\code{occasion} or \code{alt_var} is set to \verb{zonal centroid}. Use
\code{list_tables("project", type = "centroid")} to view existing centroid tables.
See \code{\link[=create_centroid]{create_centroid()}} to create centroid tables or \code{\link[=centroid_to_fsdb]{centroid_to_fsdb()}} to
create a centroid table from columns found in \code{dat}.}

\item{fish.cent.name}{The name of the fishing centroid table to use when
\code{occasion} or \code{alt_var} is set to \verb{fishing centroid}. Use
\code{list_tables("project", type = "centroid")} to view existing centroid tables.
See \code{\link[=create_centroid]{create_centroid()}} to create centroid tables or \code{\link[=centroid_to_fsdb]{centroid_to_fsdb()}} to
create a centroid table from columns found in \code{dat}.}

\item{spatname}{Required when \code{alt_var = 'nearest point'}. \code{spat} is a spatial
data file  containing information on fishery management or regulatory zones
boundaries. \code{sf} objects are recommended, but \code{sp} objects can be used as
well. See \code{\link[=dat_to_sf]{dat_to_sf()}} to convert a spatial table read from a csv file to
an \code{sf} object. To upload your spatial data to the FishSETFolder see
\code{\link[=load_spatial]{load_spatial()}}.If \code{spat} should come from the FishSET database, it should
be the name of the original file name, in quotes. For example,
\code{"pollockNMFSZonesSpatTable"}. Use \code{\link[=tables_database]{tables_database()}} or
\code{list_tables("project", type = "spat")} to view the names of spatial tables
in the FishSET database.}

\item{spatID}{Required when \code{alt_var = 'nearest point'}. Variable in \code{spat}
that identifies the individual zones or areas.}

\item{outsample}{Logical, indicating whether this is for primary data or out-of
sample data.}
}
\value{
Saves the alternative choice list to the FishSET database as a list.
Output includes: \cr
\tabular{rlll}{
dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that
zone is to be included in the model\cr
greaterNZ: \tab Zones which pass numofNecessary test\cr
numOfNecessary: \tab Minimum number of hauls for zone to be included\cr
altChoiceUnits: \tab Set to miles\cr
altChoiceType: \tab Set to distance\cr
occasion: \tab Identifies how to find latitude and longitude for starting point\cr
occasion_var: \tab Identifies how to find latitude and longitude for starting point\cr
alt_var: \tab Identifies how to find latitude and longitude for alternative choice \cr
zoneRow: \tab Zones and choices array\cr
zone_cent: \tab Geographic centroid for each zone. Generated from \code{\link[=find_centroid]{find_centroid()}}\cr
fish_cent: \tab Fishing centroid for each zone. Generated from \code{\link[=find_fishing_centroid]{find_fishing_centroid()}}\cr
zone_cent_name: \tab Name of the zonal centroid table\cr
fish_cent_name: \tab Name of the fishing centroid table\cr
spat: \tab Spatial data object\cr
spatID: \tab Variable in spat that identifies individuals zones
}
}
\description{
Required step. Creates a list identifying how alternative fishing choices should
be defined. Output is saved to the FishSET database. Run this function before
running models. \code{dat} must have a zone assignment column (see
\code{\link[=assignment_column]{assignment_column()}}). In certain cases a centroid table must be saved to
the FishSET Database, see \code{occasion_var} for details.
}
\details{
Defines the alternative fishing choices. These choices are used to develop
the matrix of distances between observed and alternative fishing choices (where
they could have fished but did not). The distance matrix is calculated by the
\code{\link[=make_model_design]{make_model_design()}} function. \code{occasion} defines the observed fishing
location and \code{alt_var} the alternative fishing location. \code{occasion_var}
identifies an ID column or set of lon-lat variables needed to create the
distance matrix.

Parts of the alternative choice list are pulled by \code{\link[=create_expectations]{create_expectations()}},
\code{\link[=make_model_design]{make_model_design()}}, and the model run \code{\link[=discretefish_subroutine]{discretefish_subroutine()}})
functions. These output include choices of which variable to use for catch and
which zones to include in analyses based on a minimum number of hauls per trip
within a zone. Note that if the alternative choice list is modified, the
\code{\link[=create_expectations]{create_expectations()}} and \code{\link[=make_model_design]{make_model_design()}} functions
should also be updated before rerunning models.
}
