% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternative_choice.R
\name{create_alternative_choice}
\alias{create_alternative_choice}
\title{Create alternative choice matrix}
\usage{
create_alternative_choice(dataset, gridfile, case = c("Centroid", "Port",
  "Other"), contents, Haul.Trip = c("Haul", "Trip"), alt_var, occasion,
  lon.dat, lat.dat, lon.grid, lat.grid, cat, use.grid = c(TRUE, FALSE),
  weight.var = NULL, hull.polygon = c(TRUE, FALSE),
  remove.na = FALSE, closest.pt = FALSE)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{gridfile}{name of gridded dataset}

\item{case}{Centroid='Centroid of Zonal Assignment', Port, Other}

\item{contents}{Value of variable to subset dataset by. For instance, include only zones with at least 10 hauls.}

\item{Haul.Trip}{Whether data is at trip or haul. Default to haul.}

\item{alt_var}{# Identifies how to find lat/lon for starting point (must have a lat/lon associated with it)}

\item{occasion}{# Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment'}

\item{lon.dat}{#Longitude variable in dataset}

\item{lat.dat}{#Latitude variable in dataset}

\item{lon.grid}{#Longitude variable in gridfile}

\item{lat.grid}{#Latitude variable in gridfile}

\item{cat}{#Variable defining zones or areas. Must be defined for dataset or gridfile.}

\item{use.grid}{#TRUE/FALSE. If TRUE, gridded data is used to create centroids}

\item{weight.var}{#Variable for weighted centroids}

\item{hull.polygon}{Using in assignmentColumn function. Creates polying using convex hull method.}

\item{remove.na}{TRUE/FALSE Remove points where zone ID not identified. Called in assignment_column function.}

\item{closest.pt}{TRUE/FALSE If true, zone ID identified as the closest polygon to the point. Called in assignment_column function.}
}
\value{
returns list containing information on alternative choice
}
\description{
Create alternative choice matrix
}
\details{
Functions returns alternative choice matrix
#outputs:                                                                                                                                                                                                                   
        Alt.dataZoneTrue - # array of logical values to identify which are to be used in model                                                                                                                           
        Alt.greaterNZ                                                                                                                                                                                                     
        Alt.numOfNecessary                                                                                                                                                                                                
        Alt.altChoiceUnits                                                                                                                                                                                                
        Alt.alt_var- alternative variable                                                                                                                                                                                 
        Alt.occassion- occassion variable                                                                                                                                                                                   
        Alt.functionType                                                                                                                                                                                                  
        Alt.zoneRow                                                                                                                                                                                                       
        Alt.zoneType 
        Alt.int
}
