% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternative_choice.R
\name{create_alternative_choice}
\alias{create_alternative_choice}
\title{Define alternative fishing choice}
\usage{
create_alternative_choice(
  dat,
  project,
  occasion = "centroid",
  alt_var = "centroid",
  dist.unit = "miles",
  min.haul = 0,
  spat,
  cat = NULL,
  zoneID = NULL,
  lon.dat = NULL,
  lat.dat = NULL,
  hull.polygon = FALSE,
  closest.pt = FALSE,
  grid = NULL,
  lon.spat = NULL,
  lat.spat = NULL
)
}
\arguments{
\item{dat}{Required, main data frame containing data on hauls or trips.
Table in FishSET database should contain the string `MainDataTable`.}

\item{project}{Required, name of project}

\item{occasion}{String, identifies how to find lat/lon for starting point 
(must have a lon/lat associated with it). \code{occasion} may be the ‘centroid 
of zonal assignment’, a port variable or lon/lat variable(s) in the primary 
dataset. If a port variable is defined, a corresponding port table must exist 
which contains the port name and the longitude and latitude of each port. For 
lon/lat variables, longitude must be specified first.}

\item{alt_var}{Identifies how to find lat/lon for alternative choices. \code{alt_var} 
may be the centroid of zonal assignment \code{"centroid"}, or lon/lat variables 
in the primary dataset. Longitude must be specified first.}

\item{dist.unit}{String, how distance measure should be returned. Choices are 
\code{"meters"} or \code{"M"}, \code{"kilometers"} or \code{"KM"}, \code{"miles"}, 
or units of \code{grid}. Defaults to miles.}

\item{min.haul}{Required, numeric, minimum number of hauls. Zones with fewer 
hauls than the \code{min.haul} value will not be included in model data.}

\item{spat}{Required, data file or character. \code{spat} is a spatial data file 
containing information on fishery management or regulatory zones boundaries.
Shape, json, geojson, and csv formats are supported. geojson is the preferred 
format. json files must be converted into geojson. This is done automatically 
when the file is loaded with \code{\link{read_dat}} with \code{is.map = TRUE}.
\code{lon.dat}, \code{lat.dat}, \code{lon.spat}, and \code{lat.spat} are required 
for specific \code{spat} file formats. \code{cat} must be specified for all 
file formats. If a zonal centroid table exists in the FishSET database and a 
zonal assignment column exists in \code{dat} then \code{spat} may be a table 
from the FishSET database or a data file. If \code{spat} should come from the 
FishSET database, it should be the name of the original file name, in quotes. 
For example, 'NMFS_zones_polygons'. \code{cat} would then be the name of the 
column in \code{dat} containing fishing zone assignments. Use \code{tables_database()} 
to view names of tables in the FishSET database.}

\item{cat}{Required, variable in either \code{dat} or \code{spat} that identifies 
the individual areas or zones. If \code{cat} is a variable in \code{dat} that 
identifies zone assignments for each occurrence record, set \code{spat} to 
\code{NULL}. Otherwise, if \code{spat} is class \code{sf}, \code{cat} should be the 
name of the list containing information on zones.}

\item{zoneID}{Variable in \code{dat} that identifies the individual zones or 
areas. Define if exists in \code{dat} and is not named `ZoneID`. Defaults to 
\code{NULL}.}

\item{lon.dat}{Longitude variable from \code{dat}. Required if zonal assignments 
in \code{dat} should be identified and \code{spat} is not \code{NULL}.}

\item{lat.dat}{Latitude variable from \code{dat}. Required if zonal assignments 
in \code{dat} should be identified and \code{spat} is not \code{NULL}.}

\item{hull.polygon}{Used in \code{\link{assignment_column}} function. Creates 
polygon using convex hull method. Required if zonal assignments in \code{dat} 
should be identified and \code{spat} is not \code{NULL}.}

\item{closest.pt}{Logical, if true, zone ID identified as the closest polygon 
to the point. Called in \code{\link{assignment_column}}. Required if zonal 
assignments for observations in \code{dat} should be identified and \code{spat} 
is not \code{NULL}.}

\item{grid}{Data must contain a variable that varies by the spatial dataset 
\code{spat}. First variable in \code{grid} should match a column in \code{dat}. 
The remaining columns should match the zone IDs in the \code{spat}.}

\item{lon.spat}{Variable or list from \code{spat} containing longitude data. 
Required for csv files. Leave as \code{NULL} if \code{spat} is a shape or json 
file, Required if zonal assignments in \code{dat} should be identified and 
\code{spat} is not \code{NULL}.}

\item{lat.spat}{Variable or list from \code{spat} containing latitude data. Required 
for csv files. Leave as \code{NULL} if \code{spat} is a shape or json file, 
Required if zonal assignments in \code{dat} should be identified and \code{spat} 
is not \code{NULL}.}
}
\value{
Saves the alternative choice list to the FishSET database as a list.
  Output includes: \cr
\tabular{rlll}{
        dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that 
        zone is to be included in the model\cr
        greaterNZ: \tab Zones which pass numofNecessary test\cr
        numOfNecessary: \tab Minimum number of hauls for zone to be included\cr
        altChoiceUnits: \tab Set to miles\cr
        altChoiceType: \tab Set to distance\cr
        occasion: \tab Identifies how to find latitude and longitude for starting point\cr
        alt_var: \tab Identifies how to find latitude and longitude for alternative choice \cr
        zoneRow: \tab Zones and choices array\cr
        int: \tab Geographic centroid for each zone. Generated from \code{\link{find_centroid}}\cr
        matrix: \tab Distance matrix is alternative choices comes from gridded dataset
        }
}
\description{
Required step. Creates a list identifying how alternative fishing choices should 
be defined. Output is saved to the FishSET database. Run this function before 
running models.
}
\details{
Defines the alternative fishing choices. These choices are used to develop 
  the matrix of distances between observed and alternative fishing choices (where 
  they could have fished but did not). The distance matrix is calculated by the 
  \code{\link{make_model_design}} function. The distance matrix can come from 
  \code{dat} or from the gridded data frame \code{grid}. If the distance 
  matrix is to come from \code{dat}, then \code{occasion} (observed fishing location) 
  and \code{alt_var} (alternative fishing location) must be specified. \code{grid}, 
  if used, must be a variable that varies by the spatial dataset, such as wind 
  speed. Each column must be a unique zone that matches the zones in \code{dat}.
  
  Parts of the alternative choice list are pulled by \code{\link{create_expectations}}, 
  \code{\link{make_model_design}}, and the model run (\code{\link{discretefish_subroutine}}) 
  functions. These output include choices of which variable to use for catch and 
  which zones to include in analyses based on a minimum number of hauls per trip 
  within a zone. Note that if the alternative choice list is modified, the 
  \code{\link{create_expectations}} and \code{\link{make_model_design}} functions 
  should also be updated before rerunning models.
}
