% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternative_choice.R
\name{create_alternative_choice}
\alias{create_alternative_choice}
\title{Create alternative choice matrix}
\usage{
create_alternative_choice(dat, gridfile, case = c("Centroid", "Port",
  "Other"), contents, haul.trip = c("Haul", "Trip"), alt_var, occasion,
  lon.dat, lat.dat, lon.grid, lat.grid, cat, use.grid = c(TRUE, FALSE),
  hull.polygon = c(TRUE, FALSE), remove.na = FALSE,
  closest.pt = FALSE, project, griddedDat = NULL, weight.var = NULL)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{case}{Centroid='Centroid of Zonal Assignment', Port, Other}

\item{contents}{Value of variable to subset dataset by. For instance, include only zones with at least 10 hauls.}

\item{haul.trip}{Should data be at trip or haul level. Defaults to haul.}

\item{alt_var}{Identifies how to find lat/lon for starting point (must have a lat/lon associated with it)}

\item{occasion}{Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment'}

\item{lon.dat}{Longitude variable in dataset}

\item{lat.dat}{Latitude variable in dataset}

\item{lon.grid}{Longitude variable in gridfile}

\item{lat.grid}{Latitude variable in gridfile}

\item{cat}{Variable defining zones or areas. Must be defined for dataset or gridfile.}

\item{use.grid}{TRUE/FALSE. If TRUE, griddedDat is used to create centroids}

\item{hull.polygon}{Used in assignment_column function. Creates polygon using convex hull method.}

\item{remove.na}{TRUE/FALSE Remove points where zone ID not identified. Called in assignment_column function.}

\item{closest.pt}{TRUE/FALSE If true, zone ID identified as the closest polygon to the point. Called in assignment_column function.}

\item{project}{Name of project. Used for naming table saved to database}

\item{griddedDat}{Data frame containing a variable that varies by the map grid. First column should match a column in the dataset. The remaining columns should match the zone IDs in the gridfile.}

\item{weight.var}{variable weighted centroids}
}
\value{
Alternative choice matrix\cr
\tabular{rlll}{                                                                                                                                                                                                           
        dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that zone is to be included in the model\cr
        greaterNZ: \tab zone which pass numofNecessary test\cr
        numOfNecessary: \tab Minimum number of hauls necessary within a zone for zone to be included\cr
        altChoiceUnits: \tab set to miles\cr
        altChoiceType: \tab set to distance\cr
        alt_var: \tab Identifies how to find lat/lon for starting point\cr
        occasion: \tab Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment' \cr
        zoneRow: \tab zones and choices array\cr
        zoneType: \tab haul or trip\cr
        int: \tab Centroid for each zone. Generated from find_centroid function
        }
}
\description{
Creates a list containing information on alternative choices.
}
\details{
Functions returns alternative choice matrix. Function must be called before the create_expectations, make_model_design, or discretefish_subroutine functions can be called.
Function identifies which zones are to be included, based on number of hauls per trip, and assigns each observation to a zone.
}
