% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_dist_matrix}
\alias{create_dist_matrix}
\title{Create the distance matrix}
\usage{
create_dist_matrix(
  dataset,
  spat = NULL,
  spatID = NULL,
  alt_var,
  occasion,
  occasion_var = NULL,
  dataZoneTrue,
  zone_cent = NULL,
  fish_cent = NULL,
  choice,
  units,
  port = NULL,
  zoneRow,
  zoneID,
  crs = NULL
)
}
\arguments{
\item{dataset}{Primary data set}

\item{spat}{Spatial table.}

\item{spatID}{Column name for Zone/area ID.}

\item{alt_var}{Alternative choice location}

\item{occasion}{Define choice location}

\item{occasion_var}{Identify variable(s) needed to define choice location.}

\item{dataZoneTrue}{Include zone}

\item{zone_cent}{Zonal centroid table.}

\item{fish_cent}{Fishing centroid table.}

\item{choice}{Choice zone}

\item{units}{Distance units}

\item{port}{Port table}

\item{zoneRow}{Zone row}

\item{zoneID}{Zone identifier}

\item{crs}{Coordinate reference system.}
}
\value{
Distance matrix based on choices made in create_alternative_choice
}
\description{
Create the distance matrix
}
\details{
Function is called by \code{\link{make_model_design}} generate the 
  distance matrix. Alternative fishing options come from the Alternative Choice 
  list, generated from the \code{\link{create_alternative_choice}} function.
}
\keyword{internal}
