% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_dist_matrix}
\alias{create_dist_matrix}
\title{Create the distance matrix}
\usage{
create_dist_matrix(
  dataset,
  alt_var,
  occasion,
  dataZoneTrue,
  int,
  choice,
  units,
  port,
  zoneRow,
  X,
  zoneID
)
}
\arguments{
\item{dataset}{Primary data set}

\item{alt_var}{Alternative choice location}

\item{occasion}{Define choice location}

\item{dataZoneTrue}{Include zone}

\item{int}{Zone centroid data frame}

\item{choice}{Choice zone}

\item{units}{Distance units}

\item{port}{Port table}

\item{zoneRow}{Zone row}

\item{X}{distance matrix}

\item{zoneID}{Zone identifier}
}
\value{
Distance matrix based on choices made in create_alternative_choice
}
\description{
Create the distance matrix
}
\details{
Function is called by \code{\link{make_model_design}} generate the 
  distance matrix. Alternative fishing options come from the Alternative Choice 
  list, generated from the \code{\link{create_alternative_choice}} function.
}
\keyword{internal}
