% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_duration}
\alias{create_duration}
\title{Create duration of time variable}
\usage{
create_duration(dat, start, end, units = c("week", "day", "hour",
  "minute"), name = "create_duration")
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{start}{Variable indicating start of time period}

\item{end}{Variable indicating end of time period}

\item{units}{Unit of time for calculating duration. Must be weeks, days, hours, or minutes.}

\item{name}{Name of created vector. Used in the logging function to reproduce work flow. Defaults to name of the function if not defined.}
}
\description{
Create duration of time variable
}
\details{
Calculates the duration of time between two temporal variables based on defined unit. The new variable is added to the dataset.
}
\examples{
\dontrun{
MainDataTable$TripDur <- create_duration(MainDataTable, 'TRIP_START', 'TRIP_END',  units='minute')
}
}
