% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expectations.R
\name{create_expectations}
\alias{create_expectations}
\title{Expected catch}
\usage{
create_expectations(dat, project, gridfile, catch, defineGroup = NULL,
  temp.var = NULL, temporal = c("daily", "sequential"),
  calc.method = c("standardAverage", "simpleLag", "weights"),
  lag.method = c("simple", "grouped"), empty.catch = c(NULL, 0,
  "allCatch", "groupedCatch"), empty.expectation = c(NULL, 1e-04, 0),
  temp.window = 1, temp.lag = 1, dummy.exp = FALSE)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{project}{Name of project. Used to pull working alternative choice matrix from fishset_db database.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{catch}{Variable containing catch data.}

\item{defineGroup}{If empty, data is treated as a fleet}

\item{temp.var}{Variable containing temporal data}

\item{temporal}{Daily (Daily time line) or sequential (sequential order)}

\item{calc.method}{Select standard average (standardAverage), simple lag regression of means (simpleLag), or weights of regressed groups (weights)}

\item{lag.method}{Use regression over entire group (simple) or for grouped time periods (grouped)}

\item{empty.catch}{Replace empty catch with NA, 0, mean of all catch (allCatch), or mean of grouped catch(groupCatch)}

\item{empty.expectation}{Do not replace (NULL) or replace with 0.0001 or 0}

\item{temp.window}{Temporal window size. Defaults to 1}

\item{temp.lag}{Temporal lag time in days.}

\item{dummy.exp}{T/F. Defaults to False. If false, no dummy variable is outputted. If true, output dummy variable for originally missing value.}
}
\value{
newGridVar dataframe. Saved to the global environment. Dataframe called in make_model_design

newGridVar,  newDumV
}
\description{
Expected catch
}
\details{
Used during model creation to create an expectation of catch for alternative choices that are added to the model design file.
The expectations created have several options and are created based on the group and time averaging choices of the user.
The spatial alternatives are built in to the function and come from the structure Alt. 
The primary choices are whether to treat data as a fleet or to group the data (defineGroup) and the time frame of catch data for calculating expected catch.
Values must be provided for `defineGroup` and `temporal` parameters. If data should be treated as a fleet, set defineGroup to NULL. If the entire record of catch data is to be used, set temporal to NULL.
Empty catch values are considered to be times of no fishing activity whereas values of 0 in the catch variable 
are considered fishing activity with no catch and so those are included in the averaging and dummy creation as a point in time when fishing occurred.
}
\examples{
\dontrun{
create_expectations(MainDataTable, adfg, 'OFFICIAL_TOTAL_CATCH_MT',  temporal='daily', 
                     temp.var="DATE_FISHING_BEGAN", calc.method='standard average', 
                     lag.method='simple',  empty.catch='all catch', empty.expectation= 0.0001, 
                     temp.window=4, temp.lag=2, dummy.exp=FALSE, 
                     AltMatrixName='pcodaltmatrix20110101', defineGroup=NULL)
}
}
