% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expectations.R
\name{create_expectations}
\alias{create_expectations}
\title{Expected catch}
\usage{
create_expectations(dataset, gridfile, catch, temporal = c("daily",
  "sequential"), temp.var = NULL, calc.method = c("standard average",
  "simple lag", "weights"), lag.method = c("simple", "grouped"),
  empty.catch = c(NULL, 0, "all catch", "grouped catch"),
  empty.expectation = c(NULL, 1e-04, 0), temp.window = 1,
  temp.lag = 1, dummy.exp = FALSE, AltMatrixName = NULL,
  defineGroup = NULL)
}
\arguments{
\item{dataset}{Main data frame containing data on hauls or trips}

\item{gridfile}{Name of file containing spatial data. Shape, json, and csv formats are supported.}

\item{catch}{Catch variable for averaging}

\item{temporal}{Daily = Daily time line or sequential = sequential order}

\item{temp.var}{Temporal variable for averaging}

\item{calc.method}{select standard average, simple lag = simple lag regression of means, or weights= weights of regressed groups}

\item{lag.method}{Simple: use region over entire group. Grouped: Use regression for individual time periods}

\item{empty.catch}{Replace empty catch with NAN, 0, all catch = mean of all catch, group catch = mean of grouped catch}

\item{empty.expectation}{Do not replace or replace with 0.0001 or 0}

\item{temp.window}{Window size for averaging. Defaults to 1}

\item{temp.lag}{Lag time for averaging}

\item{dummy.exp}{T/F. Defaults to False. If false, no dummy variable is outputted. If true, output dummy variable for originally missing value.}

\item{AltMatrixName}{Does not need to specified if ALT has been generated in createAlternativeChoice function}

\item{defineGroup}{If empty, data is treated as a fleet}
}
\value{
newGridVar dataframe. Saved to the global environment. Dataframe called in make_model_design

newGridVar,  newDumV
}
\description{
Expected catch
}
\details{
Used during model creation to create an expectation of catch for alternative choices that are added to the model design file.
The expectations created have several options and are created based on the group and time averaging choices of the user.
The spatial alternatives are built in to the function and come from the structure Alt.
NOTE: currently empty values and values ==nan are considered to be times of no fishing activity whereas values in the catch variable choosen ==0
are considered fishing activity with no catch and so those are included in the averaging and dummy creation as a point in time when fishing occurred.
}
