% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expectations.R
\name{create_expectations}
\alias{create_expectations}
\title{Expected catch/Expected Revenue}
\usage{
create_expectations(
  dat,
  project,
  catch,
  price = NULL,
  defineGroup,
  temp.var = NULL,
  temporal = c("daily", "sequential"),
  calc.method = c("standardAverage", "simpleLag", "weights"),
  lag.method = c("simple", "grouped"),
  empty.catch = c(NULL, 0, "allCatch", "groupedCatch"),
  empty.expectation = c(NULL, 1e-04, 0),
  temp.window = 7,
  temp.lag = 0,
  year.lag = 0,
  dummy.exp = FALSE,
  replace.output = FALSE
)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{project}{Name of project. Used to pull working alternative choice matrix from fishset_db database.}

\item{catch}{Variable containing catch data.}

\item{price}{Variable containing price/value data. Used in calculating expected revenue. Leave null if calculating expected catch. Multiplied against catch to generated revenue.}

\item{defineGroup}{If empty, data is treated as a fleet}

\item{temp.var}{Variable containing temporal data}

\item{temporal}{Daily (Daily time line) or sequential (sequential order)}

\item{calc.method}{Select standard average (standardAverage), simple lag regression of means (simpleLag), or weights of regressed groups (weights)}

\item{lag.method}{Use regression over entire group (simple) or for grouped time periods (grouped)}

\item{empty.catch}{Replace empty catch with NA, 0, mean of all catch (allCatch), or mean of grouped catch(groupCatch)}

\item{empty.expectation}{Do not replace (NULL) or replace with 0.0001 or 0}

\item{temp.window}{Temporal window size. In days. Defaults to 14 (14 days)}

\item{temp.lag}{Temporal lag time in days.}

\item{year.lag}{IF expected catch should be based on catch from previous year(s), set year.lag to the number of years to go back.}

\item{dummy.exp}{T/F. Defaults to False. If false, no dummy variable is outputted. If true, output dummy variable for originally missing value.}

\item{replace.output}{Default is FALSE. If TRUE, replaces existing saved expected catch dataframe with new expected catch dataframe. 
If FALSE, add new expected catch dataframes to previously saved expected catch dataframes.}
}
\value{
newGridVar dataframe. Saved to the global environment. Dataframe called in make_model_design

newGridVar,  newDumV
}
\description{
Expected catch/Expected Revenue
}
\details{
Used during model creation to create an expectation of catch or revenue for alternative choices that are added to the model design file.
IMPORTANT: Use the temp_obs_table function before using this function to assess the availability of data for desired temporal moving window size. Sparse data is not suited for shorter moving window sizes. 
The expectations created have several options and are created based on the group and time averaging choices of the user.
The spatial alternatives are built in to the function and come from the structure Alt. 
The primary choices are whether to treat data as a fleet or to group the data (defineGroup) and the time frame of catch data for calculating expected catch.
Values must be provided for `defineGroup` and `temporal` parameters. If data should be treated as a fleet, set defineGroup to NULL. If the entire record of catch data is to be used, set temporal to NULL.
Empty catch values are considered to be times of no fishing activity whereas values of 0 in the catch variable 
are considered fishing activity with no catch and so those are included in the averaging and dummy creation as a point in time when fishing occurred.
Three default expected catch cases will be run:
\itemize{
\item{Near-term: Movng window size of two days. In this case, vessels are grouped based on defineGroup parameter.}
\item{Medium-term: Moving window size of seven days. In this case, there is no grouping and catch for entire fleet is used.}
\item{Long-term: Moving window size of seven days from the previous year. In this case, there is no grouping and catch for entire fleet is used.}
}
}
\examples{
\dontrun{
create_expectations(MainDataTable, adfg, 'OFFICIAL_TOTAL_CATCH_MT',  price=NULL, temporal='daily', 
                     temp.var="DATE_FISHING_BEGAN", calc.method='standard average', 
                     lag.method='simple',  empty.catch='all catch', empty.expectation= 0.0001, 
                     temp.window=4, temp.lag=2, dummy.exp=FALSE, 
                     AltMatrixName='pcodaltmatrix20110101', defineGroup=NULL)
}
}
