% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLogitInput.R
\name{create_logit_input}
\alias{create_logit_input}
\title{create_logit_input}
\usage{
create_logit_input(choice)
}
\arguments{
\item{choice}{A (number of observations) x 1 vector of chosen locations}
}
\value{
dataCompile - a data matrix
}
\description{
Creates a data matrix that is consistent with the built in model forms
}
\details{
Called in zonalSubroutine
}
\examples{

}
