% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logit_input.R
\name{create_logit_input}
\alias{create_logit_input}
\title{Creates a data matrix consistent with built-in model forms}
\usage{
create_logit_input(choice)
}
\arguments{
\item{choice}{A (number of observations) x 1 vector of chosen locations}
}
\value{
dataCompile - a data matrix
}
\description{
Creates a data matrix consistent with built-in model forms
}
\details{
Called in the discrete_fish_subroutine function
}
\keyword{internal}
