% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_mid_haul}
\alias{create_mid_haul}
\title{Calculate latitude and longitude of haul midpoint}
\usage{
create_mid_haul(dat, start = c("lon", "lat"), end = c("lon", "lat"))
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{start}{Starting location of haul. Must be specificied as vector containing longitudinal data and vector containing latitudinal data separated by comma.}

\item{end}{Ending location of haul. Must be specificied as vector containing longitudinal data and vector containing latitudinal data separated by comma.}
}
\description{
Calculate latitude and longitude of haul midpoint
}
\details{
Returns midpoint of each haul. Each row of data must be a unique haul. Requires a start and end point for each observations.
}
\examples{
\dontrun{
MainDataTable$haulMidPoint <- create_mid_haul(MainDataTable, start = c("LonLat_START_LON", 
                         "LonLat_START_LAT"), end = c("LonLat_END_LON", "LonLat_END_LAT")) 
}
}
