% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasonal_ID.R
\name{create_seasonal_ID}
\alias{create_seasonal_ID}
\title{create_seasonal_ID
Connects the variables in a table of season dates to main data table based on two column to the data in the main workspace.
A new new variable is made with the Season identifier (TRUE/FALSE) in it.}
\usage{
create_seasonal_ID(dataset, seasonaldat, use.location = c(TRUE, FALSE),
  use.geartype = c(TRUE, FALSE), sp.col, target = NULL)
}
\arguments{
\item{dataset}{Name of table in database containing model measures of fit}

\item{seasonaldat}{Append model selection to modelChosen table or delete existing table and save new table}

\item{use.location}{TRUE/FALSE If true, season dates depend on fishery location. Colnames of location must match the two datasets}

\item{use.geartype}{TRUE/FALSE If true, season dates depend on gear type. Column names must match the two datasets}

\item{sp.col}{Column name containing species names in seasonaldat}

\item{target}{Name of target species. If not null, will return a vector with name of fishery season for each row of dataset. Will first check the dates of the fishery season for target species before checking other species.}
}
\value{
dataset Returns the input dataset with two new variables
}
\description{
create_seasonal_ID
Connects the variables in a table of season dates to main data table based on two column to the data in the main workspace.
A new new variable is made with the Season identifier (TRUE/FALSE) in it.
}
