% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_startingloc.R
\name{create_startingloc}
\alias{create_startingloc}
\title{Create starting location variable}
\usage{
create_startingloc(
  dat,
  project = NULL,
  spat,
  port,
  port_name,
  port_lon,
  port_lat,
  trip_id,
  haul_order,
  starting_port,
  zoneID,
  spatID,
  name = "startingloc"
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Name of project}

\item{spat}{Spatial data. Required if \emph{ZoneID} does not exists in \code{dat}.
Shape, json, geojson, and csv formats are supported.}

\item{port}{Port data. Contains columns: Port_Name, Port_Long, Port_Lat. 
Table is generated using the \code{\link{load_port}} and saved in the FishSET 
database as the project and port table, for example 'pollockPortTable'.}

\item{port_name}{Character string indicating the column in port table that contains the port name}

\item{port_lon}{Character string indication the column in port table that contains port longitude}

\item{port_lat}{Character string indication the column in port table that contains port latitude}

\item{trip_id}{Variable in \code{dat} that identifies unique trips.}

\item{haul_order}{Variable in \code{dat} containing information on the order 
that hauls occur within a trip. Can be time, coded variable, etc.}

\item{starting_port}{Variable in \code{dat} to identify port at start of trip.}

\item{zoneID}{Variable in \code{dat} that identifies the individual zones or 
areas.}

\item{spatID}{Variable in \code{spat} that identifies the individual zones or 
areas.}

\item{name}{String, name of created variable. Defaults to name of the function 
if not defined.}
}
\value{
Primary data set with starting location variable added.
}
\description{
Creates a variable containing the zone/area location of a vessel when choice of 
where to fish next was made. This variable is required for data with multiple sets
or hauls in a single trip and for the full information model with 
Dahl's correction (\code{\link{logit_correction}}).
}
\details{
Function creates the \code{startloc} vector that is required for the 
  full information model with Dahl's correction \code{\link{logit_correction}}. 
  The vector is the zone location of a vessel when the decision of where to fish 
  next was made. Generally, the first zone of a trip is the departure port. 
  The \code{\link{assignment_column}} function is called to assign starting port 
  locations and haul locations to zones. If ZoneID exists in \code{dat}, 
  \code{\link{assignment_column}} is not called and the following arguments are 
  not required: \code{spat, lon.dat, lat.dat, cat, lon.grid, lat.grid}.
}
\examples{
\dontrun{
pcodMainDataTable <- create_startingloc(pcodMainDataTable, 'pcod',
    map2, "pcodPortTable", "TRIP_SEQ", "HAUL_SEQ", "DISEMBARKED_PORT", 
 "START_LON", "START_LAT", "NMFS_AREA", "STARTING_LOC"
)
}
}
