% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_startingloc.R
\name{create_startingloc}
\alias{create_startingloc}
\title{Create Starting location vector}
\usage{
create_startingloc(dat, gridfile, PortTable, trip_id, haul_order,
  starting_port, lon.dat, lat.dat, lon.grid, lat.grid, cat)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{PortTable}{Port data frame. Contains columns: Port_Name, Port_Long, Port_Lat. Table is generated using the load_port function and saved in the fishset_db database.}

\item{trip_id}{Variable in 'dat' to identify unique trips.}

\item{haul_order}{Variable in `dat` containing information on the order that hauls occur within a trip. Can be time, coded variable, etc.}

\item{starting_port}{Variable in `dat` to identify port at start of trip}

\item{lon.dat}{Longitude variable in dataset}

\item{lat.dat}{Latitude variable in dataset}

\item{lon.grid}{Longitude variable in gridfile}

\item{lat.grid}{Latitude variable in gridfile}

\item{cat}{Variable defining zones or areas. Must be defined for dataset or gridfile.}
}
\description{
Creates a vector containing zone/area location when choice of where to fish next was chosen.
}
\details{
Function creates the startloc vector that is needed for the logit_correction function. The assignment_column function is called to assign 
port locations and haul locations to zones. The starting port is used to define the starting location at the start of the trip.
}
\examples{
\dontrun{
MainDataTable$startloc <- create_startingloc(MainDataTable, map2, pollockPortTable, 
                              'TRIP_SEQ','HAUL_SEQ','DISEMBARKED_PORT',
                              "LonLat_START_LON","LonLat_START_LAT", "","", 'NMFS_AREA')
}
}
