% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_trip_centroid}
\alias{create_trip_centroid}
\title{Create trip centroid variable}
\usage{
create_trip_centroid(dat, project, lon, lat, tripID, weight.var = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{lon}{Variable in \code{dat} containing longitudinal data.}

\item{lat}{Variable in \code{dat} containing latitudinal data.}

\item{tripID}{Variable in \code{dat} containing trip identifier. If trip identifier should be defined by more than one variable then list as \code{c('var1', 'var2')}.}

\item{weight.var}{Variable in \code{dat} for computing the weighted average.}
}
\value{
Returns the primary dataset with centroid latitude and centroid longitude variables added.
}
\description{
Create latitude and longitude variables containing the centroid of each trip
}
\details{
Computes the average longitude and latitude for each trip. Specify \code{weight.var} to calculate the weighted centroid.
  Additional arguments can be added that define unique trips. If no additional arguments are added, each row will be treated as a unique trip.
}
\examples{
\dontrun{
pollockMainDataTable <- create_trip_centroid(pollockMainDataTable, 'pollock', 'LonLat_START_LON', 
  'LonLat_START_LAT', weight.var = NULL, 'DISEMBARKED_PORT', 'EMBARKED_PORT')
}
}
