% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_distance.R
\name{create_trip_distance}
\alias{create_trip_distance}
\title{Generate haul level trip distance}
\usage{
create_trip_distance(dataset, PortTable, trip_id, starting_port,
  starting_haul = c("Lon", "Lat"), ending_haul = c("Lon", "Lat"),
  ending_port, haul_order, a = 6378137, f = 1/298.257223563)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{PortTable}{dataframe with columns Port_Name, Port_Long, Port_Lat}

\item{trip_id}{Vector defining unique trips}

\item{starting_port}{Vector defining port at start of trip}

\item{starting_haul}{Vector defining Lat/Long of start of haul. Should contain two vectors.}

\item{ending_haul}{Vector defining Lat/Long of end of haul. Should contain two vectors.}

\item{ending_port}{Vector defining port at end of trip}

\item{haul_order}{Vector defining order that hauls occur}

\item{a}{numeric. Major (equatorial) radius of the ellipsoid. The default value is for WGS84}

\item{f}{numeric. Ellipsoid flattening. The default value is for WGS84}
}
\value{
vector of trip distance
}
\description{
Generate haul level trip distance
}
\details{
Summation of distance across a trip based on choices by the user for starting and ending ports, and hauls in between.
 Uses longitude and latitude to calculate the distance between and returns as a new variable. Relies on the distGeo function to calculate distances
 between hauls. Inputs are a dataframe or matrix of trips, ports, etc and a datatable on ports including latitude and longitude of ports.
}
