% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_var_num}
\alias{create_var_num}
\title{Create numeric variables}
\usage{
create_var_num(dataset, x, y, method, name)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{Variable to create}

\item{y}{Second variable Note that in division  x is divided by y.}

\item{method}{Addition, substraction, multiplication, division}

\item{name}{Name of new variable}
}
\description{
Create numeric variables
}
\details{
Function for generating new or specialized variables. create_var_num create a new numeric variable based on defined arithmetic function. New variable is added to the dataset.
}
