% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{create_var_temp}
\alias{create_var_temp}
\title{Create temporal variables}
\usage{
create_var_temp(dataset, start, end, name, units = c("week", "day",
  "hour", "minute"))
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{start}{Variable indicating start of time period}

\item{end}{Variable indicating end of time period}

\item{name}{Name of new variable}

\item{units}{Units of time varibles. Must be weeks, days, hours, or minutes}
}
\description{
Create temporal variables
}
\details{
Function for generating new or specialized variables. create_var_temp calculates the duration of time between two temporal variables based on defined time format. The new variable is added to the dataset.
}
