% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{dat_to_sf}
\alias{dat_to_sf}
\title{Convert dataframe to sf}
\usage{
dat_to_sf(spatdat, lon, lat, id)
}
\arguments{
\item{spatdat}{Spatial data containing information on fishery management or 
regulatory zones.}

\item{lon}{Longitude variable in \code{spatdat}.}

\item{lat}{Latitude variable in \code{spatdat}.}

\item{id}{Polygon ID column.}
}
\description{
Used to convert spatial data with no spatial class to a \code{sf} object. 
This is useful if the spatial data was read from a non-spatial file type, 
e.g. a CSV file.
}
