% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Guided steps to cleaning the datafile}
\usage{
data_check(dataset, x, dataindex)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{column in dataframe to check for outliers}

\item{dataindex}{dataframe that contains information on each column of the dataset. Must be in quotes.}
}
\description{
Guided steps to cleaning the datafile
}
\details{
Prints summary stats for all variable in the dataframe. Checks for NaNs and prints vectors that contain NaNs.
Checks for outliers for specified vector. Further actions may be taken to further evaluate and remove NaNs and outliers.
Checks that all columnn names in the dataset are unique, whether specialized variables have been identified in the index dataset, whether any columns in the dataset are empty, if units are defined and recognized, whether each row is a unique choice occurrence at the haul or trip level, and that data for either lat/long or fishing area are included.
}
