% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{data_verification_call}
\alias{data_verification_call}
\title{Contains one function that tests several if statements. The function stops if an if statement does not pass.}
\usage{
data_verification_call(dataset)
}
\arguments{
\item{dataset}{dataframe or matrix}
}
\value{
Returns statements as to whether issues in the data may exist
}
\description{
Contains one function that tests several if statements. The function stops if an if statement does not pass.
}
\details{
checks that all columnn names in the dataset are unique, whether any columns in the dataset are empty, whether each row is a unique choice 
occurrence at the haul or trip level, and that data for either lat/long or fishing area are included.
}
