% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_cols}
\alias{date_cols}
\title{Find columns that can be converted to Date or Date-time class}
\usage{
date_cols(dat, out = "names", type = "both")
}
\arguments{
\item{dat}{MainDataTable or dataframe to check.}

\item{out}{Whether to return the column \code{"names"} (the default) or a logical vector 
(\code{"logical"}).}

\item{type}{String, the type of date column to test for. Options are 
\code{"date"}, \code{"date_time"}, or \code{"both"}.}
}
\description{
Find columns that can be converted to Date or Date-time class
}
\examples{
\dontrun{
date_cols(pollockMainDataTable) # returns column names
date_cols(pollockMainDataTable, "logical")
}
}
\keyword{internal}
