% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degree}
\alias{degree}
\title{Convert lat/long coordinates to decimate degrees}
\usage{
degree(dat, lat, long)
}
\arguments{
\item{dat}{Data table containing latitude and longitude data}

\item{lat}{Name of vector containing latitude data}

\item{long}{Name of vector containg longitude data}
}
\value{
The original dataframe with the converted latitudes and longitudes
}
\description{
Convert lat/long coordinates to decimate degrees
}
\details{
Uses the degree function to convert lat long coordinates to decimal degrees.
}
\examples{
\dontrun{
dat <- degree(MainDataTable, 'LatLon_START_LAT', 'LatLon_START_LON')}

}
