% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Check and convert lat/lon to decimal degrees}
\usage{
degree(
  dat,
  project,
  lat = NULL,
  lon = NULL,
  latsign = FALSE,
  lonsign = FALSE,
  replace = TRUE
)
}
\arguments{
\item{dat}{Dataset containing latitude and longitude data.}

\item{project}{Project name.}

\item{lat}{Variable(s) containing latitude data. If \code{NULL} the function
will attempt to search for all latitude variables by name (e.g. by matching
"lat" or "LAT").}

\item{lon}{Variable(s) containing longitude data. If \code{NULL} the function
will attempt to search for all longitude variables by name (e.g. by matching
"lon" or "LON").}

\item{latsign}{How should the sign value of \code{lat} be changed? Choices 
are \code{NULL} for no change, \code{"neg"} to convert all positive values to 
negative, \code{"pos"} to convert all negative values to positive, and
\code{"all"} to change all values.}

\item{lonsign}{How should the sign value of \code{lon} be changed? Choices 
are \code{NULL} for no change, \code{"neg"} to convert all positive values to 
negative, \code{"pos"} to convert all negative values to positive, and
\code{"all"} to change all values.}

\item{replace}{Logical, should \code{lat} and \code{lon} in \code{dat} be 
converted to decimal degrees? Defaults to \code{TRUE}. Set to \code{FALSE} 
if checking for compliance.}
}
\value{
Returns the primary dataset with the latitudes and longitudes converted 
  to decimal degrees if \code{replace = TRUE} or if Changing the sign. 
  Otherwise, a message indicating whether selected longitude and latitude
  variables are in the correct format.
}
\description{
Check that latitude and longitude are in decimal degrees and the variable 
sign is correct. Correct lat/lon if required.
}
\details{
First checks whether any variables containing 'lat' or 'lon' in their 
  names are numeric. Returns a message on results. To convert a variable to 
  decimal degrees, identify the \code{lat} or \code{lon} variable(s) and set 
  \code{replace = TRUE}. To change the sign, set \code{latsign} (for \code{lat}) 
  or \code{lonsign} (for \code{lon = TRUE}. FishSET requires that latitude 
  and longitude be in decimal degrees.
}
\examples{
\dontrun{
# check format
degree(pollockMainDataTable, 'pollock', lat = 'LatLon_START_LAT',
       lon = 'LatLon_START_LON')

# change signs and convert to decimal degrees
pollockMainDataTable <- degree(pollockMainDataTable, 'pollock', 
                               lat = 'LatLon_START_LAT', 
                               lon = 'LatLon_START_LON', latsign = FALSE, 
                               lonsign = FALSE, replace = TRUE)
}


}
