% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretefish_subroutine.R
\name{discretefish_subroutine}
\alias{discretefish_subroutine}
\title{discretefish_subroutine
Subroutine to run chosen discrete choice model}
\usage{
discretefish_subroutine(x, initparams, optimOpt, func, methodname,
  func.name, select.model = FALSE, project)
}
\arguments{
\item{x}{Table containing catch, choice, distance, and otherdat. Generated by make_model_design function. Should have 'modelInputData' in the table name.}

\item{initparams}{Initial parameter estimates for revenue/location-specific covariates then cost/distance}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x, trace]}

\item{func}{Name of likelihood function}

\item{methodname}{Optimization method (see optim options)}

\item{func.name}{Name of likelihood function for model result output table}

\item{select.model}{Return an interactive data table that allows users to select and save table of best models based on measures of fit}

\item{project}{Name of project for naming output table in sql database}
}
\value{
OutLogit - [outmat1 se1 tEPM2] (coefs, ses, tstats) \cr 
optoutput - optimization information \cr 
seoumat2 - ses \cr 
MCM - Model Comparison metrics \cr
}
\description{
discretefish_subroutine
Subroutine to run chosen discrete choice model
}
