% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretefish_subroutine.R
\name{discretefish_subroutine}
\alias{discretefish_subroutine}
\title{discretefish_subroutine
Subroutine to run chosen discrete choice model}
\usage{
discretefish_subroutine(catch, choice, distance, otherdat, initparams,
  optimOpt, func, methodname, func.name, select.model = FALSE)
}
\arguments{
\item{catch}{Data corresponding to actual zonal choice}

\item{choice}{Data corresponding to actual catch}

\item{distance}{Data corresponding to distance}

\item{otherdat}{Other data (as a list)}

\item{initparams}{Initial parameter estimates for revenue/location-specific covariates then cost/distance}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x]}

\item{func}{Name of likelihood function}

\item{methodname}{Optimization method (see optim options)}

\item{func.name}{Name of likelihood function for model result output table}

\item{select.model}{Return an interactive data table that allows users to select and save table of best models based on measures of fit}
}
\value{
OutLogit - [outmat1 se1 tEPM2] (coefs, ses, tstats) \cr 
optoutput - optimization information \cr 
seoumat2 - ses \cr 
MCM - Model Comparison metrics \cr
}
\description{
discretefish_subroutine
Subroutine to run chosen discrete choice model
}
