% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretefish_subroutine.R
\name{discretefish_subroutine}
\alias{discretefish_subroutine}
\title{Subroutine to run chosen discrete choice model}
\usage{
discretefish_subroutine(project, initparams, optimOpt, func, methodname,
  mod.name, select.model = FALSE, name = "discretefish_subroutine")
}
\arguments{
\item{project}{Name of project. For obtaining catch, choice, distance, and otherdat data generated from make_model_design function. 
Working modelInputData table (table without date) will be putlled from fishset_db database.}

\item{initparams}{Initial parameter estimates for revenue/location-specific covariates then cost/distance}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x, trace]}

\item{func}{Name of likelihood function}

\item{methodname}{Optimization method (see optim options)}

\item{mod.name}{Name of model run for model result output table}

\item{select.model}{Return an interactive data table that allows users to select and save table of best models based on measures of fit}

\item{name}{Name of created vector. Used in the logging function to reproduce work flow. Defaults to name of the function if not defined.}
}
\value{
\tabular{rlll}{
OutLogit: \tab [outmat1 se1 tEPM2] (coefs, ses, tstats) \cr 
optoutput: \tab optimization information \cr 
seoumat2: \tab ses \cr 
MCM: \tab Model Comparison metrics \cr 
}
}
\description{
Subroutine to run chosen discrete choice model
}
\examples{
\dontrun{
results <- discretefish_subroutine(modelInputData, initparams= c(0.5, -2.8), 
                                   optimOpt=c(100000,1.00000000000000e-08,1,1),
                                   func=logit_c, methodname="BFGS", 'newlogit4', 
                                   select.model=TRUE, project='projectName')
}
}
