% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_matrix}
\alias{dummy_matrix}
\title{Create dummy matrix from a coded ID variable}
\usage{
dummy_matrix(dataset, x)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{Variable to create}
}
\description{
Create dummy matrix from a coded ID variable
}
\details{
Function for generating new or specialized variables. dummy_matrix creates a dummy matrix. setQuants creates a coded variable based on the quantiles of x.
}
