% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_var}
\alias{dummy_var}
\title{Create dummy variable}
\usage{
dummy_var(dat, project, DumFill = 1, name = "dummy_var")
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{DumFill}{Fill the dummy variable with 1 or 0}

\item{name}{String, name of created dummy variable. Defaults to name of the function if not defined.}
}
\value{
Primary dataset with dummy variable added.
}
\description{
Create dummy variable
}
\details{
Creates a dummy variable of either 0 or 1 with length of the number of rows of the data set.
}
\examples{
\dontrun{
pollockMainDataTable <- dummy_var(pollockMainDataTable, 'pollock', DumFill=1, 'dummyvar')
}
}
