% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{dummy_var}
\alias{dummy_var}
\title{Create new dummy variable}
\usage{
dummy_var(dat, DumFill = "TRUE", name = "dummy_var")
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{DumFill}{Fill the dummy variable with TRUE or FALSE}

\item{name}{Name of created dummy variable. Used in the logging function to reproduce work flow. Defaults to name of the function if not defined.}
}
\description{
Create new dummy variable
}
\details{
Creates a dummy variable of either FALSE or TRUE with length of the number of rows of the data set.
}
\examples{
\dontrun{
MainDataTable$dummyvar <- dummy_var(MainDataTable, DumFill=TRUE)
}
}
