% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{empty_vars_filter}
\alias{empty_vars_filter}
\title{Check variables are not empty}
\usage{
empty_vars_filter(dat, project, remove = FALSE)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{remove}{Logical, whether to remove empty variables. Defaults to FALSE.}
}
\value{
Returns the dataset with empty variables removed if \code{remove} is TRUE.
}
\description{
Check for and remove empty variables.
}
\details{
Function checks for empty variables and prints an outcome message to the console. 
  If empty variables are present and \code{remove} = TRUE, then empty variables will be removed from the dataset.
}
\examples{
\dontrun{
empty_vars_filter(pollockMainDataTable)
mod.dat <- empty_vars_filter(pollockMainDataTable, 'pollock', remove = TRUE)
}
}
