% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_lognormal.R
\name{epm_lognormal}
\alias{epm_lognormal}
\title{epm_lognormal}
\usage{
epm_lognormal(starts3, dat, otherdat, alts)
}
\arguments{
\item{starts3}{Starting values. e.g. c([grid-varying variables], [interaction variables], [catch variance], [scale parameter]).}

\item{dat}{Data matrix, see output from shift_sort_x, alternatives with distance by column bind}

\item{otherdat}{Other data used in model (as list). Any number of grid-varying variables (e.g. expected catch that varies by location) or 
interaction variables (e.g. vessel characteristics that affect how much disutility is suffered by traveling a greater distance) are allowed. \cr \cr
However, the user must place these in `otherdat` as list objects named `griddat` and `intdat` respectively. Note the variables #' within `griddat` 
and `intdat` have no naming restrictions. \cr \cr
Also note that `griddat` variables are  dimension *(number of observations) x #' (number of alternatives)*, while `intdat` variables are 
dimension *(number of observations) x 1*, to be interacted with the distance to each #' alternative. \cr \cr
If there are no other data, the user can set `griddat` as ones with dimension *(number of observations) x (number of alternatives)*
and `intdat` variables as ones with dimension *(number of observations) x 1*.}

\item{alts}{Number of alternative choices in model}
}
\value{
ld - negative log likelihood
}
\description{
Expected profit model lognormal catch function
}
