% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epm_predict.R
\name{epm_predict}
\alias{epm_predict}
\title{EPM predict}
\usage{
epm_predict(
  project,
  mod.name,
  mod.type,
  use.scalers = FALSE,
  scaler.func = NULL,
  outsample = FALSE,
  outsample.mod.name = NULL
)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Name of saved model to use}

\item{mod.type}{String. Options are weibull, lognormal, normal}

\item{use.scalers}{Input for \code{\link{create_model_input}}. Logical, should data be normalized? Defaults to \code{FALSE}. Rescaling factors are the mean of the 
numeric vector unless specified with \code{scaler.func}.}

\item{scaler.func}{Input for \code{\link{create_model_input}}. Function to calculate rescaling factors.}

\item{outsample}{Logical, \code{FALSE} if predicting probabilities for main data, and \code{TRUE} if predicting for out-of-sample data. \code{outsample = FALSE} 
is the default setting.}
}
\value{
Returns probability of epm model by choice
}
\description{
Prediction component for expected profit models (EPMs). This function is called in 
\code{\link{model_prediction}}, which is called within \code{\link{run_policy}}.
}
\keyword{internal}
