% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_dat}
\alias{filter_dat}
\title{Remove rows based on user-defined filter}
\usage{
filter_dat(dataset, exp, use.filter.table = F)
}
\arguments{
\item{dataset}{dataframe or matrix over which to apply filter}

\item{exp}{Filter expression. Use row row number if use.filter.table==TRUE. Otherwise, should take on the form of 'x<100' or is.na(x)==F.}

\item{use.filter.table}{TRUE or FALSE. If true then data is subsetted based on a filter in filterTable and exp is the row containing the filter.}
}
\value{
filter_dat applies user-definted filters. Output must be saved
}
\description{
Remove rows based on user-defined filter
}
\details{
This function allows users to define and store data filters which can then be applied to the data. The filter dataframe can be saved and will be logged in the log file.
}
\keyword{filter,}
\keyword{subset}
