% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_dat}
\alias{filter_dat}
\title{Remove rows based on user-defined filter expressions}
\usage{
filter_dat(dat, exp, filterTable)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishet_db database should contain the string `MainDataTable`.}

\item{exp}{Filter expression. Use row number if use.filter.table==TRUE. Otherwise, should take on the form of `x<100` or `is.na(x)==F`.}

\item{filterTable}{Name of filter table in fishset_db database. Name should contain the phrase 'filterTable'.}
}
\value{
Filtered data frame
}
\description{
Remove rows based on user-defined filter expressions
}
\details{
Filter data frame based on pre-defined filter expression from filterTable or user-defined filter expression.
}
\examples{
\dontrun{  
newdat <- filter_dat(MainDataTable, exp=3, filterTable='pcodfilterTable01012011')
newdat <- filter_dat(MainDataTable, exp='PERFORMANCE_Code==1', filteTable='')
}

}
\keyword{filter,}
\keyword{subset}
