% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_table}
\alias{filter_table}
\title{Define and store filters}
\usage{
filter_table(dataset, x, exp, save.filter = FALSE, flog.dat = TRUE)
}
\arguments{
\item{dataset}{dataframe or matrix over which to apply filter}

\item{x}{column in dataframe over which to filter will be applied}

\item{exp}{Filter expression. Should take on the form of 'x<100' or is.na(x)==F.}

\item{save.filter}{Whether to save the filterTable as a csv file}

\item{flog.dat}{Whether to print filterTable to the log file}
}
\value{
filterTable saved into the global environment. The data table will grow with each run of the function.
}
\description{
Define and store filters
}
\details{
This function allows users to define and store data filters which can then be applied to the data. The filter dataframe can be saved and will be logged in the log file.
}
\keyword{filter,}
\keyword{subset}
