% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dat.R
\name{filter_table}
\alias{filter_table}
\title{Define and store filter expressions}
\usage{
filter_table(dat, project, x, exp)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{x}{Variable in \code{dat} over which filter will be applied.}

\item{exp}{Filter expression. Should take on the form of \code{"x<100"} or \code{"is.na(x) == F"}.}
}
\value{
Filter expressions saved as a table to the FishSET database.
}
\description{
Define and store filter expressions
}
\details{
This function allows users to define and store data filter expressions which can then be applied to the data.
The filter table will be saved in the FishSET database under the \code{project} name and `filterTable`.
The new filter functions are added each time the function is run and the table is  automatically updated in the
FishSET database. The function call will be logged in the log file.
}
\examples{
\dontrun{
filter_table(pcodMainDataTable, 'pcod', 'PERFORMANCE_Code','PERFORMANCE_Code==1')
}

}
\keyword{filter,}
\keyword{subset}
