% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Generate centroid of polygon of zone or area}
\usage{
find_centroid(dat, gridfile, lon.dat = NULL, lat.dat = NULL, cat,
  lon.grid = NULL, lat.grid = NULL, weight.var = NULL)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishet_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data set. Can be shape file, data frame, or list.}

\item{lon.dat}{Longitude of points from dataset.}

\item{lat.dat}{Latitude of points from dataset.}

\item{cat}{Variable defining the individual areas or zones.}

\item{lon.grid}{Longitude of points from gridfile.}

\item{lat.grid}{Latitude of points from gridfile.}

\item{weight.var}{Variable for weighted average.}
}
\value{
Data frame where each row is a unique zone and columns are the latitude and longitude defining the centroid of each zone.
}
\description{
Generate centroid of polygon of zone or area
}
\details{
Functions returns the center of a zone or area based on set of latitude and longitudes. Zone centroids are used in calculating distance matrices.
 Function can also return a weighted centroid. Code works for data frames and shape files. Lists can also be used as long as inputs are a list of 
 latitudes, longitudes, and areas or zones. Calls \code{\link{assignment_column}} function.
}
\keyword{centroid,}
\keyword{polygon}
\keyword{zone,}
