% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{fishset_compare}
\alias{fishset_compare}
\title{Compare new dataset to previous version}
\usage{
fishset_compare(x, y, compare = c(TRUE, FALSE))
}
\arguments{
\item{x}{name dataframe to be saved}

\item{y}{name of previously saved dataframe}

\item{compare}{TRUE/FALSE Compare new dataframe to previously saved dataframe before saving dataframe x to databases}
}
\description{
Compare new dataset to previous version
}
\details{
If compare is TRUE, colnames of the new and previously saved dataframe are compared for consistency. If false, no comparison is made and the new file is saved to the database.
}
