% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getis_ord_stats.r
\name{getis_ord_stats}
\alias{getis_ord_stats}
\title{Wrapper function to calculate Getis-Ord statistics}
\usage{
getis_ord_stats(
  dat,
  varofint,
  gridfile,
  lon.dat = NULL,
  lat.dat = NULL,
  cat = NULL,
  lon.grid = NULL,
  lat.grid = NULL
)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishet_db database should contain the string `MainDataTable`.}

\item{varofint}{Numeric variable of interest to test for spatial high/low clustering.}

\item{gridfile}{Spatial data set. Can be shape file, data frame, or list.}

\item{lon.dat}{Longitude of points from dataset.}

\item{lat.dat}{Latitude of points from dataset.}

\item{cat}{Variable defining the individual areas or zones.}

\item{lon.grid}{Longitude of points from gridfile.}

\item{lat.grid}{Latitude of points from gridfile.}
}
\value{
moranmap: ggplot2 object; morantable: table of statistics
}
\description{
Wrapper function to calculate global and local Getis-Ord by discrete area
}
\details{
Calculates degree, within each zone, that high or low values of the varofint cluster in space. 
Function utilizes the localmoran and knearneigh functions from the spdep package. All parameters are set to NULL except the varofint and location variables.
The spatial input is a row-standardized spatial weights matrix for computed nearest neighbor matrix, which is the null setting for the nb2listw function.
Requires A data frame with area as a factor, the lon/lat centroid for each area ("centroid_lon",
    "centroidlat"), the path_lon/path_lat outlining each area
    ("path_lon", "path_lat"), and the variable of interest ("varofint") or 
 a map file with lat/lon defining boundaries of area/zones and variable of interest for weighting.
Also required is the lat/lon defining the center of a zone/area. If the centroid is not included in the map file, 
then the find_centorid function can be called to calculate the centroid of each zone. If the varible of interest is not
associated with a area/zone than the assignement_column function can be used to assign each observation to a zone.  
Parameters to identify centroid and assign variable of interest to area/zone are optional and default to NULL.
}
\examples{
\dontrun{
names(datatomap)[1] <- "path_lon"
names(datatomap)[2] <- "path_lat"
names(datatomap)[3] <- "centroid_lon"
names(datatomap)[4] <- "centroid_lat"
names(datatomap)[5] <- "ADFGstat6"
names(datatomap)[6] <- "varofint"
getis_ord_stats(datatomap)
}

}
