% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{globalcheck_view}
\alias{globalcheck_view}
\title{View error output from discrete choice model for the defined project}
\usage{
globalcheck_view(table, project)
}
\arguments{
\item{table}{Table name in FishSET database. Should contain the project, the 
  phrase 'LDGlobalCheck', and a date in YMD format (20200101). 
Table name must be in quotes.}

\item{project}{Name of project}
}
\description{
Returns error output from running the \code{\link{discretefish_subroutine}} 
function. The table argument must be the full name of the table name in the 
FishSET database. Use \code{\link{tables_database}}to view table names in 
FishSET database.
}
\examples{
\dontrun{
globalcheck_view('pcodLDGlobalCheck20190604', 'pcod')
}
}
