% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haul_to_trip.R
\name{haul_to_trip}
\alias{haul_to_trip}
\title{Collapse data frame from haul to trip}
\usage{
haul_to_trip(
  dat,
  project,
  fun.numeric = mean,
  fun.time = mean,
  tripID,
  haul_count = TRUE,
  log_fun = TRUE
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{fun.numeric}{How to collapse numeric or temporal data. For example, \code{min}, \code{mean}, 
\code{max}, \code{sum}. Defaults to \code{mean}.}

\item{fun.time}{How to collapse temporal data. For example, \code{min}, \code{mean}, \code{max}. 
Cannot be \code{sum} for temporal variables.}

\item{tripID}{Column(s) that identify the individual trip.}

\item{haul_count}{Logical, whether to return a column of the number of hauls per trip.}

\item{log_fun}{Logical, whether to log function call (for internal use).}
}
\value{
Returns the primary dataset where each row is a trip.
}
\description{
Collapse data frame from haul to trip
}
\details{
Collapses primary dataset from haul to trip level. Unique trips are defined based on selected column(s), for 
  example, landing permit number and disembarked port. This id column is used to collapse the
  data to trip level.  \code{fun.numeric} and \code{fun.time} define how multiple observations for a trip
  are collapsed. For variables that are not numeric or dates, the first observation is used.
}
\examples{
\dontrun{
pollockMainDataTable <- haul_to_trip("pollockMainDataTable","pollock",
    min, mean, "PERMIT", "DISEMBARKED_PORT"
    )
}

}
