% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haul_to_trip.R
\name{haul_to_trip}
\alias{haul_to_trip}
\title{Collapse the dataframe from haul to trip.}
\usage{
haul_to_trip(dat, project, fun.numeric = mean, fun.time = mean, ...)
}
\arguments{
\item{dat}{Data frame containing haul level data. In the fishset_db database, the table will contain the phrase `MainDataTable`}

\item{project}{Name of project}

\item{fun.numeric}{Numeric function defining how to collapse numeric or temporal data. For example, min, mean, max, sum. Defaults to mean.}

\item{fun.time}{Numeric function defining how to collapse temporal data. For example, min, mean, max. Cannot be sum for temporal variables.}

\item{...}{Column(s) that identify the individual trip.}
}
\value{
Data frame with each row representing a trip or haul
}
\description{
Collapse the dataframe from haul to trip.
}
\details{
Collapses the main data table from haul to trip level. Requires the MainDataTableInfo table associated with the main data table. The MainDataTableInfo table is first updated
using the dataindex_update function. Unique trips are defined based on selected column(s). For example, landing permit number and disembarked port.
 This id column is used to collapse the data to trip level. Users can define how time and numeric variables with multiple observations for a trip are collapsed.
For instance, the mean value of the numeric observations. For non-numeric and non-time variables, the first observation is used.
}
\examples{
\dontrun{
 dat <- haul_to_trip(pollockMainDataTable, 'pollock',min,mean,'PERMIT','DISEMBARKED_PORT')
 }
}
