% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaulToTrip.R
\name{HaulToTrip}
\alias{HaulToTrip}
\title{Collapse the dataframe from haul to trip.
Each row should be a unique trip. Reduces the number of rows.
Unique trips are defined based on selected column(s). For example, landing permit number and disembared port. This id column is used to collapse the data to trip level.}
\usage{
HaulToTrip(dataset, dataindex, varnameindex, genTypeName, ...,
  fun.time = min, fun.numeric = mean)
}
\arguments{
\item{dataset}{dataframe or matrix over which to apply filter}

\item{dataindex}{Dataframe that contains information on each column of the dataset}

\item{varnameindex}{The column in dataindex that contains the column names of dataset}

\item{genTypeName}{column in dataindex containing information on the general category of each column in dataset (time, numeric, etc)}

\item{...}{Column(s) that define the individual trip.}

\item{rowID}{created in the first half of the HaulToTrip function. It is based on the column(s) identified as defining individual trips}
}
\value{
Dataframe with each row representing a trip or haul
}
\description{
Collapse the dataframe from haul to trip.
Each row should be a unique trip. Reduces the number of rows.
Unique trips are defined based on selected column(s). For example, landing permit number and disembared port. This id column is used to collapse the data to trip level.
}
\examples{
MainDataTable and MainDataTableInfo
dat <- HaulToTrip(MainDataTable, MainDataTableInfo,'variable_name','generalType','PERMIT','DISEMBARKED_PORT')
}
