% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_var.R
\name{ID_var}
\alias{ID_var}
\title{Create ID variable from one or more variables}
\usage{
ID_var(dat, newID, ...)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{newID}{name of new ID column}

\item{...}{Column(s) in main data frame that define unique observations.}
}
\value{
Data frame with `newID` ID variable included.
}
\description{
Create ID variable from one or more variables
}
\details{
Function is used to create a variable that indicates distinct hauls or trips.
}
\examples{
\dontrun{
dat <- ID_var(MainDataTable, newID='PermitID','GEAR_TYPE','TRIP_SEQ')
}
}
