% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_var.R
\name{ID_var}
\alias{ID_var}
\title{Create ID variable from another variable.}
\usage{
ID_var(dataset, newID, ...)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{newID}{name of new ID column}

\item{...}{Column(s) that define the variable(s) over which to make new ID}
}
\value{
Returns the dataframe with new ID variable included
}
\description{
Create ID variable from another variable.
}
\details{
Creates an variable to indicate distinct hauls or trips
}
