% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{insert_table}
\alias{insert_table}
\title{Insert table from user folder}
\usage{
insert_table(out, project)
}
\arguments{
\item{out}{String, table file name.}

\item{project}{Name of project.}
}
\description{
Insert table from user folder
}
\examples{
\dontrun{
insert_table("pollock_table.csv")
}
}
