% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{list_tables}
\alias{list_tables}
\title{Display FishSET database tables by type}
\usage{
list_tables(project, type = "main")
}
\arguments{
\item{project}{A project name to show main tables by.}

\item{type}{the type of fishset_db table to search for. Options include 
"main" (MainDataTable), "port" (PortTable), "spat" (SpatTable), "grid" 
(GridTable), "aux" (AuxTable) "ec" (ExpectedCatch),  "altc" (AltMatrix), 
"info" (MainDataTableInfo), "gc" (ldglobalcheck), "fleet" (FleetTable), 
"filter" (FilterTable), "centroid" (Centroid or FishCentroid),  "model" 
(ModelOut), "model data" or "model design" (ModelInputData).}
}
\description{
Show project table names by table type. To see all tables for all projects 
in the FishSETFolder, use \code{\link{fishset_tables}}.
}
\examples{
\dontrun{
list_tables("pollock", type = "main")
list_tables("pollock", "ec")
}

}
