% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_grid}
\alias{load_grid}
\title{Import, parse, and save gridded data to FishSET database}
\usage{
load_grid(dat, grid, name, over_write = TRUE, project = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. 
Table in FishSET database contains the string 'MainDataTable'.}

\item{grid}{File name, including path, of gridded data.}

\item{name}{Name gridded data should be saved as in FishSET database.}

\item{over_write}{Logical, If TRUE, saves dat over previously saved data table 
in the FishSET database.}

\item{project}{String, name of project.}
}
\description{
Gridded data is data that varies by two dimensions. Column names must be zone 
names. Load, parse, and save gridded data to FishSET database.  A project must 
exist before running \code{load_grid()}. See \code{\link{load_maindata}}
to create a new project.
}
\details{
Grid data is an optional data frame that contains a variable that 
  varies by the map grid (ex. sea surface temperature, wind speed). Data can 
  also vary by a second dimension (e.g., date/time). Both dimensions in the 
  gridded data file need to be variables included in the primary data set.
  The grid locations (zones) must define the columns and the optional second 
  dimension defines the rows. The row variable must have the exact name as the 
  variable in the main data frame that it will be linked to. The function DOES 
  NOT check that column and row variables match a variable in the primary data set.
  The function checks that each row is unique, that no variables are empty, 
  and that column names are case-insensitive unique. These data issues are 
  resolved before the data is saved to the database. The data is saved in the 
  FishSET database as the raw data and the working data. In both cases, the 
  table name is the \code{project} and the file name \code{x}. Date is attached 
  to the name for the raw data. The naming convention for gridded tables is 
  "projectNameGridTable". See \code{\link{table_view}} to view/load gridded 
  tables into the working environment.
}
\examples{
\dontrun{
load_grid(dat = 'pcodMainDataTable', name = 'SeaSurfaceTemp', 
          over_write = TRUE, project = 'pcod')
}

}
\seealso{
\code{\link{table_view}}, \code{\link{load_maindata}},
  \code{\link{write_dat}}
}
