% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_maindata}
\alias{load_maindata}
\title{Load data into sql database}
\usage{
load_maindata(dataset, over_write = TRUE, project = NULL,
  compare = FALSE, y = NULL)
}
\arguments{
\item{dataset}{name dataframe to be saved}

\item{over_write}{TRUE/FALSE Save over previously saved file or not}

\item{project}{name of project}

\item{compare}{TRUE/FALSE Compare new dataframe to previously saved dataframe before saving dataframe x to databases}

\item{y}{name of previously saved dataframe. y must be defined if compare==TRUE}
}
\description{
Load data into sql database
}
\details{
Runs the fishset.compare function. Then uses the new dataframe x to generate the the information table that contains information for each variable on units, data format, and specialied variable.
}
