% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{load_spatial}
\alias{load_spatial}
\title{Import, parse, and save spatial data}
\usage{
load_spatial(
  spat,
  name = NULL,
  over_write = TRUE,
  project,
  data.type = NULL,
  lon = NULL,
  lat = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{spat}{File name, including path, of spatial data.}

\item{name}{Name spatial data should be saved as in FishSET project folder.
Cannot be empty or contain spaces.}

\item{over_write}{Logical, If TRUE, saves \code{spat} over previously saved data 
table in the FishSET project folder.}

\item{project}{String, name of project.}

\item{data.type}{Data type argument passed to \code{\link{read_dat}}. Must 
be given if reading from a shape folder, e.g. \code{data.type = "shape"}. 
Otherwise, this argument is optional.}

\item{lon}{Variable or list from \code{spat} containing longitude data. 
Required for csv files. Leave as NULL if \code{spat} is a shape or json file.}

\item{lat}{Variable or list from \code{spat} containing latitude data. 
Required for csv files. Leave as NULL if \code{spat} is a shape or json file}

\item{id}{Polygon ID column. Required for csv files. Leave as NULL if 
\code{spat} is a shape or json file.}

\item{...}{Additional argument passed to \code{\link{read_dat}}.}
}
\description{
Import, parse, and save spatial data
}
\details{
Function to import, parse, and saved project folder in `FishSETFolder` 
 directory. Note, the spatial file is saved as a geojson. To export as  shape 
 file, use \code{\link{write_dat}} specifying `type='shp'`. \code{load_spatial} 
 performs basic quality check before saving spatial tables to the project 
 data folder. To be saved, the spatial must pass the checks in 
 \code{\link{check_spatdat}}. The spatial table is converted to an \code{sf} 
 object, and checked for unique rows and empty columns.
}
\examples{
\dontrun{
load_spatial(spat = 'pcodMainDataTable', name = 'tenMinSqr', over_write = TRUE, project = 'pcod')
}
}
