% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reset.R
\name{log_call}
\alias{log_call}
\title{Reset function calls saved in log file}
\usage{
log_call(fun.name)
}
\arguments{
\item{fun.name}{Function name}
}
\description{
Reset function calls saved in log file
}
\details{
uplodate log file
}
