% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_func_model.R
\name{log_func_model}
\alias{log_func_model}
\title{Log user-created functions or models}
\usage{
log_func_model(x, project)
}
\arguments{
\item{x}{Name of function.}

\item{project}{Project name.}
}
\description{
Log user-created functions or models
}
\details{
Logs function name, arguments, and, call. Use this function to log user-defined likelihood functions.
}
\examples{
\dontrun{
my_func <- function(a, b) {
  a + b
}
log_func_model(my_func)
}

}
