% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_func_model.R
\name{log_func_model}
\alias{log_func_model}
\title{Log user created functions or models}
\usage{
log_func_model(x)
}
\arguments{
\item{x}{Name of cunction}
}
\description{
Log user created functions or models
}
\details{
Logs names of the function, parameters in function, and function call in the dated log file.
}
