% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rerun.R
\name{log_rerun}
\alias{log_rerun}
\title{Console function for rerunning project log}
\usage{
log_rerun(log_file, dat = NULL, ind = NULL, run = FALSE)
}
\arguments{
\item{log_file}{String, name of the log file starting with the date (YYYY-MM-DD) and
ending in ".json".}

\item{dat}{String, new data table to rerun log with.}

\item{ind}{Numeric, indices of function calls to rerun.}

\item{run}{Logical, whether to run the logged function calls (TRUE) or simply
list all function calls (FALSE).}
}
\description{
Console function for rerunning project log
}
\examples{
\dontrun{
log_rerun("2020-10-23.json", run = TRUE) # reruns entire log with original data table
log_rerun("2020-10-23.json", dat = "pollockMainDataTable", run = TRUE) # runs log with new data table
}
}
\seealso{
\code{\link{log_rerun_gui}}
}
