% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reset.R
\name{log_reset}
\alias{log_reset}
\title{Reset function calls saved in log file}
\usage{
log_reset()
}
\description{
Reset function calls saved in log file
}
\details{
Removes the three log file lists from the working directory
}
