% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_avgcat.R
\name{logit_avgcat}
\alias{logit_avgcat}
\title{Average catch multinomial logit procedure}
\usage{
logit_avgcat(starts3, dat, otherdat, alts, project, expname, mod.name)
}
\arguments{
\item{starts3}{Starting values as a vector (num). For this likelihood, the order takes:
c([catch function parameters], [cost (distance) parameters]). \cr \cr
The catch function and cost parameters are of length (# of catch variables)*(kk-1) and (# of cost variables) respectively,
where kk equals the number of alternatives.}

\item{dat}{Data matrix, see output from shift_sort_x, alternatives with distance.}

\item{otherdat}{Other data used in model (as list containing objects griddat and intdat). \cr \cr
For grid-specific variables griddat and cost variables to be interacted with distance intdat, any number of variables are allowed, as a list of matrices. 
Note the variables (each as a matrix) within `griddat` and `intdat` have no naming restrictions.
Also note that `griddat` variables are dimension *(number of observations) x 1*, 
while `intdat` variables are dimension *(number of observations) x 1*, to be interacted with the distance to each alternative.
Grid-specific variables may correspond to catches that vary by location, 
or interaction variables may be vessel characteristics that affect how much disutility is suffered by traveling a greater distance.
Note in this likelihood the the grid-specific variables are are only of dimension *(number of observations) x 1* (compared to other likelihoods), and
each variable varies across observations but not for each location: they are grid-specific due to the location-specific coefficients. 
Also note that only (kk-1) grid-specific parameters are estimated: one alternative must be dropped and the remaining parameters are estimated relative to the
dropped (first) alternative.
If there are no other data, the user can set `griddat` as ones with dimension *(number of observations) x 1*
and `intdat` variables as ones with dimension *(number of observations) x 1*.}

\item{alts}{Number of alternative choices in model as length 1 vector (num).}

\item{project}{Name of project}

\item{expname}{Expected catch table}

\item{mod.name}{Name of model run for model result output table}
}
\value{
ld: negative log likelihood
}
\description{
Average catch multinomial logit procedure
}
\examples{
\dontrun{
data(zi)
data(catch)
data(choice)
data(distance)
data(si)

optimOpt <- c(1000,1.00000000000000e-08,1,0)

methodname <- "BFGS"

si2 <- sample(1:5,dim(si)[1],replace=TRUE)
zi2 <- sample(1:10,dim(zi)[1],replace=TRUE)

otherdat <- list(griddat=list(si=as.matrix(si),si2=as.matrix(si2)),intdat=list(zi=as.matrix(zi),zi2=as.matrix(zi2)))

initparams <- c(1.5, 1.25, 1.0, 0.9, 0.8, 0.75, -1, -0.5)

func <- logit_avgcat

results <- discretefish_subroutine(catch,choice,distance,otherdat,initparams,optimOpt,func,methodname)
}
}
