% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_predict.R
\name{logit_predict}
\alias{logit_predict}
\title{Logit predict}
\usage{
logit_predict(project, mod.name, expected.catch = NULL)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Name of saved model to use. Argument can be the name of the model or can pull the name 
of the saved "best" model. Leave \code{mod.name} empty to use the saved "best" model. If more than
one model is saved, \code{mod.name} should be the numeric indicator of which model to use.
Use \code{table_view("modelChosen", project)} to view a table of saved models.}

\item{expected.catch}{Required for conditonal logit (\code{logit_c}) model. 
Name of expected catch table to use. 
 Can be the expected catch from the short-term scenario (\code{short}), the medium-term scenario (\code{med}), the 
 long-term scenario (\code{long}), or the user-defined temporal parameters (\code{user}).}
}
\value{
Returns probability of logit model by choice
}
\description{
Prediction component from logit models (non mixed) called in Policy3, under predict_model_tempNew.m
}
\keyword{internal}
