% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{main_tables}
\alias{main_tables}
\title{Display MainDataTables}
\usage{
main_tables(project = NULL)
}
\arguments{
\item{project}{A project name to filter main tables by. Returns all MainDataTables
if \code{NULL}.}
}
\description{
Display MainDataTables
}
\examples{
\dontrun{
main_tables()
main_tables("pollock")
}
}
