% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_design.R
\name{make_model_design}
\alias{make_model_design}
\title{Make model design}
\usage{
make_model_design(dat, catchID, alternativeMatrix = c("loadedData",
  "griddedData"), lon.dat, lat.dat, project, indeVarsForModel = NULL,
  gridVariablesInclude = NULL, priceCol = NULL)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{catchID}{Name of variable that contains catch data such as 'HAUL'}

\item{alternativeMatrix}{Whether the alternative choice matrix should come from 'loaded data' or 'gridded data'}

\item{lon.dat}{longitude Column containing longitude data}

\item{lat.dat}{latitude Column containing latitude data}

\item{project}{name. name of project. For name of output table saved in sql database}

\item{indeVarsForModel}{list List columns names of independent variables to include in the model using `c()`.}

\item{gridVariablesInclude}{list List data set that varies over a grid to include in the model using `c()`.}

\item{priceCol}{NULL If required, specify which variable contains price data.}
}
\value{
Model design matrix containing \cr
  \tabular{rlll}{
    choice: \tab Data corresponding to actual zonal choice\cr 
    catch: \tab Data corresponding to actual zonal catch\cr 
    scales: \tab Scale vectors to put catch data, zonal data, and other data on same scale\cr 
    distance: \tab Data corresponding to distance\cr
    instances: \tab Number of observations\cr 
    alt: \tab Number of alternative zones\cr
    epmDefaultPrice: \tab Price data\cr 
    dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that zone is to be included.\cr 
    numOfNecessary: \tab Minimum number of hauls/trips per zone for data from that zone to be included\cr
    typeOfNecessary: \tab Haul or trip\cr
    altChoiceType: \tab Function choice. Set to distance\cr
    altChoiceUnits: \tab Units of distance\cr
    altToLocal: \tab Identifies how to find lat/lon for starting point. Can be zonal centroid, port, etc\cr
    altToLocal2: \tab Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment'\cr 
    bCHeader: \tab Variables to include in the model that do not vary by zone. Includes independent variables and interactions\cr
    gridVaryingVariables: \tab Variables to include in the model that do vary by zone such as expected catch (from \code{\link{create_expectations}} function)
 }
}
\description{
Make model design
}
\details{
Functions returns model design matrix. Calls the Alternative Choice matrix from `create_alternative_choice` function which defines alternative fishing options
and the expected catch from the `create_expectations` function. The distance from the starting point to alternative choices is calculated.
}
\examples{
\dontrun{
make_model_design(MainDataTable, catchID = 'HAUL', alternativeMatrix = "loadedData", 
                  'LonLat_START_LON', 'LonLat_START_LAT', project = 'pcod')
}
}
