% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_design.R
\name{make_model_design}
\alias{make_model_design}
\title{Make model design}
\usage{
make_model_design(
  dat,
  catchID,
  alternativeMatrix = c("loadedData", "griddedData"),
  lon.dat,
  lat.dat,
  project,
  likelihood = NULL,
  vars1 = NULL,
  vars2 = NULL,
  priceCol = NULL,
  startloc = NULL,
  polyn = NULL
)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in fishset_db database should contain the string `MainDataTable`.}

\item{catchID}{Name of variable that contains catch data such as 'HAUL'}

\item{alternativeMatrix}{Whether the alternative choice matrix should come from 'loaded data' or 'gridded data'}

\item{lon.dat}{longitude Column containing longitude data}

\item{lat.dat}{latitude Column containing latitude data}

\item{project}{name. name of project. For name of output table saved in sql database}

\item{likelihood}{Name of likelihood function. Current choices are logit_c, logit_avgcat, logit_correction, epm_normal, epm_weibull, epm_ognormal.}

\item{vars1}{List varialbes using `c()`. These depend on the likelihood 
the user chooses, so please see the Detail section for how to specify for each likelihood function.}

\item{vars2}{List varialbes using `c()`. These depend on the likelihood
the user chooses, so please see the Detail section for how to specify for each likelihood function.}

\item{priceCol}{NULL If required, specify which variable contains price data.}

\item{startloc}{Vector required for logit_correction likelihood. startloc is a matrix of dimension (number of observations) 
by (unity), that corresponds to the starting location when the agent decides between alternatives.}

\item{polyn}{Vector required for logit_correction likelihood. Correction polynomial degree.}
}
\value{
Model design matrix containing \cr
  \tabular{rlll}{
    likelihood: \tab Name of likelihood function\cr
    choice: \tab Data corresponding to actual zonal choice\cr 
    catch: \tab Data corresponding to actual zonal catch\cr 
    scales: \tab Scale vectors to put catch data, zonal data, and other data on same scale\cr 
    distance: \tab Data corresponding to distance\cr
    instances: \tab Number of observations\cr 
    alt: \tab Number of alternative zones\cr
    epmDefaultPrice: \tab Price data\cr 
    dataZoneTrue: \tab Vector of 0/1 indicating whether the data from that zone is to be included.\cr 
    numOfNecessary: \tab Minimum number of hauls/trips per zone for data from that zone to be included\cr
    typeOfNecessary: \tab Haul or trip\cr
    altChoiceType: \tab Function choice. Set to distance\cr
    altChoiceUnits: \tab Units of distance\cr
    altToLocal: \tab Identifies how to find lat/lon for starting point. Can be zonal centroid, port, etc\cr
    altToLocal2: \tab Identifies how to find lat/lon for alternative choices such as 'Centroid of Zonal Assignment'\cr 
    bCHeader: \tab Variables to include in the model that do not vary by zone. Includes independent variables and interactions\cr
    gridVaryingVariables: \tab Variables to include in the model that do vary by zone such as expected catch (from \code{\link{create_expectations}} function)
  }
}
\description{
Create a list containing likelihood function, parameters, and data to be pass to model call function
}
\details{
Functions returns model design matrix. Calls the Alternative Choice matrix from `create_alternative_choice` function which defines alternative fishing options
and the expected catch from the `create_expectations` function. The distance from the starting point to alternative choices is calculated.
Variable names details \cr
\tabular{lllllll}{
\tab \tab \strong{vars1} \tab \tab \strong{vars2} \tab \cr \cr
\strong{logit_c}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"alternative-specific variables" vary across alternatives,
    e.g. catch rates. Each variable name therefore corresponds to data with
    dimensions (number of observations) by (number of alternatives), and
    returns a single parameter for each variable (e.g. the marginal utility
    from catch).} \cr \cr
\strong{logit_avgcat}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"average-catch variables" are alternative-invariant
    variables, e.g. vessel gross tonnage. Each variable name therefore
    corresponds to data with dimensions (number of observations) by (unity),
    and returns (k-1) parameters where (k) equals the number of alternatives,
    as a normalization of parameters is needed as the probabilities sum to
    one. Interpretation is therefore relative to the first
    alternative.} \cr \cr
\strong{epm_normal}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"catch-function variables" are alternative-invariant
    variables that are interacted with zonal constants to form the catch
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns (k)
    parameters where (k) equals the number of alternatives.} \cr \cr
\strong{epm_lognormal}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"catch-function variables" are alternative-invariant
    variables that are interacted with zonal constants to form the catch
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns (k)
    parameters where (k) equals the number of alternatives.} \cr \cr
\strong{epm_weibull}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"catch-function variables" are alternative-invariant
    variables that are interacted with zonal constants to form the catch
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns (k)
    parameters where (k) equals the number of alternatives.} \cr \cr
\strong{logit_correction}: \tab \tab
    \preformatted{"travel-distance variables" are alternative-invariant
    variables that are interacted with travel distance to form the cost
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns a
    single parameter.} \tab \tab
    \preformatted{"catch-function variables" are alternative-invariant
    variables that are interacted with zonal constants to form the catch
    portion of the likelihood. Each variable name therefore corresponds to
    data with dimensions (number of observations) by (unity), and returns (k)
    parameters where (k) equals the number of alternatives.} \cr \cr
}
}
\examples{
\dontrun{
make_model_design(MainDataTable, catchID = 'HAUL', alternativeMatrix = "loadedData", 
                  'LonLat_START_LON', 'LonLat_START_LAT', project = 'pcod')
}
}
