% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_out_view}
\alias{model_out_view}
\title{Load discrete choice model output to console for the defined project}
\usage{
model_out_view(table, project)
}
\arguments{
\item{table}{Table name in FishSET database. Should contain the phrase 
'modelout'. Table name must be in quotes.}

\item{project}{Name of project}
}
\description{
Returns output from running the discretefish_subroutine function.
  The table parameter must be the full name of the table name in the FishSET database.
}
\details{
Returns output from running \code{\link{discretefish_subroutine}}. The table 
argument must be the full name of the table name in the FishSET database. 
Output includes information on model convergence, standard errors, t-stats, etc.
}
\examples{
\dontrun{
model_out_view('pcodmodelout20190604', 'pcod')
}
}
