% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{model_params}
\alias{model_params}
\title{Load model parameter estimates, standard errors, and t-statistic to console 
for the defined project}
\usage{
model_params(project, output = "list")
}
\arguments{
\item{project}{Name of project}

\item{output}{Options include list, table, or print.}
}
\description{
Returns parameter estimates, standard errors, and t-statistic 
 from running the discretefish_subroutine function. The table parameter must 
 be the full name of the table name in the FishSET database.
}
\details{
Returns parameter estimates from running \code{\link{discretefish_subroutine}}. 
The table argument must be the full name of the table name in the FishSET 
database.
}
\examples{
\dontrun{
model_params('pcod')
}

}
