% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction.R
\name{model_prediction}
\alias{model_prediction}
\title{Model prediction function}
\usage{
model_prediction(project, mod.name, expected.catch = NULL, enteredPrice = NULL)
}
\arguments{
\item{project}{Name of project}

\item{mod.name}{Model name. Argument can be the name of the model or the name can be pulled the modelChosen table.
 Leave \code{mod.name} empty to use the name of the saved "best" model. If more than
one model is saved, \code{mod.name} should be the numeric indicator of which model to use.
Use \code{table_view("modelChosen", project)} to view a table of saved models.}

\item{expected.catch}{Required for conditional logit (\code{logit_c}) model. 
Name of expected catch table to use. 
 Can be the expected catch from the short-term scenario (\code{short}), the medium-term scenario (\code{med}), the 
 long-term scenario (\code{long}), or the user-defined temporal parameters (\code{user}).}

\item{enteredPrice}{NEED TO FIGURE OUT WHAT EXACTLY THIS IS}
}
\description{
Model prediction function
}
\details{
Calls \code{\link{logit_predict}}, \code{\link{epm_predict}}, and \code{\link{predict_probability}}.
   Closure scenarios and TAC must be define using \code{\link{zone_closure}} function before function can be run.
}
\keyword{(ASC)}
\keyword{Logit'}
\keyword{Zonal}
\keyword{internal}
