% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_stats.r
\name{moran_stats}
\alias{moran_stats}
\title{Wrapper function to calculate Moran's I}
\usage{
moran_stats(datatomap)
}
\arguments{
\item{datatomap}{A data frame with the ADFG Stat6 area as a factor
"ADFGstat6", the lon/lat centroid for each area ("centroid_lon",
"centroidlat"), the path_lon/path_lat outlining each area
("path_lon", "path_lat"), and the variable of interest ("varofint").}
}
\value{
moranmap: ggplot2 object; morantable: table of statistics
}
\description{
Wrapper function to calculate global and local Moran's I by discrete area
}
\examples{
\dontrun{
names(datatomap)[1] <- "path_lon"
names(datatomap)[2] <- "path_lat"
names(datatomap)[3] <- "centroid_lon"
names(datatomap)[4] <- "centroid_lat"
names(datatomap)[5] <- "ADFGstat6"
names(datatomap)[5] <- "varofint"
moran_stats(datatomap)
}

}
