% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_stats.r
\name{moran_stats}
\alias{moran_stats}
\title{Calculate Moran's I statistic}
\usage{
moran_stats(dat, var, dat_zone, spat, spat_zone, project)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in FishSET database 
contains the string 'MaindataTable'.}

\item{var}{Numeric variable from \code{dat} to test for spatial autocorrelation.}

\item{dat_zone}{Variable in \code{dat} that identifies the individual zones or
areas. Define if exists in \code{dat} and is not named `ZoneID`. Defaults to NULL.}

\item{spat}{Spatial data containing information on fishery management or regulatory zones.
Shape, json, geojson, and csv formats are supported.}

\item{spat_zone}{Variable or list in \code{spat} that identifies the individual areas or zones.
If \code{spat} is class sf, \code{cat} should be name of list containing information on zones.}
}
\value{
Returns a list with (1) global Moran's I stats, (2) Moran lagged plot, (3) LISA cluster
 map summarizing local Moran's values
}
\description{
This wrapper function calculates global and local Moran's I statistics to measure spatial 
autocorrelation by discrete area.
}
\details{
The function measures the degree of spatial autocorrelation and utilizes functions from the 
`spdep` package. The function requires a spatial file with latitude and longitude 
coordinates defining the boundaries of areas or zones and a variable of interest (`var`) to 
test for spatial autocorrelation.
}
\examples{
\dontrun{
moran_stats(pcodMainDataTable, var='OFFICIAL_MT_TONS', dat_zone='zoneID',
  spat=spatdat, spat_zone='NMFS_AREA', project = 'pcod')
}

}
