% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_filter.R
\name{na_filter}
\alias{na_filter}
\title{Filters NaN's from variable}
\usage{
na_filter(dataset, x, replace = F, remove = F, rep.value = NA)
}
\arguments{
\item{dataset}{dataframe or matrix over which to check for NAs}

\item{x}{column in dataframe over which to remove or replace NAs}

\item{replace}{whether to (TRUE) or not to (FALSE) replace NAs in a column. Defaults to FALSE.}

\item{remove}{whether to (TRUE) or not to (FALSE) remove all remove the entire row of the dataframe where NA is present in a specified column. Defaults to FALSE.}

\item{rep.value}{value to replace all NAs in a column. Defaults to the mean value of the column.}
}
\value{
Returns the modified dataframe
}
\description{
Function to return a modified dataframe where NaNshave been replaced or removed.
}
\keyword{NA}
