% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_filter.R
\name{nan_filter}
\alias{nan_filter}
\title{Filters NaN's from variable}
\usage{
nan_filter(dataset, x, replace = F, remove = F, rep.value = NA)
}
\arguments{
\item{dataset}{dataframe or matrix over which to check for NaNs}

\item{x}{column in dataframe over which to remove or replace NaNs}

\item{replace}{whether to (TRUE) or not to (FALSE) replace NaNs in a column. Defaults to FALSE.}

\item{remove}{whether to (TRUE) or not to (FALSE) remove all remove the entire row of the dataframe where NaN is present in a specified column. Defaults to FALSE.}

\item{rep.value}{value to replace all NaNs in a column. Defaults to the mean value of the column.}
}
\value{
Returns the modified dataframe
}
\description{
Function to return a modified dataframe where NaNshave been replaced or removed.
}
\keyword{NaN}
