% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_filter.R
\name{nan_identify}
\alias{nan_identify}
\title{Check whether any columns in the primary dataset contain NAs or NaNs.}
\usage{
nan_identify(dat, project)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. 
Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}
}
\value{
Returns names of columns containing NAs or NaNs, if any.
}
\description{
Check whether any columns in the primary dataset contain NAs or NaNs. 
  Returns column names containing NAs or NaNs.
}
\examples{
\dontrun{
nan_identify(pcodMainDataTable, "pcod")
}
}
\keyword{NA}
\keyword{NaN,}
