% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_filter.R
\name{nan_identify}
\alias{nan_identify}
\title{Filters NaN's from variable
Identify NaN's in dataset}
\usage{
nan_identify(dataset)
}
\arguments{
\item{dataset}{dataframe or matrix over which to check for NaNs}
}
\description{
Filters NaN's from variable
Identify NaN's in dataset
}
\keyword{NaN}
