% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{outlier_remove}
\alias{outlier_remove}
\title{Evaluate and edit outliers from variable
 Contains functions to evaluate the data through histogram, QQ, and x-y plots.
 Contains functions to evaulate the fit of the data in terms of distributions and analysis}
\usage{
outlier_remove(dataset, x, outlier.mod = "none", remove = T)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{column in dataframe}

\item{outlier.mod}{defines method to ..
Choices include:
                none: No data points are removed
                5_95_quant: Removes data points outside the 5th and 95th quantiles
                25_75_quant: Removes data points outside the 25th and 75th quantiles
                mean_2SD: Removes data points outside +/- 2SD of the mean
                median_2SD: Removes data points outside +/- 2SD of the meadian
                mean_3SD: Removes data points outside +/- 3SD of the mean
                median_3SD: Removes data points outside +/- 3SD of the meadian}

\item{remove}{Save data with outliers removed. If TRUE, the revised data tables, based on outlier.mod, is returned.}
}
\value{
Returns the modified dataframe
}
\description{
Evaluate and edit outliers from variable
 Contains functions to evaluate the data through histogram, QQ, and x-y plots.
 Contains functions to evaulate the fit of the data in terms of distributions and analysis
}
\details{
outlier_check is a series of functions allowing users to evaluate the occurrence of outliers and their effect on the data using summary data and plots. Outliers can onlyl be assessed for numeric vectors.
}
\keyword{outliers}
