% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{outlier_table}
\alias{outlier_table}
\title{Outlier check functions
Evaluate outliers in a table output}
\usage{
outlier_table(dataset, x)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{column in dataframe}
}
\value{
Returns table with quantiles for all numeric variables in the dataframe
}
\description{
Outlier check functions
Evaluate outliers in a table output
}
\keyword{outliers}
