% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_map.R
\name{predict_map}
\alias{predict_map}
\title{Map of predicted probabilities}
\usage{
predict_map(project, policy.name, spat, zone.spat)
}
\arguments{
\item{project}{Name of project}

\item{policy.name}{Name of policy scenario}

\item{spat}{A spatial data file containing information on fishery management 
or regulatory zones boundaries. `sf` objects are recommended, but `sp` objects 
can be used as well. See [dat_to_sf()] to convert a spatial table read from 
a csv file to an `sf` object. To upload your spatial data to the FishSETFolder 
see [load_spatial()].}

\item{zone.spat}{Name of zone ID column in `spat`.}
}
\value{
A map showing predicted probabilities
}
\description{
Create a map showing predicted probabilities by zone for a chosen policy scenario
}
\details{
This function requires that model and prediction output tables exist in the FishSET database.
}
\examples{
\dontrun{

predict_map(project = "scallop", policy.name = "logit_c_mod1 closure_1", 
            spat = spat, zone.spat = "TEN_ID")

}
}
