% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{randomize_value_range}
\alias{randomize_value_range}
\title{Randomize value by percentage range}
\usage{
randomize_value_range(dat, project, value, perc = NULL)
}
\arguments{
\item{dat}{Main data frame over which to apply function. Table in FishSET 
database should contain the string `MainDataTable`.}

\item{project}{Project name.}

\item{value}{String, name of variable to jitter.}

\item{perc}{Numeric, a vector of percentages to randomly adjust a column of values by. 
Defaults to a range of 0.05 - 0.15 (i.e. 5-15 percent of original value).}
}
\description{
Randomize value by percentage range
}
\details{
This is one of the FishSET confidentiality functions. It adjusts a 
  value by randomly sampling a range of percentages provided in the "perc" argument.
}
\examples{
\dontrun{
randomize_value_range(pollockMainDataTable, "LBS_270_POLLOCK_LBS")
} 
}
