% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fishset.R
\name{run_fishset_gui}
\alias{run_fishset_gui}
\title{Interactive tool to run FishSET functions 
Runs functions associated with loading data, exploring data, checking for data quality issues, generating new variables, and basic data analyis function.
run_fishset_gui}
\usage{
run_fishset_gui(dat, project)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{project}{Name of project. Parameter is used to generate meaningful table names in fishset_db database.}
}
\description{
Interactive tool to run FishSET functions 
Runs functions associated with loading data, exploring data, checking for data quality issues, generating new variables, and basic data analyis function.
run_fishset_gui
}
\details{
Opens an interactive page that allows users to select which functions to run by clicking check boxes.
}
\examples{
\dontrun{
run_fishset_gui(pcodMainDataTable, 'pcod')
}
}
