% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model_gui.R
\name{select_model}
\alias{select_model}
\title{View and select models}
\usage{
select_model(table, overwrite.table = TRUE)
}
\arguments{
\item{table}{Name of table in database containing model measures of fit}

\item{overwrite.table}{If FALSE, appends model selection to out.mod table. If true, overwrites new table with model selection columns. Set to TRUE if models are deleted or selected models are changed.}
}
\description{
View and select models
}
\details{
Opens an interactive data table that displays model measures of fit for each model run saved in the model measures of fit table in the fishet_db database. 
The name of this table should contain the string `out.mod`. 
Users can delete models from the table and select the preferred model by checking the `selected` box. The table is then saved to the fishet_db database with two new columns added, a TRUE/FALSE selected column
and the date it was selected. The table is saved with the phrase `modelChosen` in the fishset_db database. The function can also be called indirectly in the 
\code{\link{discretefish_subroutine}} by specifying 
the `select.model` parameter as TRUE. The `modelChose` table is not used in any functions. The purpose of this function and the `modelChosen` table is to save a
reference of the preferred model.
}
\examples{
\dontrun{
select_model('out.mod', overwrite.table=FALSE)
}
}
