% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variables.R
\name{set_quants}
\alias{set_quants}
\title{Create quantile variable}
\usage{
set_quants(dataset, x, quant.cat = c(0.2, 0.25, 0.4))
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{Variable to create}

\item{quant.cat}{Quantile categories. Includes 0.2, 0.25, 0.4.}
}
\description{
Create quantile variable
}
\details{
Function for generating new or specialized variables. setQuants creates a coded variable based on the quantiles of x.
}
