% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{skewness}
\alias{skewness}
\title{Evaluate and edit outliers from variable}
\usage{
skewness(x)
}
\arguments{
\item{x}{column in dataframe}

\item{dataset}{dataframe or matrix}

\item{outlier.mod}{defines method to ..
Choices include:
                none: No data points are removed
                5_95_quant: Removes data points outside the 5th and 95th quantiles
                25_75_quant: Removes data points outside the 25th and 75th quantiles
                mean_2SD: Removes data points outside +/- 2SD of the mean
                median_2SD: Removes data points outside +/- 2SD of the meadian
                mean_3SD: Removes data points outside +/- 3SD of the mean
                median_3SD: Removes data points outside +/- 3SD of the meadian}

\item{x.dist}{Distribution of the data. 
Choices include: normal, lognormal, exponential, weibull, poisson, negative binomial}

\item{save.output}{Save plots as pdf file. If TRUE, three plots are returned as pdf file. If false, plots are printed to the screen.}

\item{table.dat}{Print a summary data table. If TRUE, eight summary measures are printed}

\item{remove}{Save data with outliers removed. If TRUE, the revised data tables, based on outlier.mod, is returned.}
}
\value{
Returns the modified dataframe
}
\description{
Contains functions to evaluate the data through histogram, QQ, and x-y plots.
 contains functions to evaulate the fit of the data in terms of distributions and analysis
}
\details{
outlier_check is a series of functions allowing users to evaluate the occurrence of outliers and their effect on the data
using summary data and plots. Outliers can onlyl be assessed for numeric vectors.
}
\examples{
Generate data for example. Inject outliers into cars dataset.
cars1 <- cbind(rbind(cars[1:30, ], data.frame(speed=c(19,19,20,20,20), dist=c(190, 186, 210, 220, 218))), ID=rep(c('a','b','c','d','e'), 7))
outlier_table(cars1, 'dist')
outlier_plot(cars1, 'dist', outlier.mod='mean_2SD', x.dist='normal', save.output=FALSE)
cars1 <- outlier_remove(cars1, 'dist', outlier.mod='mean_2SD', save.output=TRUE)
}
\keyword{outliers}
