% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparstable.R
\name{sparsetable}
\alias{sparsetable}
\title{Evaluate sparsity in data over time in table format}
\usage{
sparsetable(dat, project, timevar, zonevar, var)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips. Table in the FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{timevar}{Variable in \code{dat} containing temporal data}

\item{zonevar}{Variable in \code{dat} containing zone observation assigned to}

\item{var}{Variable in \code{dat} containing catch data}
}
\description{
Create table of data sparsity by predefined time periods.
}
