% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparstable.R
\name{sparsplot}
\alias{sparsplot}
\title{Evaluate sparsity in data over time in plot format}
\usage{
sparsplot(project, x = NULL)
}
\arguments{
\item{project}{String, name of project.}

\item{x}{Output from \code{sparsetable}. If \code{x} is null, the sparsity table will be pulled from the output folder if it exists.}
}
\description{
Evaluate sparsity in data over time in plot format
}
\details{
Returns a plot of sparsity values over time.
Requires sparsity table generated by \code{sparsetable}.
}
