% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_output_funcs.R
\name{spatial_hist}
\alias{spatial_hist}
\title{Histogram of latitude and longitude by grouping variable}
\usage{
spatial_hist(dat, project, group = NULL)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{group}{Column in \code{dat} containing grouping categories.}
}
\value{
Returns histogram of latitude and longitude by grouping variable.
  Output returned to console and saved to Output folder.
}
\description{
Histogram of latitude and longitude by grouping variable
}
\details{
Returns a histogram of observed lat/lon split by grouping variable.
  Output printed to console and saved to Output folder.  Function is used to
  assess spatial variance/clumping of selected grouping variable.
}
\examples{
\dontrun{
spatial_hist(pollockMainDataTable, 'pollock', 'GEAR_TYPE')
}
}
