% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_output_funcs.R
\name{spatial_summary}
\alias{spatial_summary}
\title{Summarize variable over data and time}
\usage{
spatial_summary(
  dat,
  project,
  stat.var = c("length", "no_unique_obs", "perc_total", "mean", "median", "min", "max",
    "sum"),
  variable,
  spat,
  lon.spat = NULL,
  lat.spat = NULL,
  lon.dat = NULL,
  lat.dat = NULL,
  cat
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{String, name of project.}

\item{stat.var}{Options are \code{"length"}, \code{"no_unique_obs"}, \code{"perc_total"},
\code{"mean"}, \code{"median"}, \code{"min"}, \code{"max"}, and \code{"sum"}.}

\item{variable}{Variable in \code{dat} to summarize over date and zone.}

\item{spat}{Spatial data containing information on fishery management or regulatory zones.
Shape, json, geojson, and csv formats are supported. Leave as NULL if the variable ‘ZoneID’
assigning observations to zones exists in \code{dat}.}

\item{lon.spat}{Variable or list from \code{spat} containing longitude data. Required for csv files. 
Leave as NULL if \code{spat} is a shape or json file or if the variable ‘ZoneID’ exists in \code{dat}.}

\item{lat.spat}{Variable or list from \code{spat} containing latitude data. Required for csv files. Leave as NULL if \code{spat}
is a shape or json file, or if the variable ‘ZoneID’ exists in \code{dat}.}

\item{lon.dat}{Longitude variable in \code{dat}. Leave as NULL if the variable ‘ZoneID’ (zonal 
assignment) exists in \code{dat}.}

\item{lat.dat}{Latitude variable in \code{dat}. Leave as NULL if the variable ‘ZoneID’ (zonal 
assignments) exists in \code{dat}.}

\item{cat}{Variable or list in \code{spat} that identifies the individual areas or zones. If \code{spat} is class sf, \code{cat}
should be name of list containing information on zones. Leave as NULL if the variable ‘ZoneID’ exists in \code{dat}.}
}
\value{
Returns two plots, the variable aggregated by \code{stat.var} plotted against date and against zone.
}
\description{
View summary and exploratory statistics of selected variable by date and zone.
}
\details{
\code{stat.var} details:
\tabular{rlll}{
length: \tab Number of observations \cr
no_unique_obs: \tab Number of unique observations \cr
perc_total: \tab Percent of total observations \cr
mean: \tab Mean \cr
median: \tab  Median \cr
min: \tab  Minimum\cr
max: \tab Maximum \cr
sum: \tab Sum \cr
}
}
\examples{
\dontrun{
Example where ZoneID exists in dataset
   spatial_summary(pcodMainDataTable, project = 'pcod', 
      stat.var = "no_unique_obs", variable = 'HAUL')

Example where obs. have not been assigned to zones
    spatial_summary(pcodMainDataTable, project = 'pcod', stat.var = "no_unique_obs",
       variable = 'HAUL', spat = spatdat, lon.dat = 'MidLat', lat.dat = 'MidLat',
       cat = 'NMFS_AREA')
}
}
