% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_exists}
\alias{table_exists}
\title{Check whether a specific table exists in the fishset_db database}
\usage{
table_exists(table)
}
\arguments{
\item{table}{name of tablel in sql database}
}
\description{
Check whether a specific table exists in the fishset_db database
}
