% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_exists}
\alias{table_exists}
\title{Check if table exists in the FishSET database for the defined project}
\usage{
table_exists(table, project)
}
\arguments{
\item{table}{Name of table in FishSET database.Table name must be in quotes.}

\item{project}{Name of project}
}
\value{
Returns a logical statement of table existence.
}
\description{
Wrapper for \code{\link[DBI]{dbExistsTable}}. Check if a table 
  exists in the FishSET database.
}
\examples{
\dontrun{
table_exists('pollockMainDataTable')
}
}
