% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_fields}
\alias{table_fields}
\title{Lists fields for FishSET database table}
\usage{
table_fields(table, project)
}
\arguments{
\item{table}{String, name of table in FishSET database. Table name must be in quotes.}

\item{project}{Project name}
}
\description{
Wrapper for \code{\link[DBI]{dbListFields}}.  View fields of selected table.
}
\examples{
\dontrun{
table_fields('pollockMainDataTable', 'pollock')
}
}
