% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_fields}
\alias{table_fields}
\title{Lists fields for the selected fishset_db table}
\usage{
table_fields(table)
}
\arguments{
\item{table}{name of table in sql database}
}
\description{
Function utilizes sql functions to view tables in the SQLite fishset_db database. View fields of selected table.
}
\examples{
\dontrun{  
table_fields('MainDataTable') 
}
}
