% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_format}
\alias{table_format}
\title{Import and format saved tables to notebook file}
\usage{
table_format(x)
}
\arguments{
\item{x}{Name of table saved in inst/output}
}
\description{
Import and format saved tables to notebook file
}
