% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info_verification.R
\name{table_info_verification}
\alias{table_info_verification}
\title{Check for common data issues}
\usage{
table_info_verification(dataindex)
}
\arguments{
\item{dataindex}{Data table saved in fishset_db database that contains information on each column of the main dataset. 
Table name should contain the phrase `MainDataTableInfo`.
Can be called from the global environment or the fishset_db database.}
}
\value{
Returns statements as to whether issues in the `dataindex` table may exist
}
\description{
Check for common data issues
}
\details{
Checks whether specialized variables have been identified in the `dataindex` table and if units are defined and recognized.
}
\examples{
\dontrun{
#Table loaded from fishet_db database
   table_info_verification('MainDataTableInfo') 
#Table loaded from the global environment.
   table_info_verification(MainDataTableInfo) 
}
}
