% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_remove}
\alias{table_remove}
\title{remove a table from the fishset_db database}
\usage{
table_remove(table)
}
\arguments{
\item{table}{name of tablel in sql database}
}
\description{
remove a table from the fishset_db database
}
