% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_remove}
\alias{table_remove}
\title{Remove a table from the fishset_db database}
\usage{
table_remove(table)
}
\arguments{
\item{table}{Name of table in sql database}
}
\description{
Remove a table from the fishset_db database
}
\details{
Function utilizes sql functions to permanently remove a table from the SQLite fishset_db database.
}
\examples{
\dontrun{  
table_remove('MainDataTable') 
}
}
