% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{table_view}
\alias{table_view}
\title{view the selected table in the fishset_db database}
\usage{
table_view(table)
}
\arguments{
\item{table}{name of tablel in sql database}
}
\description{
view the selected table in the fishset_db database
}
