% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{tables_database}
\alias{tables_database}
\title{View names of project tables}
\usage{
tables_database(project)
}
\arguments{
\item{project}{Project name}
}
\description{
Wrapper for \code{\link[DBI]{dbListTables}}. View names of tables in a  
project's FishSET database.
}
\examples{
\dontrun{
tables_database('pollock')
}
}
