% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_obs_table.R
\name{temp_obs_table}
\alias{temp_obs_table}
\title{Table of observations by year, month, and zone}
\usage{
temp_obs_table(dat, gridfile, x, lon.grid, lat.grid, lon.dat, lat.dat, cat)
}
\arguments{
\item{dat}{Main data frame containing data on hauls or trips. Table in fishset_db database should contain the string `MainDataTable`.}

\item{gridfile}{Spatial data. Shape, json, and csv formats are supported.}

\item{x}{Column in data containing date variable}

\item{lon.grid}{Column containing longitude data in gridfile.}

\item{lat.grid}{lColumn containing latitude data in gridfile.}

\item{lon.dat}{Column containing longitude data in main data frame.}

\item{lat.dat}{Column containing latitude data in main data frame.}

\item{cat}{Column in gridfile that identifies the individual areas or zones. If gridfile is class sf, `cat` should be name of list containing information on zones.}
}
\description{
Table of observations by year, month, and zone
}
\details{
Prints tables displaying the number of observations by year, month, and zone. The assignment_column function is called to assign observations to zones. 
Use this function before calling the create_expectations function to inform choices in appropriate window size for analysis.
}
\examples{
\dontrun{
temp_obs_table('pollockMainDataTable', gridfile=map2, x="DATE_FISHING_BEGAN", 
                lon.grid="", lat.grid="", lon.dat="LonLat_START_LON", 
                lat.dat="LonLat_START_LAT", cat="NMFS_AREA")
}
}
