% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot.R
\name{temp_plot_helper}
\alias{temp_plot_helper}
\title{plot help function for spatial plots}
\usage{
temp_plot_helper(
  dataset,
  len_df,
  agg_df,
  var.select,
  date.var,
  agg_per,
  len.fun,
  agg.fun,
  plot_by_yr,
  alpha,
  pages,
  text.size
)
}
\arguments{
\item{dataset}{dataset name}

\item{len_df}{length bar plot}

\item{agg_df}{aggregation plot}

\item{var.select}{variable}

\item{date.var}{date variable}

\item{agg_per}{character}

\item{len.fun}{length, unique, percent}

\item{agg.fun}{aggregation function}

\item{plot_by_yr}{Whether to plot by year.}

\item{alpha}{transparency of data points.}

\item{pages}{Whether to output plots on a single page (\code{"single"}, the 
default) or multiple pages (\code{"multi"}).}

\item{text.size}{Text size of x-axes.}
}
\description{
plot help function for spatial plots
}
\details{
Used in temp_plot
}
\keyword{internal}
