% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalModify.R
\name{TempMod}
\alias{TempMod}
\title{Transform units of a temporal variables}
\usage{
TempMod(dataset, x, fun.mod = "", define.format)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{Time variable to modify}
}
\value{
Returns variable with modified time unites
}
\description{
Transform units of a temporal variables
}
\details{
define.format defines the format that the variable should take on. Examples include '%Y%m%d', '%Y-%m-%d %H:%M:%S'.
The function also has options to extract year, month, minute, or month/day. These are defined through fun.mod
}
\examples{
Date_Landed_YMD <- TempMod(MainDataTable, 'DATE_LANDED', fun.mod='',define.format = '\%Y\%m\%d')
Date_Landed_year <- TempMod(MainDataTable, 'DATE_LANDED', fun.mod='year',define.format = '')
}
\keyword{Date,}
\keyword{as.Date}
