% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_mod.R
\name{temporal_mod}
\alias{temporal_mod}
\title{Transform units of date variables}
\usage{
temporal_mod(
  dat,
  project,
  x,
  define.format = NULL,
  timezone = NULL,
  name = NULL,
  log_fun = TRUE,
  ...
)
}
\arguments{
\item{dat}{Primary data containing information on hauls or trips.
Table in FishSET database contains the string 'MainDataTable'.}

\item{project}{Project name.}

\item{x}{Time variable to modify from \code{dat}.}

\item{define.format}{Format of temporal data. \code{define.format} should be NULL if converting timezone for \code{x} but not
  changing format. Format can be user-defined or from pre-defined choices. Format follows \code{\link{as.Date}} format.
See Details for more information.}

\item{timezone}{String, defaults to NULL. Returns the date-time in the specified time zone. 
Must be a recognizable timezone, such as "UTC", "America/New_York", "Europe/Amsterdam".}

\item{name}{String, name of created variables. Defaults to `TempMod`.}

\item{log_fun}{Logical, whether to log function call (for internal use).}

\item{...}{Additional arguments. Use \code{tz=''} to specify time zone.}
}
\value{
Primary data set with new variable added.
}
\description{
Creates a new temporal variable by extracting temporal unit, such as year, month, or day
from a date variable.
}
\details{
Converts a date variable to desired timezone or units using \code{\link[base]{as.Date}}. \code{\link{date_parser}} is
also called to ensure the date variable is in an acceptable format for \code{\link[base]{as.Date}}.
 \code{define.format} defines the format that the variable should take on. Examples include \code{"\%Y\%m\%d"}, \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.
   Users can define their own format or use one of the predefined ones. Hours is 0-23. 
   To return a list of time-zone name in the Olson/IANA database paste \code{OlsonNames()} to the console.
\itemize{
Predefined formats:
\item{year: Takes on the format \code{"\%Y"} and returns the year.}
\item{month: Takes on the format \code{"\%Y/\%m"} and returns the year and month.}
\item{day: Takes on the format \code{"\%Y/\%m/\%d"} and returns the year, month, and day.}
\item{hour: Takes on the format \code{"\%Y/\%m/\%d \%H"} and returns the year, month, day and hour.}
\item{minute: Takes on the format \code{"\%Y/\%m/\%d \%H:\%M"} and returns the year, month, day, hour, and minute.}
}
For more information on formats, see \url{https://www.stat.berkeley.edu/~s133/dates.html}.
}
\examples{
\dontrun{
pcodMainDataTable <- temporal_mod(pcodMainDataTable, "pcod", 
   "DATE_LANDED", define.format = "\%Y\%m\%d")
pcodMainDataTable <- temporal_mod(pcodMainDataTable, "pcod", 
   "DATE_LANDED", define.format = "year")
}


# Change to Year, month, day, minutes
}
\keyword{Date}
\keyword{as.Date}
