% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_mod.R
\name{temporal_mod}
\alias{temporal_mod}
\title{Transform units of a temporal variables}
\usage{
temporal_mod(dataset, x, fun.mod = "", define.format)
}
\arguments{
\item{dataset}{dataframe or matrix}

\item{x}{Time variable to modify}

\item{fun.mod}{Defines temporal time frame such as year, month, day}

\item{define.format}{Format of temporal data}
}
\value{
Returns variable with modified time unites
}
\description{
Transform units of a temporal variables
}
\details{
define.format defines the format that the variable should take on. Examples include '%Y%m%d', '%Y-%m-%d %H:%M:%S'.
The function also has options to extract year, month, minute, or month/day. These are defined through fun.mod

Date_Landed_year <- temp_mod(MainDataTable, 'DATE_LANDED', fun.mod = 'year', define.format = '')
}
\keyword{Date,}
\keyword{as.Date}
