% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_grid_dat.R
\name{view_grid_dat}
\alias{view_grid_dat}
\title{Visualize gridded data on a map}
\usage{
view_grid_dat(
  gridfile,
  project,
  lon,
  lat,
  value,
  split_by = NULL,
  agg_by = NULL,
  agg_fun = "mean",
  gmap = FALSE
)
}
\arguments{
\item{gridfile}{Gridded data table to visualize. Use string if visualizing 
a gridded data table in the FishSET Database.}

\item{project}{String, project name.}

\item{lon}{String, variable name containing longitude.}

\item{lat}{String, variable name containing latitude.}

\item{value}{String, variable name containing gridded values, e.g. sea surface 
temperature, wind speed, etc.}

\item{split_by}{String, variable in gridded data table to split by.}

\item{agg_by}{String, variable in gridded data table to group \code{value} by.
By default, the mean is aggregated. The string "latlon" is a shortcut for 
\code{agg_by = c("lon", "lat")} which aggregates the "value" for each 
latitude-longitude pair across the entire dataset.}

\item{agg_fun}{Aggregating function applied to \code{agg_by}. Defaults to mean.}

\item{gmap}{A ggmap object to be passed to \code{\link[ggmap]{ggmap}}. If FALSE, 
then a ggmap is automatically retrieved. Defaults to FALSE.}
}
\description{
Visualize gridded data on a map
}
\examples{
\dontrun{
view_grid_dat('SST', "pollock", "lon", "lat", value = "analysed_sst")
}

}
