% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser.R
\name{write_dat}
\alias{write_dat}
\title{Write a dataset to local file}
\usage{
write_dat(dat, file, file_type = "csv", project)
}
\arguments{
\item{dat}{Name of data frame in working environment to save to file.}

\item{file}{String, file or connection to write to.}

\item{file_type}{String, the type of file to write to. Options include \code{"csv"},
\code{"txt"} (tab-separated text file), \code{"xlsx"} (excel), \code{"rdata"}, \code{"json"}, 
\code{"stata"}, \code{"spss"},
\code{"sas"}, and \code{"matlab"}.}

\item{project}{String, project name.}

\item{...}{Additional arguments passed to writing function. See "details" for 
the list of functions.}
}
\description{
Write a dataset to local file
}
\details{
See \code{\link[utils]{write.csv}} for saving csv files, \code{\link[utils]{write.table}}
 for tab-separated files, \code{\link[base]{save}} for R data files, \code{\link[openxlsx]{write.xlsx}},
 \code{\link[jsonlite]{read_json}} for json files, \code{\link[haven]{write_dta}} 
 for Stata files, \code{\link[haven]{write_sav}} for SPSS files, \code{\link[haven]{write_sas}}
 for SAS files, and \code{\link[R.matlab]{writeMat}} for Matlab files.
}
\examples{
\dontrun{
write_dat(pollockMainDataTable, file = "C://data/pollock_dataset.csv", type = "csv", "pollock")
}
}
