% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalSubroutine.R
\name{zonal_subroutine}
\alias{zonal_subroutine}
\title{zonal_subroutine}
\usage{
zonal_subroutine(catch, choice, distance, otherdat, initparams, optimOpt,
  func)
}
\arguments{
\item{catch}{Data corresponding to actual catch}

\item{choice}{Data corresponding to actual zonal. Should be as.integer.}

\item{distance}{Data corresponding to distance f}

\item{otherdat}{Other data (as a list)}

\item{initparams}{Initial parameter estimates for revenue/location-specific covariates then cost/distance}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x]}

\item{func}{Name of likelihood function}
}
\value{
OutLogit - [outmat1 se1 tEPM2] (coefs, ses, tstats) \cr 
clogitoutput - optimization information \cr 
seoumat2 - ses \cr 
MCM - Model Comparison metrics \cr 
H1 - inverse hessian \cr 
mod.out - dataframe with model comparison metrics. These are saved to the fishset database
Output also saved as a list. Use the following code to extract. some_object <- unserialize(RSQLite::dbGetQuery (mydb, 'SELECT name FROM data LIMIT 1')$name[[1]]);
}
\description{
Subroutine to run chosen discrete choice model
}
\details{
Called the ShiftSortX, CreateLogitInput and likekihood functions. Inputs from ModelInputData generated in make_model_design
}
