% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_closure.R
\name{zone_closure}
\alias{zone_closure}
\title{Define zone closure scenarios}
\usage{
zone_closure(
  project,
  spatdat,
  cat,
  lon.spat = NULL,
  lat.spat = NULL,
  epsg = NULL
)
}
\arguments{
\item{project}{Required, name of project.}

\item{spatdat}{Required, data file or character. 
\code{spatdat} is a spatial data file containing information on fishery 
management or regulatory zones boundaries. Shape, json, geojson, and csv 
formats are supported. geojson is the preferred format. json files must be 
converted into geoson. This is done automatically when the file is loaded 
with \code{\link{read_dat}} with \code{is.map} set to true. \code{spatdat} 
cannot, at this time, be loaded from the FishSET database. \cr}

\item{cat}{Variable in \code{spatdat} that identifies the individual areas or zones.}

\item{lon.spat}{Required for csv files. Variable or list from \code{spatdat} 
containing longitude data. Leave as NULL if \code{spatdat} is a shape or json file.}

\item{lat.spat}{Required for csv files. Variable or list from \code{spatdat} 
containing latitude data.  Leave as NULL if \code{spatdat} is a shape or json file.}

\item{epsg}{EPSG number. Set the epsg to ensure that \code{spatdat} have the correct projections. 
If epsg is not specified but is defined for \code{spatdat}. 
  See \url{http://spatialreference.org/} to help identify optimal epsg number.}
}
\value{
Returns a yaml file to the project output folder.
}
\description{
Define zone closure scenarios
}
\details{
Define zone closure scenarios. Function opens an interactive map. 
  Define zone closures by clicking on one or more zones and clicking the 
  'Close zones' button. To define another closure scenario, unclick zones and 
  then click the desired zones. Press the 'Save closures' button to save choices.
  The saved choices are called in the policy scenario function.
}
